bcbio single-cell RNA-seq data set

bcbioSingleCell is an S4 class that extends SingleCellExperiment, and is designed to store a bcbio single-cell RNA-seq analysis. This class contains read counts saved as a sparse matrix (sparseMatrix), sample metadata, and cell quality control metrics.

bcbioSingleCell(uploadDir, sampleMetadataFile = NULL, organism = NULL,
  ensemblRelease = NULL, genomeBuild = NULL, gffFile = NULL,
  transgeneNames = NULL, spikeNames = NULL,
  interestingGroups = "sampleName", ...)

Arguments

uploadDir
character(1). Path to final upload directory. This path is set when running "bcbio_nextgen -w template".

sampleMetadataFile
character(1). Sample metadata file path. CSV or TSV is preferred, but Excel worksheets are also supported. Check the documentation for conventions and required columns.

organism
character(1). Full Latin organism name (e.g. "Homo sapiens").

ensemblRelease
integer(1). Ensembl release version (e.g. 90). We recommend setting this value if possible, for improved reproducibility. When left unset, the latest release available via AnnotationHub/ensembldb is used. Note that the latest version available can vary, depending on the versions of AnnotationHub and ensembldb in use.

genomeBuild
character(1). Ensembl genome build assembly name (e.g. "GRCh38"). If set NULL, defaults to the most recent build available. Note: don't pass in UCSC build IDs (e.g. "hg38").

gffFile
character(1). GFF/GTF (General Feature Format) file. Generally, we recommend using a GTF (GFFv2) instead of a GFFv3 file if possible.

transgeneNames
character. Vector indicating which assay rows denote transgenes (e.g. EGFP, TDTOMATO).

spikeNames
character. Vector indicating which assay rows denote spike-in sequences (e.g. ERCCs).

interestingGroups
character. Groups of interest that define the samples. If left unset, defaults to sampleName.

...
Additional arguments.

Value

bcbioSingleCell.

Note

Updated 2019-07-30.

Remote data

When working in RStudio, we recommend connecting to the bcbio-nextgen run directory as a remote connection over sshfs.

See also

Examples

uploadDir <- system.file("extdata/indrops", package = "bcbioSingleCell")

x <- bcbioSingleCell(uploadDir)

#> Dated project directory: 2018-01-01_bcbio

#> 1 sample detected:
#> [1] "multiplexed_AAAAAAAA"

#> Failed to detect sequencing lanes.

#> Importing project-summary.yaml using yaml::yaml.load_file().

#> Data versions are missing.

#> Importing bcbio-nextgen.log using readr::read_lines().

#> bcbio-nextgen.log file is empty.

#> Importing bcbio-nextgen-commands.log using readr::read_lines().

#> 1000 reads per cellular barcode cutoff detected.

#> Counts will imported as genes.

#> UMI type: harvard-indrop-v3

#> Importing unfiltered cellular barcode distributions...

#> Importing counts...

#> Imported multiplexed-AAAAAAAA.

#> bcbio GTF file: /n/app/bcbio/dev/genomes/Hsapiens/hg38/rnaseq/ref-transcripts.gtf

#> bcbio GTF file is not accessible.

#> Slotting empty ranges into rowRanges().

#> `sampleMetadataFile` is recommended for multiplexed samples (e.g. harvard-indrop-v3).

#> Calculating cellular barcode metrics.

#> 100 cells detected.

#> Calculating metrics without biotype information.
#> `rowRanges` is required to calculate: nCoding, nMito, mitoRatio

#> 100 / 100 cellular barcodes passed pre-filtering (100%)

print(x)

#> bcbioSingleCell 0.3.18
#> uploadDir: /data00/R/site-library/3.6/bcbioSingleCell/extdata/indrops
#> dates(2): [bcbio] 2018-01-01; [R] 2019-07-30
#> level: genes
#> interestingGroups: sampleName
#> filtered: FALSE
#> class: SingleCellExperiment 
#> dim: 50 100 
#> metadata(27): version pipeline ... wd sessionInfo
#> assays(1): counts
#> rownames(50): ENSG00000071082 ENSG00000100316 ... ENSG00000269028
#>   ENSG00000282105
#> rowData names(0):
#> colnames(100): AAACACTA_CTTCGATT AAACTACA_CCACATTA ...
#>   TGGGAATT_ATATAGGA TGTTATCA_ACGCAGAG
#> colData names(9): nUMI nGene ... sampleID sampleName
#> reducedDimNames(0):
#> spikeNames(0):

x <- bcbioSingleCell(
    uploadDir = uploadDir,
    sampleMetadataFile = file.path(uploadDir, "metadata.csv")
)

#> Dated project directory: 2018-01-01_bcbio

#> 1 sample detected:
#> [1] "multiplexed_AAAAAAAA"

#> Failed to detect sequencing lanes.

#> Importing project-summary.yaml using yaml::yaml.load_file().

#> Data versions are missing.

#> Importing bcbio-nextgen.log using readr::read_lines().

#> bcbio-nextgen.log file is empty.

#> Importing bcbio-nextgen-commands.log using readr::read_lines().

#> 1000 reads per cellular barcode cutoff detected.

#> Counts will imported as genes.

#> UMI type: harvard-indrop-v3

#> Importing metadata.csv using data.table::fread().

#> Multiplexed samples detected.

#> Importing unfiltered cellular barcode distributions...

#> Importing counts...

#> Imported multiplexed-AAAAAAAA.

#> bcbio GTF file: /n/app/bcbio/dev/genomes/Hsapiens/hg38/rnaseq/ref-transcripts.gtf

#> bcbio GTF file is not accessible.

#> Slotting empty ranges into rowRanges().

#> Calculating cellular barcode metrics.

#> 100 cells detected.

#> Calculating metrics without biotype information.
#> `rowRanges` is required to calculate: nCoding, nMito, mitoRatio

#> 100 / 100 cellular barcodes passed pre-filtering (100%)

print(x)

#> bcbioSingleCell 0.3.18
#> uploadDir: /data00/R/site-library/3.6/bcbioSingleCell/extdata/indrops
#> dates(2): [bcbio] 2018-01-01; [R] 2019-07-30
#> level: genes
#> sampleMetadataFile: /data00/R/site-library/3.6/bcbioSingleCell/extdata/indrops/metadata.csv
#> interestingGroups: sampleName
#> filtered: FALSE
#> class: SingleCellExperiment 
#> dim: 50 100 
#> metadata(27): version pipeline ... wd sessionInfo
#> assays(1): counts
#> rownames(50): ENSG00000071082 ENSG00000100316 ... ENSG00000269028
#>   ENSG00000282105
#> rowData names(0):
#> colnames(100): AAACACTA_CTTCGATT AAACTACA_CCACATTA ...
#>   TGGGAATT_ATATAGGA TGTTATCA_ACGCAGAG
#> colData names(15): nUMI nGene ... aggregate revcomp
#> reducedDimNames(0):
#> spikeNames(0):