Coerce Object

Coerce Object

Arguments

from
Class for which the coerce method will perform coercion.

Upgrade bcbioSingleCell to current version

This method adds support for upgrading bcbioSCDataSet objects to the latest bcbioSingleCell class version. This should be backwards compatible to bcbioSingleCell version 0.0.17. Previous objects saved using bcbioSinglecell (note case) will likely fail to load with newer versions of the package.

bcbioSingleCell to seurat

Interally, this begins by calling Seurat::CreateSeuratObject() without any additional filtering cutoffs, since we already applied them during our quality control analysis. Here we are passing the raw gene-level counts of the filtered cells into a new seurat class object, using as() object coercion. Next, global-scaling normalization is applied to the raw counts with Seurat::NormalizeData(), which (1) normalizes the gene expression measurements for each cell by the total expression, (2) multiplies this by a scale factor (10,000 by default), and (3) log-transforms the result. Seurat::FindVariableGenes() is then called, which calculates the average expression and dispersion for each gene, places these genes into bins, and then calculates a z-score for dispersion within each bin. This helps control for the relationship between variability and average expression. Finally, the genes are scaled and centered using the Seurat::ScaleData() function.

See also

help(topic = "coerce", package = "methods").

Examples

# NOT RUN {
seurat <- as(bcb, "seurat")
# }