Combine Multiple SingleCellExperiment Objects

Combine Multiple SingleCellExperiment Objects

# S4 method for SingleCellExperiment,SingleCellExperiment
combine(x, y, ...)

Arguments

x
One of the values.

y
A second value.

...
Any other objects of the same class as x and y.

Value

SingleCellExperiment.

Note

We're attempting to make this as strict as possible, requiring:

Rows (genes) across objects must be identical.

Organism and genome build must be identical.

rowRanges metadata must be identical.

colData (cellular barcode metrics) must contain the same columns.

sampleData must contain the same columns.

See also

Examples

# For this simple working example, let's duplicate our minimal dataset,
# modify the cell barcodes, and combine

x <- indrops_small
validObject(x)

#> [1] TRUE

sampleData(x)

#> DataFrame with 1 row and 8 columns
#>                      sampleName             fileName          description
#>                        <factor>             <factor>             <factor>
#> multiplexed_AAAAAAAA      rep_1 multiplexed.fastq.gz multiplexed-AAAAAAAA
#>                         index sequence aggregate  revcomp interestingGroups
#>                      <factor> <factor>  <factor> <factor>          <factor>
#> multiplexed_AAAAAAAA        1 TTTTTTTT    sample AAAAAAAA             rep_1

sampleNames(x)

#> multiplexed_AAAAAAAA 
#>              "rep_1" 

# Here we're faking a distinct replicate, just as an example
y <- indrops_small
# Modify the cellular barcodes prefix
colnames(y) <- gsub(
    pattern = "^multiplexed_AAAAAAAA",
    replacement = "multiplexed_CCCCCCCC",
    x = colnames(y)
)
# Modify the column data
colData(y)[["description"]] <- factor("multiplexed-CCCCCCCC")
colData(y)[["sampleID"]] <- factor("multiplexed_CCCCCCCC")
colData(y)[["sampleName"]] <- factor("rep_2")
colData(y)[["sequence"]] <- factor("GGGGGGGG")
colData(y)[["revcomp"]] <- factor("CCCCCCCC")
# Modify the sample data
sampleData(y)[["sampleName"]] <- factor("rep_2")
sampleData(y)[["description"]] <- factor("multiplexed-CCCCCCCC")
sampleData(y)[["sequence"]] <- factor("GGGGGGGG")
sampleData(y)[["revcomp"]] <- factor("CCCCCCCC")
rownames(sampleData(y)) <- "multiplexed_CCCCCCCC"
validObject(y)

#> [1] TRUE

sampleData(y)

#> DataFrame with 1 row and 8 columns
#>                      sampleName             fileName          description
#>                        <factor>             <factor>             <factor>
#> multiplexed_CCCCCCCC      rep_2 multiplexed.fastq.gz multiplexed-CCCCCCCC
#>                         index sequence aggregate  revcomp interestingGroups
#>                      <factor> <factor>  <factor> <factor>          <factor>
#> multiplexed_CCCCCCCC        1 GGGGGGGG    sample CCCCCCCC             rep_2

sampleNames(y)

#> multiplexed_CCCCCCCC 
#>              "rep_2" 

# Combine two SingleCellExperiment objects
c <- combine(x, y)
print(c)

#> class: SingleCellExperiment 
#> dim: 500 1000 
#> metadata(12): interestingGroups organism ... utilsSessionInfo
#>   devtoolsSessionInfo
#> assays(1): counts
#> rownames(500): ENSG00000000457 ENSG00000002586 ... ENSG00000283060
#>   ENSG00000283391
#> rowData names(7): broadClass description ... geneName seqCoordSystem
#> colnames(1000): multiplexed_AAAAAAAA_AAACACTA_ATCCCACG
#>   multiplexed_AAAAAAAA_AAACACTA_CTTAGGTA ...
#>   multiplexed_CCCCCCCC_TGTTATCA_ACGCAGAG
#>   multiplexed_CCCCCCCC_TGTTATCA_CAAGTCAT
#> colData names(15): nCount sampleID ... aggregate revcomp
#> reducedDimNames(0):
#> spikeNames(0):

sampleNames(c)

#> multiplexed_AAAAAAAA multiplexed_CCCCCCCC 
#>              "rep_1"              "rep_2"