Differential Expression

Differential Expression

diffExp(object, ...)

# S4 method for SingleCellExperiment
diffExp(object, numerator, denominator,
  minCells = 10L, maxit = 1000L)

# S4 method for seurat
diffExp(object, numerator, denominator, minCells = 10L,
  maxit = 1000L)

Arguments

object
Object.

...
Additional arguments.

numerator
Group of cells to use in the numerator of the contrast (e.g. treatment).

denominator
Group of cells to use in the denominator of the contrast (e.g. control).

minCells
Minimum number of cells required per group.

maxit
The number of iterations for the EM-algorithm. 200 by default, but larger may be useful for large datasets (many samples). Convergence of the posterior probabilities can be checked by following the distribution of posterior probabilities over iterations with plotW. The EM-algorithm will automatically stop if convergence is achieved before the maximum number of iterations.

Value

DEGLRT.

zingeR-edgeR

We first perform a differential expression analysis using zingeR posterior probabilities and focussing on the count component of the ZINB model. We use edgeR to model the count component.

The weights are estimated with the core function of zingeR, zeroWeightsLS(). It is important to be consistent with the normalization procedure, i.e. if you use TMM normalization for the analysis, you should also use it for estimating the zingeR posterior probabilities. The weights can then be provided in the DGEList object which will be used for dispersion estimation in the native edgeR estimateDisp() function.

After estimation of the dispersions and zingeR posterior probabilities, the glmWeightedF() function is used for statistical inference. This is an adapted function from the glmLRT() function of edgeR. It uses an F-test for which the denominator degrees of freedom are by default adjusted according to the downweighting of excess zeros (ZI = TRUE). Also, independent filtering can be performed on the obtained p-values (independentFiltering = TRUE). We use the independent filtering strategy that was originally implemented in DESeq2. By default, the average fitted values are used as a filter criterion.

Consult the zingeR vignette v2 for additional information.

See also

Examples

# SingleCellExperiment ====
x <- metrics(cellranger_small)
numerator <- rownames(x)[which(x[["sampleName"]] == "proximal")]
denominator <- rownames(x)[which(x[["sampleName"]] == "distal")]
lrt <- diffExp(
    object = cellranger_small,
    numerator = numerator,
    denominator = denominator,
    maxit = 100L
)

#> Loading required namespace: zingeR

#> Loading required namespace: edgeR

#> converting counts to integer mode

#> Loading required package: edgeR

#> Loading required package: limma

#> 
#> Attaching package: 'limma'

#> The following object is masked from 'package:BiocGenerics':
#> 
#>     plotMA

#> 
#> Attaching package: 'edgeR'

#> The following object is masked from 'package:SingleCellExperiment':
#> 
#>     cpm

#> iteration: 1. mean conv.: 0
#> iteration: 2. mean conv.: 0.16
#> iteration: 3. mean conv.: 0.958
#> iteration: 4. mean conv.: 0.986
#> iteration: 5. mean conv.: 0.994
#> iteration: 6. mean conv.: 0.998
#> iteration: 7. mean conv.: 0.998
#> iteration: 8. mean conv.: 1
#> iteration: 9. mean conv.: 0.98
#> iteration: 10. mean conv.: 1
#> converged. 

#> Loading required package: genefilter

#> 
#> Attaching package: 'genefilter'

#> The following objects are masked from 'package:matrixStats':
#> 
#>     rowSds, rowVars

glimpse(lrt[["table"]])

#> Observations: 500
#> Variables: 5
#> $ logFC      <dbl> 0.17740165, 0.67929074, 1.25917593, -0.53085841, -1.4541...
#> $ logCPM     <dbl> 10.050208, 8.730282, 10.027433, 12.389401, 8.905963, 8.6...
#> $ LR         <dbl> 9.1513701, 106.3974491, 168.5623434, 309.4309766, 25.178...
#> $ PValue     <dbl> 2.612343e-03, 9.251915e-23, 1.893677e-33, 2.417904e-54, ...
#> $ padjFilter <dbl> 3.178032e-03, 2.141647e-22, 6.667875e-33, 2.281041e-53, ...

# seurat ====
# Expression in cluster 3 relative to cluster 2
numerator <- Seurat::WhichCells(Seurat::pbmc_small, ident = 3L)
denominator <- Seurat::WhichCells(Seurat::pbmc_small, ident = 2L)
lrt <- diffExp(
    object = Seurat::pbmc_small,
    numerator = numerator,
    denominator = denominator,
    maxit = 100L
)

#> converting counts to integer mode

#> iteration: 1. mean conv.: 0
#> iteration: 2. mean conv.: 0.1913
#> iteration: 3. mean conv.: 0.18696
#> iteration: 4. mean conv.: 0.1913
#> iteration: 5. mean conv.: 0.21304
#> iteration: 6. mean conv.: 0.18696
#> iteration: 7. mean conv.: 0.2087
#> iteration: 8. mean conv.: 0.23043
#> iteration: 9. mean conv.: 0.23043
#> iteration: 10. mean conv.: 0.22174
#> iteration: 11. mean conv.: 0.23478
#> iteration: 12. mean conv.: 0.23913
#> iteration: 13. mean conv.: 0.21739
#> iteration: 14. mean conv.: 0.2087
#> iteration: 15. mean conv.: 0.2087
#> iteration: 16. mean conv.: 0.20435
#> iteration: 17. mean conv.: 0.2087
#> iteration: 18. mean conv.: 0.21739
#> iteration: 19. mean conv.: 0.22609
#> iteration: 20. mean conv.: 0.22609
#> iteration: 21. mean conv.: 0.26957
#> iteration: 22. mean conv.: 0.27391
#> iteration: 23. mean conv.: 0.27391
#> iteration: 24. mean conv.: 0.28261
#> iteration: 25. mean conv.: 0.29565
#> iteration: 26. mean conv.: 0.3
#> iteration: 27. mean conv.: 0.3087
#> iteration: 28. mean conv.: 0.32609
#> iteration: 29. mean conv.: 0.33913
#> iteration: 30. mean conv.: 0.33913
#> iteration: 31. mean conv.: 0.34348
#> iteration: 32. mean conv.: 0.33913
#> iteration: 33. mean conv.: 0.34348
#> iteration: 34. mean conv.: 0.34783
#> iteration: 35. mean conv.: 0.35217
#> iteration: 36. mean conv.: 0.36957
#> iteration: 37. mean conv.: 0.38696
#> iteration: 38. mean conv.: 0.39565
#> iteration: 39. mean conv.: 0.40435
#> iteration: 40. mean conv.: 0.41304
#> iteration: 41. mean conv.: 0.41739
#> iteration: 42. mean conv.: 0.42609
#> iteration: 43. mean conv.: 0.43478
#> iteration: 44. mean conv.: 0.44348
#> iteration: 45. mean conv.: 0.44783
#> iteration: 46. mean conv.: 0.46087
#> iteration: 47. mean conv.: 0.47391
#> iteration: 48. mean conv.: 0.48696
#> iteration: 49. mean conv.: 0.5
#> iteration: 50. mean conv.: 0.5
#> iteration: 51. mean conv.: 0.5087
#> iteration: 52. mean conv.: 0.52174
#> iteration: 53. mean conv.: 0.52609
#> iteration: 54. mean conv.: 0.54348
#> iteration: 55. mean conv.: 0.55652
#> iteration: 56. mean conv.: 0.57826
#> iteration: 57. mean conv.: 0.57826
#> iteration: 58. mean conv.: 0.58696
#> iteration: 59. mean conv.: 0.60435
#> iteration: 60. mean conv.: 0.6087
#> iteration: 61. mean conv.: 0.62609
#> iteration: 62. mean conv.: 0.64783
#> iteration: 63. mean conv.: 0.66957
#> iteration: 64. mean conv.: 0.69565
#> iteration: 65. mean conv.: 0.72174
#> iteration: 66. mean conv.: 0.74783
#> iteration: 67. mean conv.: 0.77391
#> iteration: 68. mean conv.: 0.77826
#> iteration: 69. mean conv.: 0.78696
#> iteration: 70. mean conv.: 0.7913
#> iteration: 71. mean conv.: 0.79565
#> iteration: 72. mean conv.: 0.80435
#> iteration: 73. mean conv.: 0.82609
#> iteration: 74. mean conv.: 0.83043
#> iteration: 75. mean conv.: 0.83913
#> iteration: 76. mean conv.: 0.85217
#> iteration: 77. mean conv.: 0.86957
#> iteration: 78. mean conv.: 0.88261
#> iteration: 79. mean conv.: 0.89565
#> iteration: 80. mean conv.: 0.91304
#> iteration: 81. mean conv.: 0.92174
#> iteration: 82. mean conv.: 0.93043
#> iteration: 83. mean conv.: 0.93478
#> iteration: 84. mean conv.: 0.93913
#> iteration: 85. mean conv.: 0.94783
#> iteration: 86. mean conv.: 0.95217
#> iteration: 87. mean conv.: 0.95217
#> iteration: 88. mean conv.: 0.95652
#> iteration: 89. mean conv.: 0.98261
#> iteration: 90. mean conv.: 0.98696
#> iteration: 91. mean conv.: 0.98696
#> iteration: 92. mean conv.: 0.9913
#> iteration: 93. mean conv.: 0.9913
#> iteration: 94. mean conv.: 0.9913
#> iteration: 95. mean conv.: 0.9913
#> iteration: 96. mean conv.: 0.9913
#> iteration: 97. mean conv.: 0.9913
#> iteration: 98. mean conv.: 0.99565
#> iteration: 99. mean conv.: 0.99565
#> iteration: 100. mean conv.: 1

glimpse(lrt[["table"]])

#> Observations: 230
#> Variables: 5
#> $ logFC      <dbl> 0.0000000, -1.4941343, 0.0000000, 0.8996484, 0.0000000, ...
#> $ logCPM     <dbl> 12.98492, 13.08969, 12.98492, 13.71783, 12.98492, 13.005...
#> $ LR         <dbl> 0.0000000, 2.0319785, 0.0000000, 1.0090244, 0.0000000, 1...
#> $ PValue     <dbl> 1.00000000, 0.15417754, 1.00000000, 0.31525937, 1.000000...
#> $ padjFilter <dbl> NA, 0.2459308, NA, 0.3972781, NA, NA, NA, NA, 0.3572848,...