Fetch Data Functions

Fetch Data Functions

fetchGeneData(object, ...)

fetchPCAData(object, ...)

fetchTSNEData(object, ...)

fetchTSNEExpressionData(object, ...)

fetchUMAPData(object, ...)

fetchUMAPExpressionData(object, ...)

# S4 method for SingleCellExperiment
fetchGeneData(object, genes,
  gene2symbol = FALSE)

# S4 method for seurat
fetchGeneData(object, genes, gene2symbol = FALSE)

# S4 method for SingleCellExperiment
fetchPCAData(object, minimal = FALSE)

# S4 method for seurat
fetchPCAData(object, minimal = FALSE)

# S4 method for SingleCellExperiment
fetchTSNEData(object, minimal = FALSE)

# S4 method for seurat
fetchTSNEData(object, minimal = FALSE)

# S4 method for SingleCellExperiment
fetchTSNEExpressionData(object, genes)

# S4 method for seurat
fetchTSNEExpressionData(object, genes)

# S4 method for SingleCellExperiment
fetchUMAPData(object, minimal = FALSE)

# S4 method for seurat
fetchUMAPData(object, minimal = FALSE)

# S4 method for SingleCellExperiment
fetchUMAPExpressionData(object, genes)

# S4 method for seurat
fetchUMAPExpressionData(object, genes)

Arguments

object
Object.

...
Additional arguments.

genes
character. Gene identifiers. Must match the rownames of the object.

gene2symbol
data.frame. Gene-to-symbol mappings. Columns must contain geneID and geneName.

minimal
boolean. Return minimal data without metrics.

Value

fetchGeneData(): matrix.

fetchTSNEExpressionData(): data.frame containing t-SNE coordinates, sample metadata, and aggregate marker expression values (mean, median, and sum).

Other functions: data.frame, containing metrics.

Note

Some of these functions require "ident" to be defined in colData(). For t-SNE, UMAP, and PCA, the reduced dimensions must be defined in the reducedDims() slot.

See also

Examples

# SingleCellExperiment ====
object <- indrops_small
genes <- head(rownames(object))
glimpse(genes)

#>  chr [1:6] "ENSG00000000457" "ENSG00000002586" "ENSG00000004939" ...

# Genes
x <- fetchGeneData(object, genes = genes)
glimpse(x)

#>  num [1:500, 1:6] 1 1 4 3 3 2 5 1 3 3 ...
#>  - attr(*, "dimnames")=List of 2
#>   ..$ : chr [1:500] "multiplexed_AAAAAAAA_AAACACTA_ATCCCACG" "multiplexed_AAAAAAAA_AAACACTA_CTTAGGTA" "multiplexed_AAAAAAAA_AAACACTA_CTTCGATT" "multiplexed_AAAAAAAA_AAACCGCC_GAGTTTCG" ...
#>   ..$ : chr [1:6] "ENSG00000000457" "ENSG00000002586" "ENSG00000004939" "ENSG00000005022" ...

# t-SNE
x <- fetchTSNEData(object)
glimpse(x)

#> Observations: 500
#> Variables: 24
#> $ nGene            <int> 289, 405, 423, 337, 350, 378, 442, 217, 332, 405, ...
#> $ nUMI             <int> 1199, 2766, 4399, 1485, 1771, 2208, 4095, 1406, 15...
#> $ nCount           <int> 25114, 55337, 71802, 34607, 34005, 44760, 76590, 3...
#> $ sampleID         <fct> multiplexed_AAAAAAAA, multiplexed_AAAAAAAA, multip...
#> $ nCoding          <int> 1105, 2385, 3864, 1296, 1574, 1599, 3630, 695, 117...
#> $ nMito            <int> 24, 241, 234, 103, 83, 452, 234, 556, 246, 225, 28...
#> $ log10GenesPerUMI <dbl> 0.7992992, 0.7575732, 0.7208579, 0.7969256, 0.7832...
#> $ mitoRatio        <dbl> 0.02001668, 0.08712943, 0.05319391, 0.06936027, 0....
#> $ sampleName       <fct> rep_1, rep_1, rep_1, rep_1, rep_1, rep_1, rep_1, r...
#> $ fileName         <fct> multiplexed.fastq.gz, multiplexed.fastq.gz, multip...
#> $ description      <fct> multiplexed-AAAAAAAA, multiplexed-AAAAAAAA, multip...
#> $ index            <fct> 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1,...
#> $ sequence         <fct> TTTTTTTT, TTTTTTTT, TTTTTTTT, TTTTTTTT, TTTTTTTT, ...
#> $ aggregate        <fct> sample, sample, sample, sample, sample, sample, sa...
#> $ revcomp          <fct> AAAAAAAA, AAAAAAAA, AAAAAAAA, AAAAAAAA, AAAAAAAA, ...
#> $ orig.ident       <fct> multiplexed, multiplexed, multiplexed, multiplexed...
#> $ res.0.4          <chr> "0", "2", "0", "0", "0", "1", "0", "0", "0", "2", ...
#> $ res.0.8          <chr> "3", "2", "1", "1", "3", "0", "1", "1", "3", "2", ...
#> $ res.1.2          <chr> "4", "0", "1", "1", "4", "2", "1", "1", "4", "0", ...
#> $ ident            <fct> 0, 2, 0, 0, 0, 1, 0, 0, 0, 2, 0, 1, 1, 1, 0, 1, 2,...
#> $ tSNE_1           <dbl> -17.8005980, 14.3021396, 5.5353587, 5.5824008, -17...
#> $ tSNE_2           <dbl> -8.7635904, -1.2800499, -15.9179439, -9.7737431, -...
#> $ centerX          <dbl> -1.020748, 13.796989, -1.020748, -1.020748, -1.020...
#> $ centerY          <dbl> -9.92682961, 0.06908376, -9.92682961, -9.92682961,...

# PCA
x <- fetchPCAData(object)
glimpse(x)

#> Observations: 500
#> Variables: 24
#> $ nGene            <int> 289, 405, 423, 337, 350, 378, 442, 217, 332, 405, ...
#> $ nUMI             <int> 1199, 2766, 4399, 1485, 1771, 2208, 4095, 1406, 15...
#> $ nCount           <int> 25114, 55337, 71802, 34607, 34005, 44760, 76590, 3...
#> $ sampleID         <fct> multiplexed_AAAAAAAA, multiplexed_AAAAAAAA, multip...
#> $ nCoding          <int> 1105, 2385, 3864, 1296, 1574, 1599, 3630, 695, 117...
#> $ nMito            <int> 24, 241, 234, 103, 83, 452, 234, 556, 246, 225, 28...
#> $ log10GenesPerUMI <dbl> 0.7992992, 0.7575732, 0.7208579, 0.7969256, 0.7832...
#> $ mitoRatio        <dbl> 0.02001668, 0.08712943, 0.05319391, 0.06936027, 0....
#> $ sampleName       <fct> rep_1, rep_1, rep_1, rep_1, rep_1, rep_1, rep_1, r...
#> $ fileName         <fct> multiplexed.fastq.gz, multiplexed.fastq.gz, multip...
#> $ description      <fct> multiplexed-AAAAAAAA, multiplexed-AAAAAAAA, multip...
#> $ index            <fct> 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1,...
#> $ sequence         <fct> TTTTTTTT, TTTTTTTT, TTTTTTTT, TTTTTTTT, TTTTTTTT, ...
#> $ aggregate        <fct> sample, sample, sample, sample, sample, sample, sa...
#> $ revcomp          <fct> AAAAAAAA, AAAAAAAA, AAAAAAAA, AAAAAAAA, AAAAAAAA, ...
#> $ orig.ident       <fct> multiplexed, multiplexed, multiplexed, multiplexed...
#> $ res.0.4          <chr> "0", "2", "0", "0", "0", "1", "0", "0", "0", "2", ...
#> $ res.0.8          <chr> "3", "2", "1", "1", "3", "0", "1", "1", "3", "2", ...
#> $ res.1.2          <chr> "4", "0", "1", "1", "4", "2", "1", "1", "4", "0", ...
#> $ ident            <fct> 0, 2, 0, 0, 0, 1, 0, 0, 0, 2, 0, 1, 1, 1, 0, 1, 2,...
#> $ PC1              <dbl> 1.0316848, 5.3377977, 2.2791735, 3.3898010, 1.1814...
#> $ PC2              <dbl> 1.39906902, -2.94220122, 0.22684589, 1.44953021, 0...
#> $ centerX          <dbl> 1.962767, 3.819283, 1.962767, 1.962767, 1.962767, ...
#> $ centerY          <dbl> 1.84123959, -2.55660965, 1.84123959, 1.84123959, 1...

# UMAP
x <- fetchUMAPData(object)
glimpse(x)

#> Observations: 500
#> Variables: 24
#> $ nGene            <int> 289, 405, 423, 337, 350, 378, 442, 217, 332, 405, ...
#> $ nUMI             <int> 1199, 2766, 4399, 1485, 1771, 2208, 4095, 1406, 15...
#> $ nCount           <int> 25114, 55337, 71802, 34607, 34005, 44760, 76590, 3...
#> $ sampleID         <fct> multiplexed_AAAAAAAA, multiplexed_AAAAAAAA, multip...
#> $ nCoding          <int> 1105, 2385, 3864, 1296, 1574, 1599, 3630, 695, 117...
#> $ nMito            <int> 24, 241, 234, 103, 83, 452, 234, 556, 246, 225, 28...
#> $ log10GenesPerUMI <dbl> 0.7992992, 0.7575732, 0.7208579, 0.7969256, 0.7832...
#> $ mitoRatio        <dbl> 0.02001668, 0.08712943, 0.05319391, 0.06936027, 0....
#> $ sampleName       <fct> rep_1, rep_1, rep_1, rep_1, rep_1, rep_1, rep_1, r...
#> $ fileName         <fct> multiplexed.fastq.gz, multiplexed.fastq.gz, multip...
#> $ description      <fct> multiplexed-AAAAAAAA, multiplexed-AAAAAAAA, multip...
#> $ index            <fct> 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1,...
#> $ sequence         <fct> TTTTTTTT, TTTTTTTT, TTTTTTTT, TTTTTTTT, TTTTTTTT, ...
#> $ aggregate        <fct> sample, sample, sample, sample, sample, sample, sa...
#> $ revcomp          <fct> AAAAAAAA, AAAAAAAA, AAAAAAAA, AAAAAAAA, AAAAAAAA, ...
#> $ orig.ident       <fct> multiplexed, multiplexed, multiplexed, multiplexed...
#> $ res.0.4          <chr> "0", "2", "0", "0", "0", "1", "0", "0", "0", "2", ...
#> $ res.0.8          <chr> "3", "2", "1", "1", "3", "0", "1", "1", "3", "2", ...
#> $ res.1.2          <chr> "4", "0", "1", "1", "4", "2", "1", "1", "4", "0", ...
#> $ ident            <fct> 0, 2, 0, 0, 0, 1, 0, 0, 0, 2, 0, 1, 1, 1, 0, 1, 2,...
#> $ UMAP1            <dbl> 1.42426342, 4.41265806, 3.12128718, 1.65304247, 1....
#> $ UMAP2            <dbl> 1.2632623, 1.4619932, -5.3615592, -3.3089443, 0.88...
#> $ centerX          <dbl> 1.390433, 4.982767, 1.390433, 1.390433, 1.390433, ...
#> $ centerY          <dbl> -3.395173, 1.284410, -3.395173, -3.395173, -3.3951...

# t-SNE gene expression
x <- fetchTSNEExpressionData(object, genes = genes)
glimpse(x)

#> Observations: 500
#> Variables: 27
#> $ nGene            <int> 289, 405, 423, 337, 350, 378, 442, 217, 332, 405, ...
#> $ nUMI             <int> 1199, 2766, 4399, 1485, 1771, 2208, 4095, 1406, 15...
#> $ nCount           <int> 25114, 55337, 71802, 34607, 34005, 44760, 76590, 3...
#> $ sampleID         <fct> multiplexed_AAAAAAAA, multiplexed_AAAAAAAA, multip...
#> $ nCoding          <int> 1105, 2385, 3864, 1296, 1574, 1599, 3630, 695, 117...
#> $ nMito            <int> 24, 241, 234, 103, 83, 452, 234, 556, 246, 225, 28...
#> $ log10GenesPerUMI <dbl> 0.7992992, 0.7575732, 0.7208579, 0.7969256, 0.7832...
#> $ mitoRatio        <dbl> 0.02001668, 0.08712943, 0.05319391, 0.06936027, 0....
#> $ sampleName       <fct> rep_1, rep_1, rep_1, rep_1, rep_1, rep_1, rep_1, r...
#> $ fileName         <fct> multiplexed.fastq.gz, multiplexed.fastq.gz, multip...
#> $ description      <fct> multiplexed-AAAAAAAA, multiplexed-AAAAAAAA, multip...
#> $ index            <fct> 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1,...
#> $ sequence         <fct> TTTTTTTT, TTTTTTTT, TTTTTTTT, TTTTTTTT, TTTTTTTT, ...
#> $ aggregate        <fct> sample, sample, sample, sample, sample, sample, sa...
#> $ revcomp          <fct> AAAAAAAA, AAAAAAAA, AAAAAAAA, AAAAAAAA, AAAAAAAA, ...
#> $ orig.ident       <fct> multiplexed, multiplexed, multiplexed, multiplexed...
#> $ res.0.4          <chr> "0", "2", "0", "0", "0", "1", "0", "0", "0", "2", ...
#> $ res.0.8          <chr> "3", "2", "1", "1", "3", "0", "1", "1", "3", "2", ...
#> $ res.1.2          <chr> "4", "0", "1", "1", "4", "2", "1", "1", "4", "0", ...
#> $ ident            <fct> 0, 2, 0, 0, 0, 1, 0, 0, 0, 2, 0, 1, 1, 1, 0, 1, 2,...
#> $ tSNE_1           <dbl> -17.8005980, 14.3021396, 5.5353587, 5.5824008, -17...
#> $ tSNE_2           <dbl> -8.7635904, -1.2800499, -15.9179439, -9.7737431, -...
#> $ centerX          <dbl> -1.020748, 13.796989, -1.020748, -1.020748, -1.020...
#> $ centerY          <dbl> -9.92682961, 0.06908376, -9.92682961, -9.92682961,...
#> $ mean             <dbl> 2.8333333, 1.8333333, 5.0000000, 1.3333333, 2.8333...
#> $ median           <dbl> 1.0, 1.0, 2.0, 1.0, 1.5, 2.0, 4.0, 1.5, 1.5, 3.0, ...
#> $ sum              <dbl> 17, 11, 30, 8, 17, 14, 34, 12, 10, 26, 14, 3, 10, ...

# UMAP gene expession
x <- fetchUMAPExpressionData(object, genes = genes)
glimpse(x)

#> Observations: 500
#> Variables: 27
#> $ nGene            <int> 289, 405, 423, 337, 350, 378, 442, 217, 332, 405, ...
#> $ nUMI             <int> 1199, 2766, 4399, 1485, 1771, 2208, 4095, 1406, 15...
#> $ nCount           <int> 25114, 55337, 71802, 34607, 34005, 44760, 76590, 3...
#> $ sampleID         <fct> multiplexed_AAAAAAAA, multiplexed_AAAAAAAA, multip...
#> $ nCoding          <int> 1105, 2385, 3864, 1296, 1574, 1599, 3630, 695, 117...
#> $ nMito            <int> 24, 241, 234, 103, 83, 452, 234, 556, 246, 225, 28...
#> $ log10GenesPerUMI <dbl> 0.7992992, 0.7575732, 0.7208579, 0.7969256, 0.7832...
#> $ mitoRatio        <dbl> 0.02001668, 0.08712943, 0.05319391, 0.06936027, 0....
#> $ sampleName       <fct> rep_1, rep_1, rep_1, rep_1, rep_1, rep_1, rep_1, r...
#> $ fileName         <fct> multiplexed.fastq.gz, multiplexed.fastq.gz, multip...
#> $ description      <fct> multiplexed-AAAAAAAA, multiplexed-AAAAAAAA, multip...
#> $ index            <fct> 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1,...
#> $ sequence         <fct> TTTTTTTT, TTTTTTTT, TTTTTTTT, TTTTTTTT, TTTTTTTT, ...
#> $ aggregate        <fct> sample, sample, sample, sample, sample, sample, sa...
#> $ revcomp          <fct> AAAAAAAA, AAAAAAAA, AAAAAAAA, AAAAAAAA, AAAAAAAA, ...
#> $ orig.ident       <fct> multiplexed, multiplexed, multiplexed, multiplexed...
#> $ res.0.4          <chr> "0", "2", "0", "0", "0", "1", "0", "0", "0", "2", ...
#> $ res.0.8          <chr> "3", "2", "1", "1", "3", "0", "1", "1", "3", "2", ...
#> $ res.1.2          <chr> "4", "0", "1", "1", "4", "2", "1", "1", "4", "0", ...
#> $ ident            <fct> 0, 2, 0, 0, 0, 1, 0, 0, 0, 2, 0, 1, 1, 1, 0, 1, 2,...
#> $ UMAP1            <dbl> 1.42426342, 4.41265806, 3.12128718, 1.65304247, 1....
#> $ UMAP2            <dbl> 1.2632623, 1.4619932, -5.3615592, -3.3089443, 0.88...
#> $ centerX          <dbl> 1.390433, 4.982767, 1.390433, 1.390433, 1.390433, ...
#> $ centerY          <dbl> -3.395173, 1.284410, -3.395173, -3.395173, -3.3951...
#> $ mean             <dbl> 2.8333333, 1.8333333, 5.0000000, 1.3333333, 2.8333...
#> $ median           <dbl> 1.0, 1.0, 2.0, 1.0, 1.5, 2.0, 4.0, 1.5, 1.5, 3.0, ...
#> $ sum              <dbl> 17, 11, 30, 8, 17, 14, 34, 12, 10, 26, 14, 3, 10, ...