Fetch t-SNE Locations, Cellular Metadata and Expression of Genes

This gets t-SNE locations, cellular metadata, expression of genes and the geometric mean of the gene expression from a seurat object.

fetchTSNEExpressionData(object, ...)

# S4 method for seurat
fetchTSNEExpressionData(object, genes)

Arguments

object
Object.

...
Additional arguments (for the S4 generic definition).

genes
Genes identifiers (matching the rownames in the object), of which to get expression data.

Value

data.frame containing tSNE coordinates, sample metadata, and aggregate marker expression values (mean, median, and sum).

See also

Other t-SNE Utilities: fetchTSNEData, plotMarkerTSNE, plotTSNE

Examples

load(system.file(
    file.path("extdata", "seurat.rda"),
    package = "bcbioSingleCell"))

genes <- counts(seurat) %>% rownames() %>% head()
print(genes)

#> ENSG00000000457 ENSG00000002586 ENSG00000002933 ENSG00000004939 ENSG00000005022 
#>         "SCYL3"          "CD99"      "TMEM176A"        "SLC4A1"       "SLC25A5" 
#> ENSG00000005100 
#>         "DHX33" 

fetchTSNEExpressionData(seurat, genes = genes) %>%
    glimpse()

#> Observations: 243
#> Variables: 21
#> $ tSNE1             <dbl> -3.055881, 13.310006, 13.533313, 3.831661, -8.180...
#> $ tSNE2             <dbl> -4.30397137, -1.86121673, 2.32827003, 1.36239831,...
#> $ ident             <fct> 0, 1, 1, 3, 2, 1, 3, 0, 1, 2, 3, 0, 1, 2, 1, 0, 3...
#> $ nGene             <int> 734, 774, 541, 545, 653, 816, 525, 725, 614, 566,...
#> $ nUMI              <int> 6220, 9622, 1896, 2137, 4918, 9247, 1891, 6031, 4...
#> $ sampleID          <fct> M1, M1, M1, M1, M1, M1, M1, M1, M1, M1, M1, M1, M...
#> $ nCoding           <int> 5495, 8536, 1682, 1908, 3770, 8336, 1484, 5280, 3...
#> $ nMito             <int> 442, 438, 103, 83, 830, 423, 246, 436, 543, 208, ...
#> $ log10GenesPerUMI  <dbl> 0.7553649, 0.7252193, 0.8338414, 0.8217889, 0.762...
#> $ mitoRatio         <dbl> 0.07106109, 0.04552068, 0.05432489, 0.03883949, 0...
#> $ nCount            <int> 95155, 124786, 34607, 34005, 77663, 132228, 41090...
#> $ sampleName        <fct> M1, M1, M1, M1, M1, M1, M1, M1, M1, M1, M1, M1, M...
#> $ description       <fct> M1, M1, M1, M1, M1, M1, M1, M1, M1, M1, M1, M1, M...
#> $ interestingGroups <fct> M1, M1, M1, M1, M1, M1, M1, M1, M1, M1, M1, M1, M...
#> $ origIdent         <fct> M1, M1, M1, M1, M1, M1, M1, M1, M1, M1, M1, M1, M...
#> $ res0x8            <fct> 0, 1, 1, 3, 2, 1, 3, 0, 1, 2, 3, 0, 1, 2, 1, 0, 3...
#> $ centerX           <dbl> -6.650639, 13.494614, 13.494614, 2.638557, -6.678...
#> $ centerY           <dbl> -4.2258378, -1.0635274, -1.0635274, 0.8535128, 9....
#> $ mean              <dbl> 1.164669, 1.402145, 1.388684, 1.420361, 2.096435,...
#> $ median            <dbl> 1.360158, 1.474219, 1.836456, 1.736856, 2.186520,...
#> $ sum               <dbl> 6.988012, 8.412870, 8.332102, 8.522164, 12.578611...