Filter Cells

Filter Cells

# S4 method for SingleCellExperiment
filterCells(object, nCells = Inf,
  minUMIs = 1L, maxUMIs = Inf, minGenes = 1L, maxGenes = Inf,
  minNovelty = 0L, maxMitoRatio = 1L, minCellsPerGene = 1L)

Arguments

object
Object.

nCells
scalar integer. Expected number of cells per sample.

minUMIs
scalar integer. Minimum number of UMI disambiguated counts per cell.

maxUMIs
scalar integer. Maximum number of UMI disambiguated counts per cell.

minGenes
scalar integer. Minimum number of genes detected.

maxGenes
scalar integer. Maximum number of genes detected.

minNovelty
scalar integer (0-1). Minimum novelty score (log10 genes per UMI).

maxMitoRatio
scalar integer (0-1). Maximum relative mitochondrial abundance.

minCellsPerGene
scalar integer. Include genes with non-zero expression in at least this many cells.

Value

SingleCellExperiment, with filtering information slotted into S4Vectors::metadata() as filterCells and filterParams.

Details

Apply gene detection, novelty score, and mitochondrial abundance cutoffs to cellular barcodes. By default we recommend applying the same filtering cutoff to all samples. The filtering parameters now support per-sample cutoffs, defined using a named numeric vector. When matching per sample, be sure to use the Biobase::sampleNames() return values (i.e. the sampleName column in basejump::sampleData()).

Examples

data(indrops)

object <- indrops
show(object)

#> bcbioSingleCell 0.3.2
#> uploadDir: /data01/git/bcbioSingleCell/inst/extdata/indrops
#> dates(2): [bcbio] 2018-01-01; [R] 2018-11-16
#> level: genes
#> sampleMetadataFile: inst/extdata/indrops/metadata.csv
#> organism: Homo sapiens
#> interestingGroups: sampleName
#> filtered: FALSE
#> class: SingleCellExperiment 
#> dim: 50 100 
#> metadata(27): version pipeline ... wd sessionInfo
#> assays(1): counts
#> rownames(50): ENSG00000071082 ENSG00000100316 ... ENSG00000269028
#>   ENSG00000282105
#> rowData names(4): broadClass geneBiotype geneID geneName
#> colnames(100): AAACACTA_CTTCGATT AAACTACA_CCACATTA ...
#>   TGGGAATT_ATATAGGA TGTTATCA_ACGCAGAG
#> colData names(15): nUMI nGene ... aggregate revcomp
#> reducedDimNames(0):
#> spikeNames(0):

x <- filterCells(object)

#> Parameters:
#>   - >= 1 UMIs per cell
#>   - <= Inf UMIs per cell
#>   - >= 1 genes per cell
#>   - <= Inf genes per cell
#>   - >= 0 novelty score
#>   - <= 1 mitochondrial abundance
#>   - == Inf cells per sample
#>   - >= 1 cells per gene
#> ────────────────────────────────────────────────────────────────────────
#> Cells:
#>      100 cells | prefilter
#>      100 cells | minUMIs >= 1
#>      100 cells | maxUMIs <= Inf
#>      100 cells | minGenes >= 1
#>      100 cells | maxGenes <= Inf
#>      100 cells | minNovelty <= 0
#>      100 cells | maxMitoRatio <= 1
#>      100 cells | nCells == Inf
#>      100 of 100 cells passed filtering (100%)
#> rep_1 
#>   100 
#> ────────────────────────────────────────────────────────────────────────
#> Genes:
#>       50 genes | prefilter
#>       50 genes | minCellsPerGene >= 1
#>       50 of 50 genes passed filtering (100%)

show(x)

#> bcbioSingleCell 0.3.2
#> uploadDir: /data01/git/bcbioSingleCell/inst/extdata/indrops
#> dates(2): [bcbio] 2018-01-01; [R] 2018-11-16
#> level: genes
#> sampleMetadataFile: inst/extdata/indrops/metadata.csv
#> organism: Homo sapiens
#> interestingGroups: sampleName
#> filtered: TRUE
#> class: SingleCellExperiment 
#> dim: 50 100 
#> metadata(30): version pipeline ... filterParams filterSummary
#> assays(1): counts
#> rownames(50): ENSG00000071082 ENSG00000100316 ... ENSG00000269028
#>   ENSG00000282105
#> rowData names(4): broadClass geneBiotype geneID geneName
#> colnames(100): AAACACTA_CTTCGATT AAACTACA_CCACATTA ...
#>   TGGGAATT_ATATAGGA TGTTATCA_ACGCAGAG
#> colData names(15): nUMI nGene ... aggregate revcomp
#> reducedDimNames(0):
#> spikeNames(0):

metadata(x)$filterParams

#> $nCells
#> [1] Inf
#> 
#> $minUMIs
#> [1] 1
#> 
#> $maxUMIs
#> [1] Inf
#> 
#> $minGenes
#> [1] 1
#> 
#> $maxGenes
#> [1] Inf
#> 
#> $minNovelty
#> [1] 0
#> 
#> $maxMitoRatio
#> [1] 1
#> 
#> $minCellsPerGene
#> [1] 1
#> 

## Per sample cutoffs.
sampleNames(object)

#> multiplexed_AAAAAAAA 
#>              "rep_1" 

x <- filterCells(
    object = object,
    minUMIs = c(rep_1 = 100)
)

#> minUMIs: per sample mode
#> rep_1 
#>   100 

#> Parameters:
#>   - >= 100 UMIs per cell
#>   - <= Inf UMIs per cell
#>   - >= 1 genes per cell
#>   - <= Inf genes per cell
#>   - >= 0 novelty score
#>   - <= 1 mitochondrial abundance
#>   - == Inf cells per sample
#>   - >= 1 cells per gene
#> ────────────────────────────────────────────────────────────────────────
#> Cells:
#>      100 cells | prefilter
#>      100 cells | minUMIs >= 100
#>      100 cells | maxUMIs <= Inf
#>      100 cells | minGenes >= 1
#>      100 cells | maxGenes <= Inf
#>      100 cells | minNovelty <= 0
#>      100 cells | maxMitoRatio <= 1
#>      100 cells | nCells == Inf
#>      100 of 100 cells passed filtering (100%)
#> rep_1 
#>   100 
#> ────────────────────────────────────────────────────────────────────────
#> Genes:
#>       50 genes | prefilter
#>       50 genes | minCellsPerGene >= 1
#>       50 of 50 genes passed filtering (100%)

show(x)

#> bcbioSingleCell 0.3.2
#> uploadDir: /data01/git/bcbioSingleCell/inst/extdata/indrops
#> dates(2): [bcbio] 2018-01-01; [R] 2018-11-16
#> level: genes
#> sampleMetadataFile: inst/extdata/indrops/metadata.csv
#> organism: Homo sapiens
#> interestingGroups: sampleName
#> filtered: TRUE
#> class: SingleCellExperiment 
#> dim: 50 100 
#> metadata(30): version pipeline ... filterParams filterSummary
#> assays(1): counts
#> rownames(50): ENSG00000071082 ENSG00000100316 ... ENSG00000269028
#>   ENSG00000282105
#> rowData names(4): broadClass geneBiotype geneID geneName
#> colnames(100): AAACACTA_CTTCGATT AAACTACA_CCACATTA ...
#>   TGGGAATT_ATATAGGA TGTTATCA_ACGCAGAG
#> colData names(15): nUMI nGene ... aggregate revcomp
#> reducedDimNames(0):
#> spikeNames(0):

metadata(x)$filterParams

#> $nCells
#> [1] Inf
#> 
#> $minUMIs
#> rep_1 
#>   100 
#> 
#> $maxUMIs
#> [1] Inf
#> 
#> $minGenes
#> [1] 1
#> 
#> $maxGenes
#> [1] Inf
#> 
#> $minNovelty
#> [1] 0
#> 
#> $maxMitoRatio
#> [1] 1
#> 
#> $minCellsPerGene
#> [1] 1
#>