General Arguments

General Arguments

Arguments

aspectRatio
Aspect ratio.

color
Desired ggplot color scale. Must supply discrete values. When set to NULL, the default ggplot2 color palette will be used. If manual color definitions are desired, we recommend using ggplot2::scale_color_manual().

fill
Desired ggplot fill scale. Must supply discrete values. When set to NULL, the default ggplot2 color palette will be used. If manual color definitions are desired, we recommend using ggplot2::scale_fill_manual().

gene2symbol
data.frame containing gene-to-symbol mappings. Columns must contain geneID and geneName.

dark
Plot against a dark background using basejump::theme_midnight().

dir
Output directory.

genes
Gene identifiers. Must match the rownames of the object.

geom
Plot type. Uses match.arg() internally and defaults to the first argument in the vector.

grid
Show major grid lines but hide axis lines.

headerLevel
R Markdown header level.

interestingGroups
Character vector of interesting groups. Must be formatted in camel case and intersect with sampleData() colnames.

label
Overlay a cluster identitiy label on the plot.

labelSize
Size of the text label.

legend
Include plot legend.

min
Recommended minimum value cutoff.

max
Recommended maximum value cutoff.

object
Object.

pipeline
Pipeline used to generate the samples.

pointAlpha
Alpha transparency level. Useful when there many cells in the dataset, and some cells can be masked.

pointsAsNumbers
Plot the points as numbers (TRUE) or dots (FALSE).

pointSize
Cell point size.

return
Return type. Uses base::match.arg() internally and defaults to the first argument in the vector.

title
Plot title.

trans
Name of the axis scale transformation to apply. See help("scale_x_continuous", "ggplot2") for more information.

trendline
Include trendline on plot.

value
Object to assign.

x
Primary object.

y
Secondary object.

...
Additional arguments.

Value

No value.