Read Sparse Counts

Read single-cell RNA-seq counts saved as a MatrixMarket file into a sparse counts matrix (dgCMatrix). Transcript (rownames) and molecular barcodes (colnames) identifiers are set from the corresponding dependency files automatically.

.readSparseCounts(sampleDir, pipeline = "bcbio")

Arguments

sampleDir
Named character vector of sample directory containing the MatrixMart file.

pipeline
Pipeline used to generate the MatrixMarket file. Defaults to bcbio-nextgen (bcbio). Also supports 10X Chromium CellRanger (cellranger).

Value

dgCMatrix.

Details

This function supports automatic handling of compressed matrix files.

Note

bcbio-nextgen outputs counts at transcript level. 10X Chromium CellRanger outputs counts at gene level.

See also

help("dgCMatrix-class")