Known Markers Detected

Known Markers Detected

knownMarkersDetected(object, ...)

# S4 method for grouped_df
knownMarkersDetected(object, known, alpha = 0.05,
  filterPromiscuous = FALSE, promiscuousCutoff = 5L)

Arguments

object
All markers data, generated by Seurat::FindAllMarkers() return and sanitized with sanitizeMarkers().

...
Additional arguments.

known
Known markers data.frame imported by readCellTypeMarkers() or pulled from internal cellCycleMarkers data.

alpha
Alpha cutoff (adjusted P value; false discovery rate).

filterPromiscuous
Remove genes with poor specificity, that are present in as many clusters as defined by promiscuousCutoff.

promiscuousCutoff
Minimum number of clusters required to consider a gene marker promiscuous.

Value

grouped_df, grouped by "cellType" column.

Note

Both the all and known objects must contain Ensembl gene identifiers in the geneID column. We must avoid any matching operations based on the gene names, since these change often and can mismatch easily.

See also

Examples

# grouped_df ====
x <- knownMarkersDetected(
    object = all_markers_small,
    known = cellTypeMarkers[["homoSapiens"]]
)
head(x)

#> # A tibble: 6 x 14
#> # Groups:   cellType [1]
#>   cellType cluster geneID        geneName avgLogFC  pct1  pct2 rowname   pvalue
#>   <fct>    <fct>   <chr>         <chr>       <dbl> <dbl> <dbl> <chr>      <dbl>
#> 1 B Cell   4       ENSG00000211… IGHM        2.16  0.933 0.191 IGHM    1.87e-37
#> 2 B Cell   4       ENSG00000211… IGLC2       3.87  0.733 0.268 IGLC2   6.53e-16
#> 3 B Cell   4       ENSG00000211… IGKC        3.02  0.778 0.385 IGKC    1.17e-15
#> 4 B Cell   5       ENSG00000211… IGKC        0.946 1     0.393 IGKC    6.80e-14
#> 5 B Cell   4       ENSG00000211… IGLC3       3.14  0.444 0.099 IGLC3   2.46e-12
#> 6 B Cell   5       ENSG00000211… IGHM        0.762 0.864 0.23  IGHM    5.55e-12
#> # ... with 5 more variables: padj <dbl>, geneBiotype <fct>, description <chr>,
#> #   seqCoordSystem <fct>, broadClass <fct>