Known Markers Example

Known Markers Example

known_markers_small

Format

An object of class grouped_df (inherits from tbl_df, tbl, data.frame) with 29 rows and 19 columns.

See also

Examples

glimpse(known_markers_small)

#> Observations: 29
#> Variables: 19
#> $ cellType       <fct> B Cell, B Cell, B Cell, B Cell, B Cell, B Cell, Dend...
#> $ cluster        <fct> 3, 3, 3, 3, 1, 0, 1, 1, 1, 1, 1, 4, 4, 4, 4, 4, 0, 0...
#> $ geneID         <chr> "ENSG00000211899", "ENSG00000211592", "ENSG000002116...
#> $ geneName       <fct> IGHM, IGKC, IGLC2, IGLC3, IGKC, IGHM, HLA-DRA, HLA-D...
#> $ avgLogFC       <dbl> 2.0535014, 3.0662657, 3.4738857, 2.7277204, -1.99226...
#> $ pct1           <dbl> 0.900, 0.843, 0.614, 0.329, 0.290, 0.159, 0.758, 0.4...
#> $ pct2           <dbl> 0.153, 0.351, 0.260, 0.098, 0.463, 0.307, 1.000, 1.0...
#> $ rowname        <chr> "IGHM", "IGKC", "IGLC2", "IGLC3", "IGKC", "IGHM", "H...
#> $ pvalue         <dbl> 1.534176e-50, 1.696343e-29, 3.274191e-14, 9.446819e-...
#> $ padj           <dbl> 7.670879e-48, 8.481717e-27, 1.637096e-11, 4.723409e-...
#> $ seqnames       <fct> 14, 2, 22, 22, 2, 14, 6, 6, 6, 6, 6, 6, 6, 6, 6, 6, ...
#> $ start          <int> 105851708, 88857161, 22900976, 22906342, 88857161, 1...
#> $ end            <int> 105856218, 88857683, 22901437, 22906803, 88857683, 1...
#> $ width          <int> 4511, 523, 462, 462, 523, 4511, 5205, 11080, 8917, 1...
#> $ strand         <fct> -, -, +, +, -, -, +, -, -, +, -, +, -, +, -, -, -, -...
#> $ broadClass     <fct> ig, ig, ig, ig, ig, ig, coding, coding, coding, codi...
#> $ description    <fct> immunoglobulin heavy constant mu [Source:HGNC Symbol...
#> $ geneBiotype    <fct> IG_C_gene, IG_C_gene, IG_C_gene, IG_C_gene, IG_C_gen...
#> $ seqCoordSystem <fct> chromosome, chromosome, chromosome, chromosome, chro...