Load 10X Genomics CellRanger Data

Read 10x Genomics Chromium cell counts from barcodes.tsv, genes.tsv, and matrix.mtx files.

loadCellRanger(uploadDir, refdataDir, interestingGroups = "sampleName",
  sampleMetadataFile = NULL, annotable = TRUE, prefilter = TRUE, ...)

Arguments

uploadDir
Path to CellRanger output directory. This directory path must contain filtered_gene_bc_matrices* as a child directory.

refdataDir
Directory path to cellranger reference annotation data.

interestingGroups
Character vector of interesting groups. First entry is used for plot colors during quality control (QC) analysis. Entire vector is used for PCA and heatmap QC functions.

sampleMetadataFile
Sample barcode metadata file. Optional for runs with demultiplixed index barcodes (e.g. SureCell), but otherwise required for runs with multipliexed FASTQs containing multiple index barcodes (e.g. inDrop). Consult the GitHub repo for examples and additional information.

annotable
User-defined gene annotations (a.k.a. "annotable"), which will be slotted into rowData(). Typically this should be left set to TRUE. By default, the function will automatically generate an annotable from the annotations available on Ensembl. If set FALSE, then rowData() inside the resulting bcbioSingleCell object will be left empty. This is recommended for projects dealing with genes or transcripts that are poorly annotated. Additionally, a manually constructed annotable can be passed in as a data.frame, but this isn't generally recommended.

prefilter
Prefilter counts prior to quality control analysis.

...
Additional arguments, to be stashed in the metadata() slot.

Value

bcbioSingleCell.

Details

This function is a simplified version of loadSingleCell() optimized for handling CellRanger output.

Note

Unlike loadSingleCell(), the organism, ensemblVersion, and genomeBuild are always detected automatically, based on the refdataDir YAML metadata. Therefore, these parameters cannot be set by the user.

Examples

extdataDir <- system.file("extdata", package = "bcbioSingleCell")
uploadDir <- file.path(extdataDir, "cellranger")
refdataDir <- file.path(extdataDir, "refdata-cellranger-hg19-1.2.0")
sampleMetadataFile <- file.path(extdataDir, "cellranger.csv")
bcb <- loadCellRanger(
    uploadDir = uploadDir,
    refdataDir = refdataDir,
    sampleMetadataFile = sampleMetadataFile)

#> CellRanger output directory structure:
#> <uploadDir>/<sampleName>/outs/filtered_gene_bc_matrices*/<genomeBuild>/matrix.mtx

#> 1 samples detected

#> Reading cellranger.csv

#> Parsed with column specification:
#> cols(
#>   fileName = col_character(),
#>   description = col_character(),
#>   sampleName = col_character(),
#>   index = col_integer()
#> )

#> Organism: Homo sapiens
#> Genome build: hg19
#> Ensembl release: 82

#> Minimal working example doesn't contain a GTF file

#> Obtaining gene2symbol from Ensembl

#> Reading counts at gene level

#> Calculating barcode metrics

#> 500 cellular barcodes detected

#> 11 mitochondrial genes detected

#> 500 cellular barcodes passed pre-filtering (100%)

print(bcb)

#> class: bcbioSingleCell 
#> dim: 500 500 
#> metadata(23): version pipeline ... utilsSessionInfo devtoolsSessionInfo
#> assays(1): assay
#> rownames(500): ENSG00000005022 ENSG00000008018 ... ENSG00000265681
#>   ENSG00000269893
#> rowData names(10): ensgene symbol ... strand entrez
#> colnames(500): aggregation_AAACCTGGTTTACTCT_1
#>   aggregation_AAACGGGGTATCTGCA_1 ... aggregation_TTTGTCACACAACGTT_2
#>   aggregation_TTTGTCACACTCGACG_2
#> colData names(6): nUMI nGene ... log10GenesPerUMI mitoRatio