Load bcbio Single-Cell RNA-Seq Run

Load bcbio Single-Cell RNA-Seq Run

loadRun(object, ...)

# S4 method for character
loadRun(object = "final", sampleMetadataFile,
  wellMetadataFile = NULL, interestingGroups = "sampleName",
  tx2gene = NULL, ...)

Arguments

object
Path to final upload directory. This path is set when running bcbio_nextgen -w template.

...
Additional arguments, passed as metadata.

sampleMetadataFile
Required. Sample barcode metadata file.

wellMetadataFile
Optional. Well identifier metadata file.

interestingGroups
Character vector of interesting groups. First entry is used for plot colors during quality control (QC) analysis. Entire vector is used for PCA and heatmap QC functions.

tx2gene
tx2gene data.frame, required when using a custom transcriptome FASTA file. This argument will be removed in a future update.

Value

bcbioSCDataSet.

Details

cellranger: Read 10x GenomicsChromium cell counts from barcodes.tsv, genes.tsv, and matrix.mtx files.

Note

When working in RStudio, we recommend connecting to the bcbio-nextgen run directory as a remote connection over sshfs.

See also

Other `bcbio` Utilities: bcbio, download