Load bcbio Single-Cell RNA-Seq Data

Load bcbio Single-Cell RNA-Seq Data

loadSingleCell(uploadDir, sampleMetadataFile = NULL,
  interestingGroups = "sampleName", gtfFile = NULL, annotable = TRUE,
  organism = NULL, ensemblVersion = NULL, genomeBuild = NULL,
  prefilter = TRUE, ...)

Arguments

uploadDir
Path to final upload directory. This path is set when running bcbio_nextgen -w template.

sampleMetadataFile
Sample barcode metadata file. Optional for runs with demultiplixed index barcodes (e.g. SureCell), but otherwise required for runs with multipliexed FASTQs containing multiple index barcodes (e.g. inDrop). Consult the GitHub repo for examples and additional information.

interestingGroups
Character vector of interesting groups. First entry is used for plot colors during quality control (QC) analysis. Entire vector is used for PCA and heatmap QC functions.

gtfFile
Optional but recommended. GTF (Gene Transfer Format) file, which will be used for gene-to-symbol (gene2symbol) and transcript-to-gene (tx2gene) annotation mappings.

annotable
User-defined gene annotations (a.k.a. "annotable"), which will be slotted into rowData(). Typically this should be left set to TRUE. By default, the function will automatically generate an annotable from the annotations available on Ensembl. If set FALSE, then rowData() inside the resulting bcbioSingleCell object will be left empty. This is recommended for projects dealing with genes or transcripts that are poorly annotated. Additionally, a manually constructed annotable can be passed in as a data.frame, but this isn't generally recommended.

organism
Optional. Organism name. Use the full latin name (e.g. "Homo sapiens"), since this will be input downstream to AnnotationHub/ensembldb. If set, this genome must be supported on Ensembl. Normally this can be left NULL, and the function will attempt to detect the organism automatically using detectOrganism().

ensemblVersion
Optional. Ensembl release version. If NULL, defaults to current release, and does not typically need to be user-defined. This parameter can be useful for matching Ensembl annotations against an outdated bcbio annotation build.

genomeBuild
Optional. Genome build. Normally this can be left NULL and the build will be detected from the bcbio run data. This can be set manually (e.g. "hg19" for the older Homo sapiens reference genome). Note that this must match the genome build identifier on Ensembl for annotations to download correctly.

prefilter
Prefilter counts prior to quality control analysis.

...
Additional arguments, to be stashed in the metadata() slot.

Value

bcbioSingleCell.

Note

When working in RStudio, we recommend connecting to the bcbio-nextgen run directory as a remote connection over sshfs.

Examples

# Homo sapiens
# Minimal working example dataset
extdataDir <- system.file("extdata", package = "bcbioSingleCell")
uploadDir <- file.path(extdataDir, "harvard_indrop_v3")
sampleMetadataFile <- file.path(extdataDir, "harvard_indrop_v3.xlsx")
bcb <- loadSingleCell(
    uploadDir = uploadDir,
    sampleMetadataFile = sampleMetadataFile)

#> 2017-11-10_bcbio

#> 2 samples detected

#> Reading project-summary.yaml

#> Reading log files

#> Warning: bcbio-nextgen.log missing

#> Warning: bcbio-nextgen-commands.log missing

#> Reading data and program versions

#> Parsed with column specification:
#> cols(
#>   genome = col_character(),
#>   resource = col_character(),
#>   version = col_character()
#> )

#> Organism: Homo sapiens
#> Genome build: hg38

#> Warning: Assuming unknown UMI is harvard-indrop-v3 (default)

#> Reading harvard_indrop_v3.xlsx

#> Loading Ensembl annotations from AnnotationHub
#> 2017-10-27

#> EnsDB AH57757: Homo sapiens Ensembl 90

#> Reading cellular barcode distributions

#>    |                                                  | 0 % ~calculating  |==                                                                      |   4%|=====                                                                   |   8%|========                                                                |  11%|===========                                                             |  15%|==============                                                          |  19%|=================                                                       |  23%|===================                                                     |  27%|======================                                                  |  31%|=========================                                               |  34%|========================                                        |  38%    1 MB|===========================                                     |  42%    1 MB|=============================                                   |  45%    1 MB|================================                                |  49%    1 MB|==================================                              |  53%    1 MB|====================================                            |  56%    1 MB|=======================================                         |  60%    1 MB|=========================================                       |  63%    1 MB|===========================================                     |  67%    1 MB|=============================================                   |  70%    1 MB|================================================                |  74%    2 MB|==================================================              |  77%    2 MB|====================================================            |  81%    2 MB|=======================================================         |  84%    2 MB|=========================================================       |  88%    2 MB|===========================================================     |  91%    2 MB|=============================================================   |  95%    2 MB|================================================================|  98%    2 MB|=================================================================| 100%    2 MB
#>    |+++++++++++++++++++++++++                         | 50% ~00s             |++++++++++++++++++++++++++++++++++++++++++++++++++| 100% elapsed = 00s

#> Reading counts at gene level

#>    |                                                  | 0 % ~calculating  

#> run1_AGAGGATA

#>    |+++++++++++++++++++++++++                         | 50% ~06s          

#> run2_AGAGGATA

#>    |++++++++++++++++++++++++++++++++++++++++++++++++++| 100% elapsed = 10s

#> Error in (function (classes, fdef, mtable) {    methods <- .findInheritedMethods(classes, fdef, mtable)    if (length(methods) == 1L)         return(methods[[1L]])    else if (length(methods) == 0L) {        cnames <- paste0("\"", vapply(classes, as.character,             ""), "\"", collapse = ", ")        stop(gettextf("unable to find an inherited method for function %s for signature %s",             sQuote(fdef@generic), sQuote(cnames)), domain = NA)    }    else stop("Internal error in finding inherited methods; didn't return a unique method",         domain = NA)})(list(structure("dgCMatrix", package = "Matrix")), structure(function (object,     ...) {    standardGeneric("calculateMetrics")}, generic = structure("calculateMetrics", package = "bcbioSingleCell"), package = "bcbioSingleCell", group = list(), valueClass = character(0), signature = "object", default = `\001NULL\001`, skeleton = (function (object,     ...) stop("invalid call in method dispatch to 'calculateMetrics' (no default method)",     domain = NA))(object, ...), class = structure("nonstandardGenericFunction", package = "methods")),     <environment>): unable to find an inherited method for function 'calculateMetrics' for signature '"dgCMatrix"'

# Mus musculus
# Run with Ensembl 88 transcriptome FASTA and GTF files

# NOT RUN {
bcb <- loadSingleCell(
    uploadDir = file.path("indrop_rnaseq", "final"),
    interestingGroups = c("genotype", "treatment"),
    sampleMetadataFile = file.path("meta", "sample_metadata.xlsx"),
    gtfFile = file.path(
        "annotations",
        "Mus_musculus.GRCm38.88.chr_patch_hapl_scaff.gtf.gz"),
    ensemblVersion = 88)
# }