Sample Barcode Metrics

Sample Barcode Metrics

# S4 method for matrix
metrics(object, rowData = NULL, prefilter = TRUE)

# S4 method for dgCMatrix
metrics(object, rowData = NULL, prefilter = TRUE)

# S4 method for dgTMatrix
metrics(object, rowData = NULL, prefilter = TRUE)

# S4 method for SingleCellExperiment
metrics(object, interestingGroups)

# S4 method for seurat
metrics(object, ...)

Arguments

object
Object.

rowData
Data describing the rows of the object.

prefilter
Whether to apply pre-filtering to the cellular barcodes.

interestingGroups
Character vector of interesting groups. Must be formatted in camel case and intersect with sampleData() colnames.

...
Additional arguments.

Value

data.frame with cellular barcodes as rows.

See also

Examples

# bcbioSingleCell ====
metrics(indrops_small) %>% glimpse()

#> Observations: 500
#> Variables: 16
#> $ sampleID          <fct> multiplexed_AAAAAAAA, multiplexed_AAAAAAAA, multiplexed_AAAAAAAA, multiplexed_AAAAAA...
#> $ nUMI              <int> 1199, 2766, 4399, 1485, 1771, 2208, 4095, 1406, 1564, 2690, 1859, 1436, 997, 1987, 1...
#> $ nGene             <int> 289, 405, 423, 337, 350, 378, 442, 217, 332, 405, 351, 161, 221, 265, 231, 364, 330,...
#> $ nCoding           <int> 1105, 2385, 3864, 1296, 1574, 1599, 3630, 695, 1171, 2318, 1423, 415, 635, 903, 689,...
#> $ nMito             <int> 24, 241, 234, 103, 83, 452, 234, 556, 246, 225, 286, 766, 282, 900, 813, 208, 76, 91...
#> $ log10GenesPerUMI  <dbl> 0.7992992, 0.7575732, 0.7208579, 0.7969256, 0.7832195, 0.7707813, 0.7323467, 0.74220...
#> $ mitoRatio         <dbl> 0.020016681, 0.087129429, 0.053193908, 0.069360269, 0.046866177, 0.204710145, 0.0571...
#> $ sampleName        <fct> rep_1, rep_1, rep_1, rep_1, rep_1, rep_1, rep_1, rep_1, rep_1, rep_1, rep_1, rep_1, ...
#> $ fileName          <fct> multiplexed.fastq.gz, multiplexed.fastq.gz, multiplexed.fastq.gz, multiplexed.fastq....
#> $ description       <fct> multiplexed-AAAAAAAA, multiplexed-AAAAAAAA, multiplexed-AAAAAAAA, multiplexed-AAAAAA...
#> $ index             <fct> 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, ...
#> $ sequence          <fct> TTTTTTTT, TTTTTTTT, TTTTTTTT, TTTTTTTT, TTTTTTTT, TTTTTTTT, TTTTTTTT, TTTTTTTT, TTTT...
#> $ aggregate         <fct> sample, sample, sample, sample, sample, sample, sample, sample, sample, sample, samp...
#> $ revcomp           <fct> AAAAAAAA, AAAAAAAA, AAAAAAAA, AAAAAAAA, AAAAAAAA, AAAAAAAA, AAAAAAAA, AAAAAAAA, AAAA...
#> $ nCount            <int> 25114, 55337, 71802, 34607, 34005, 44760, 76590, 37290, 41090, 55103, 48440, 17616, ...
#> $ interestingGroups <fct> rep_1, rep_1, rep_1, rep_1, rep_1, rep_1, rep_1, rep_1, rep_1, rep_1, rep_1, rep_1, ...

# SingleCellExperiment ====
metrics(cellranger_small) %>% glimpse()

#> Observations: 500
#> Variables: 12
#> $ sampleID          <fct> aggregation_1, aggregation_1, aggregation_1, aggregation_1, aggregation_1, aggregati...
#> $ nUMI              <int> 31609, 36078, 72677, 32049, 28812, 59565, 32026, 45823, 63914, 28536, 66375, 31324, ...
#> $ nGene             <int> 495, 486, 495, 487, 489, 496, 484, 497, 492, 480, 496, 485, 496, 475, 479, 492, 495,...
#> $ nCoding           <int> 25193, 32709, 67262, 29640, 25629, 53611, 28885, 40676, 57717, 25682, 59972, 26791, ...
#> $ nMito             <int> 5353, 2968, 4710, 1824, 2426, 4937, 2188, 4375, 5200, 2614, 5690, 3397, 5557, 3071, ...
#> $ log10GenesPerUMI  <dbl> 0.5988263, 0.5895311, 0.5542862, 0.5964580, 0.6030417, 0.5644998, 0.5959036, 0.57848...
#> $ mitoRatio         <dbl> 0.16935050, 0.08226620, 0.06480730, 0.05691285, 0.08420103, 0.08288424, 0.06831949, ...
#> $ sampleName        <fct> proximal, proximal, proximal, proximal, proximal, proximal, proximal, proximal, prox...
#> $ fileName          <fct> aggregation.fastq.gz, aggregation.fastq.gz, aggregation.fastq.gz, aggregation.fastq....
#> $ description       <fct> aggregation-1, aggregation-1, aggregation-1, aggregation-1, aggregation-1, aggregati...
#> $ index             <fct> 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 2, 2, 1, 1, 1, 1, 1, 2, 2, 1, 2, 1, 2, 1, 2, ...
#> $ interestingGroups <fct> proximal, proximal, proximal, proximal, proximal, proximal, proximal, proximal, prox...

# dgCMatrix ====
counts <- counts(indrops_small)
class(counts)

#> [1] "dgCMatrix"
#> attr(,"package")
#> [1] "Matrix"

metrics(counts) %>% glimpse()

#> Calculating cellular barcode metrics

#> 500 cells detected

#> 500 / 500 cellular barcodes passed pre-filtering (100%)

#> Observations: 500
#> Variables: 6
#> $ nUMI             <int> 1199, 2766, 4399, 1485, 1771, 2208, 4095, 1406, 1564, 2690, 1859, 1436, 997, 1987, 17...
#> $ nGene            <int> 289, 405, 423, 337, 350, 378, 442, 217, 332, 405, 351, 161, 221, 265, 231, 364, 330, ...
#> $ nCoding          <int> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0...
#> $ nMito            <int> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0...
#> $ log10GenesPerUMI <dbl> 0.7992992, 0.7575732, 0.7208579, 0.7969256, 0.7832195, 0.7707813, 0.7323467, 0.742208...
#> $ mitoRatio        <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0...

# seurat ====
metrics(seurat_small) %>% glimpse()

#> Observations: 500
#> Variables: 19
#> $ nGene             <int> 289, 405, 423, 337, 350, 378, 442, 217, 332, 405, 351, 161, 221, 265, 231, 364, 330,...
#> $ nUMI              <int> 1199, 2766, 4399, 1485, 1771, 2208, 4095, 1406, 1564, 2690, 1859, 1436, 997, 1987, 1...
#> $ sampleID          <fct> multiplexed_AAAAAAAA, multiplexed_AAAAAAAA, multiplexed_AAAAAAAA, multiplexed_AAAAAA...
#> $ nCoding           <int> 1105, 2385, 3864, 1296, 1574, 1599, 3630, 695, 1171, 2318, 1423, 415, 635, 903, 689,...
#> $ nMito             <int> 24, 241, 234, 103, 83, 452, 234, 556, 246, 225, 286, 766, 282, 900, 813, 208, 76, 91...
#> $ log10GenesPerUMI  <dbl> 0.7992992, 0.7575732, 0.7208579, 0.7969256, 0.7832195, 0.7707813, 0.7323467, 0.74220...
#> $ mitoRatio         <dbl> 0.020016681, 0.087129429, 0.053193908, 0.069360269, 0.046866177, 0.204710145, 0.0571...
#> $ sampleName        <fct> rep_1, rep_1, rep_1, rep_1, rep_1, rep_1, rep_1, rep_1, rep_1, rep_1, rep_1, rep_1, ...
#> $ fileName          <fct> multiplexed.fastq.gz, multiplexed.fastq.gz, multiplexed.fastq.gz, multiplexed.fastq....
#> $ description       <fct> multiplexed-AAAAAAAA, multiplexed-AAAAAAAA, multiplexed-AAAAAAAA, multiplexed-AAAAAA...
#> $ index             <fct> 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, ...
#> $ sequence          <fct> TTTTTTTT, TTTTTTTT, TTTTTTTT, TTTTTTTT, TTTTTTTT, TTTTTTTT, TTTTTTTT, TTTTTTTT, TTTT...
#> $ aggregate         <fct> sample, sample, sample, sample, sample, sample, sample, sample, sample, sample, samp...
#> $ revcomp           <fct> AAAAAAAA, AAAAAAAA, AAAAAAAA, AAAAAAAA, AAAAAAAA, AAAAAAAA, AAAAAAAA, AAAAAAAA, AAAA...
#> $ nCount            <int> 25114, 55337, 71802, 34607, 34005, 44760, 76590, 37290, 41090, 55103, 48440, 17616, ...
#> $ interestingGroups <fct> rep_1, rep_1, rep_1, rep_1, rep_1, rep_1, rep_1, rep_1, rep_1, rep_1, rep_1, rep_1, ...
#> $ orig.ident        <fct> multiplexed, multiplexed, multiplexed, multiplexed, multiplexed, multiplexed, multip...
#> $ res.0.8           <chr> "4", "0", "1", "1", "4", "2", "1", "1", "4", "0", "1", "3", "3", "3", "1", "2", "0",...
#> $ ident             <fct> 4, 0, 1, 1, 4, 2, 1, 1, 4, 0, 1, 3, 3, 3, 1, 2, 0, 4, 0, 1, 2, 1, 3, 1, 0, 0, 4, 1, ...

metrics(Seurat::pbmc_small) %>% glimpse()

#> Error in .local(object, ...): is_subset : The element 'sampleName' in interestingGroups is not in colnames(object).