Plot UMIs per Cell

Plot the universal molecular identifiers (UMIs) per cell.

plotUMIsPerCell(object, ...)

# S4 method for bcbioSingleCell
plotUMIsPerCell(object, geom = "violin", min,
  interestingGroups, filterCells = FALSE, samplesOnYAxis = TRUE,
  fill = scale_fill_viridis(discrete = TRUE))

# S4 method for data.frame
plotUMIsPerCell(object, geom = "violin", min = 0,
  interestingGroups, multiplexed = FALSE, samplesOnYAxis = TRUE,
  fill = scale_fill_viridis(discrete = TRUE))

# S4 method for seurat
plotUMIsPerCell(object, geom = "violin", min,
  interestingGroups, multiplexed = FALSE, samplesOnYAxis = TRUE,
  fill = scale_fill_viridis(discrete = TRUE))

Arguments

object
Object.

...
Additional arguments (for the S4 generic definition).

geom
Plot type. Supported formats: boxplot, histogram, ridgeline, and violin.

min
Recommended minimum value cutoff.

interestingGroups
Interesting group, to use for colors.

filterCells
Show only the cells that have passed filtering cutoffs.

samplesOnYAxis
Plot the samples on the y axis. Doesn't apply to histogram.

fill
Desired ggplot fill scale. Defaults to viridis::scale_fill_viridis(). Must supply discrete values. When set to NULL, the default ggplot2 color palette will be used. If manual color definitions are desired, we recommend using ggplot2::scale_fill_manual().

multiplexed
Whether the FASTQ files are multiplexed. This is detected automatically for bcbioSingleCell objects.

Value

ggplot.

Details

A violin plot is a comact display of a continuous distribution. It is a blend of geom_boxplot() and geom_density(): a violin plot is a mirrored density plot displayed in the same way as a boxplot.

See also

Other Quality Control Metrics: metrics, plotCellCounts, plotGenesPerCell, plotMitoRatio, plotMitoVsCoding, plotNovelty, plotReadsPerCell, plotUMIsVsGenes, plotZerosVsDepth

Examples

# bcbioSingleCell

# NOT RUN {
plotUMIsPerCell(bcb)
# }

# seurat

# NOT RUN {
plotUMIsPerCell(seurat)
# }

# data.frame

# NOT RUN {
metrics <- metrics(bcb)
plotUMIsPerCell(metrics)
# }