Read 10X Genomics Cell Ranger Data

Read 10x Genomics Chromium cell counts from barcodes.tsv, genes.tsv, and matrix.mtx files.

readCellRanger(uploadDir, sampleMetadataFile = NULL, refdataDir = NULL,
  interestingGroups = "sampleName", transgeneNames = NULL,
  spikeNames = NULL, ...)

Arguments

uploadDir
Path to Cell Ranger output directory. This directory path must contain filtered_gene_bc_matrices* as a child directory.

sampleMetadataFile
Sample barcode metadata file. Optional for runs with demultiplixed index barcodes (e.g. SureCell), but otherwise required for runs with multipliexed FASTQs containing multiple index barcodes (e.g. inDrop).

refdataDir
Directory path to Cell Ranger reference annotation data.

interestingGroups
Character vector of interesting groups. Must be formatted in camel case and intersect with sampleData() colnames.

transgeneNames
character vector indicating which assay() rows denote transgenes (e.g. EGFP, TDTOMATO).

spikeNames
character vector indicating which assay() rows denote spike-in sequences (e.g. ERCCs).

...
Additional arguments, to be stashed in the metadata() slot.

Value

SingleCellExperiment.

Directory structure

Cell Ranger can vary in its output directory structure, but we're requiring a single, consistent data structure for datasets containing multiple samples. Note that Cell Ranger data may not always contain per sample subdirectories, or the "outs" subdirectory. We may make this more flexible in the future, but for now we're making this strict to ensure reproducibility.

file.path(
    "<uploadDir>",
    "<sampleName>",
    "outs",
    "filtered_gene_bc_matrices*",
    "outs",
    "<genomeBuild>",
    "matrix.mtx"
)

Sample metadata

A user-supplied sample metadata file defined by sampleMetadataFile is required for multiplexed datasets. Otherwise this can be left NULL, and minimal sample data will be used, based on the directory names.

Reference data

We strongly recommend supplying the corresponding reference data required for Cell Ranger with the refdataDir argument. When set, the function will detect the organism, ensemblRelease, and genomeBuild automatically, based on the 10X refdataDir YAML metadata. Additionally, it will convert the gene annotations defined in the GTF file into a GRanges object, which get slotted in rowRanges(). Otherwise, the function will attempt to use the most current annotations available from Ensembl, and some gene IDs may not match, due to deprecation in the current Ensembl release.

See also

Examples

uploadDir <- system.file("extdata/cellranger", package = "bcbioSingleCell")
x <- readCellRanger(uploadDir)

#> Making GRanges from Ensembl with AnnotationHub 2.12.0 (2018-04-30)

#> id: "AH60977"
#> organism: "Homo sapiens"
#> build: "GRCh38"
#> release: "92"
#> format: "genes"

#> Standardizing the metadata columns

#> Using geneName instead of symbol

#> Setting geneBiotype as factor

#> Setting seqCoordSystem as factor

#> Defining broadClass using: geneName, geneBiotype, seqnames

#> Reading counts at gene level

#> Calculating cellular barcode metrics

#> 500 cells detected

#> 312 coding genes detected

#> 0 mitochondrial genes detected

#> 500 / 500 cellular barcodes passed pre-filtering (100%)

#> Warning: Unannotated rows detected (9):
#> ENSG00000236743, ENSG00000278566, ENSG00000273547, ENSG00000279244, ENSG00000...

show(x)

#> class: SingleCellExperiment 
#> dim: 500 500 
#> metadata(20): version pipeline ... utilsSessionInfo devtoolsSessionInfo
#> assays(1): counts
#> rownames(500): ENSG00000243485 ENSG00000237613 ... ENSG00000230703
#>   ENSG00000142677
#> rowData names(7): geneID geneName ... entrezID broadClass
#> colnames(500): cellranger_1_AAACCTGAGAAGGCCT
#>   cellranger_1_AAACCTGAGACAGACC ... cellranger_1_ACTTACTCACTGCCAG
#>   cellranger_1_ACTTACTCAGTACACT
#> colData names(10): sampleID nUMI ... description index
#> reducedDimNames(0):
#> spikeNames(0):