Sample Data

Metadata in columns describing the samples, which are defined in the rownames. Similar to colData(), which for bcbioSingleCell and SingleCellExperiment objects describes cells in the columns, rather than the samples.

# S4 method for SingleCellExperiment
sampleData(object, interestingGroups,
  return = c("DataFrame", "data.frame", "kable"))

# S4 method for seurat
sampleData(object, interestingGroups,
  return = c("DataFrame", "data.frame"))

# S4 method for SingleCellExperiment,DataFrame
sampleData(object) <- value

# S4 method for seurat,DataFrame
sampleData(object) <- value

Arguments

object
Object.

interestingGroups
Character vector of interesting groups. Must be formatted in camel case and intersect with sampleData() colnames.

return
Return type. Uses base::match.arg() internally and defaults to the first argument in the vector.

value
Object to assign.

Value

data.frame.

Examples

# bcbioSingleCell ====
x <- bcb_small
sampleData(x) %>% glimpse()

#> Formal class 'DataFrame' [package "S4Vectors"] with 6 slots
#>   ..@ rownames       : chr "multiplexed_AAAAAAAA"
#>   ..@ nrows          : int 1
#>   ..@ listData       :List of 8
#>   .. ..$ sampleName       : Factor w/ 1 level "rep_1": 1
#>   .. ..$ description      : Factor w/ 1 level "multiplexed": 1
#>   .. ..$ fileName         : Factor w/ 1 level "multiplexed.fastq.gz": 1
#>   .. ..$ index            : Factor w/ 1 level "1": 1
#>   .. ..$ sequence         : Factor w/ 1 level "TTTTTTTT": 1
#>   .. ..$ aggregate        : Factor w/ 1 level "sample": 1
#>   .. ..$ revcomp          : Factor w/ 1 level "AAAAAAAA": 1
#>   .. ..$ interestingGroups: Factor w/ 1 level "rep_1": 1
#>   ..@ elementType    : chr "ANY"
#>   ..@ elementMetadata: NULL
#>   ..@ metadata       : list()

sampleData(x)[["batch"]] <- 1L
sampleData(x) %>% glimpse()

#> Formal class 'DataFrame' [package "S4Vectors"] with 6 slots
#>   ..@ rownames       : chr "multiplexed_AAAAAAAA"
#>   ..@ nrows          : int 1
#>   ..@ listData       :List of 9
#>   .. ..$ sampleName       : Factor w/ 1 level "rep_1": 1
#>   .. ..$ description      : Factor w/ 1 level "multiplexed": 1
#>   .. ..$ fileName         : Factor w/ 1 level "multiplexed.fastq.gz": 1
#>   .. ..$ index            : Factor w/ 1 level "1": 1
#>   .. ..$ sequence         : Factor w/ 1 level "TTTTTTTT": 1
#>   .. ..$ aggregate        : Factor w/ 1 level "sample": 1
#>   .. ..$ revcomp          : Factor w/ 1 level "AAAAAAAA": 1
#>   .. ..$ interestingGroups: Factor w/ 1 level "rep_1": 1
#>   .. ..$ batch            : Factor w/ 1 level "1": 1
#>   ..@ elementType    : chr "ANY"
#>   ..@ elementMetadata: NULL
#>   ..@ metadata       : list()

# seurat ====
x <- seurat_small
sampleData(x) %>% glimpse()

#> Formal class 'DataFrame' [package "S4Vectors"] with 6 slots
#>   ..@ rownames       : chr "multiplexed_AAAAAAAA"
#>   ..@ nrows          : int 1
#>   ..@ listData       :List of 8
#>   .. ..$ sampleName       : Factor w/ 1 level "rep_1": 1
#>   .. ..$ description      : Factor w/ 1 level "multiplexed": 1
#>   .. ..$ fileName         : Factor w/ 1 level "multiplexed.fastq.gz": 1
#>   .. ..$ index            : Factor w/ 1 level "1": 1
#>   .. ..$ sequence         : Factor w/ 1 level "TTTTTTTT": 1
#>   .. ..$ aggregate        : Factor w/ 1 level "sample": 1
#>   .. ..$ revcomp          : Factor w/ 1 level "AAAAAAAA": 1
#>   .. ..$ interestingGroups: Factor w/ 1 level "rep_1": 1
#>   ..@ elementType    : chr "ANY"
#>   ..@ elementMetadata: NULL
#>   ..@ metadata       : list()

sampleData(x)[["batch"]] <- 1L
sampleData(x) %>% glimpse()

#> Formal class 'DataFrame' [package "S4Vectors"] with 6 slots
#>   ..@ rownames       : chr "multiplexed_AAAAAAAA"
#>   ..@ nrows          : int 1
#>   ..@ listData       :List of 9
#>   .. ..$ sampleName       : Factor w/ 1 level "rep_1": 1
#>   .. ..$ description      : Factor w/ 1 level "multiplexed": 1
#>   .. ..$ fileName         : Factor w/ 1 level "multiplexed.fastq.gz": 1
#>   .. ..$ index            : Factor w/ 1 level "1": 1
#>   .. ..$ sequence         : Factor w/ 1 level "TTTTTTTT": 1
#>   .. ..$ aggregate        : Factor w/ 1 level "sample": 1
#>   .. ..$ revcomp          : Factor w/ 1 level "AAAAAAAA": 1
#>   .. ..$ interestingGroups: Factor w/ 1 level "rep_1": 1
#>   .. ..$ batch            : Factor w/ 1 level "1": 1
#>   ..@ elementType    : chr "ANY"
#>   ..@ elementMetadata: NULL
#>   ..@ metadata       : list()