Sanitize Markers Output

Sanitize Markers Output

sanitizeMarkers(object, markers, ...)

# S4 method for seurat,data.frame
sanitizeMarkers(object, markers)

Arguments

object
Object.

markers
Original Seurat::FindAllMarkers() data.frame.

...
Additional arguments.

Value

Tibble arranged by adjusted P value.

Seurat::FindAllMarkers(): grouped_df, grouped by cluster ident.

Note

Seurat::FindAllMarkers() maps the counts matrix rownames correctly in the gene column, whereas Seurat::FindMarkers() maps them correctly in the rownames of the returned marker data.frame.

See also

Other Clustering Functions: cellTypesPerCluster, knownMarkersDetected, plotCellTypesPerCluster, plotFeatureTSNE, plotKnownMarkersDetected, plotMarkerTSNE, plotMarkers, plotPCA, plotPCElbow, plotTSNE, plotTopMarkers, topBarcodes, topMarkers

Examples

# seurat ====
all_markers <- FindAllMarkers(seurat_small)

#>    |                                                  | 0 % ~calculating     |+                                                 | 1 % ~15s             |++                                                | 2 % ~11s             |++                                                | 3 % ~10s             |+++                                               | 4 % ~09s             |+++                                               | 6 % ~09s             |++++                                              | 7 % ~08s             |++++                                              | 8 % ~08s             |+++++                                             | 9 % ~08s             |+++++                                             | 10% ~08s             |++++++                                            | 11% ~08s             |+++++++                                           | 12% ~08s             |+++++++                                           | 13% ~08s             |++++++++                                          | 14% ~07s             |++++++++                                          | 16% ~07s             |+++++++++                                         | 17% ~07s             |+++++++++                                         | 18% ~07s             |++++++++++                                        | 19% ~07s             |++++++++++                                        | 20% ~07s             |+++++++++++                                       | 21% ~07s             |++++++++++++                                      | 22% ~06s             |++++++++++++                                      | 23% ~06s             |+++++++++++++                                     | 24% ~06s             |+++++++++++++                                     | 26% ~06s             |++++++++++++++                                    | 27% ~06s             |++++++++++++++                                    | 28% ~06s             |+++++++++++++++                                   | 29% ~06s             |+++++++++++++++                                   | 30% ~06s             |++++++++++++++++                                  | 31% ~06s             |+++++++++++++++++                                 | 32% ~05s             |+++++++++++++++++                                 | 33% ~05s             |++++++++++++++++++                                | 34% ~05s             |++++++++++++++++++                                | 36% ~05s             |+++++++++++++++++++                               | 37% ~05s             |+++++++++++++++++++                               | 38% ~05s             |++++++++++++++++++++                              | 39% ~05s             |++++++++++++++++++++                              | 40% ~05s             |+++++++++++++++++++++                             | 41% ~05s             |++++++++++++++++++++++                            | 42% ~05s             |++++++++++++++++++++++                            | 43% ~04s             |+++++++++++++++++++++++                           | 44% ~04s             |+++++++++++++++++++++++                           | 46% ~04s             |++++++++++++++++++++++++                          | 47% ~04s             |++++++++++++++++++++++++                          | 48% ~04s             |+++++++++++++++++++++++++                         | 49% ~04s             |+++++++++++++++++++++++++                         | 50% ~04s             |++++++++++++++++++++++++++                        | 51% ~04s             |+++++++++++++++++++++++++++                       | 52% ~04s             |+++++++++++++++++++++++++++                       | 53% ~04s             |++++++++++++++++++++++++++++                      | 54% ~04s             |++++++++++++++++++++++++++++                      | 56% ~03s             |+++++++++++++++++++++++++++++                     | 57% ~03s             |+++++++++++++++++++++++++++++                     | 58% ~03s             |++++++++++++++++++++++++++++++                    | 59% ~03s             |++++++++++++++++++++++++++++++                    | 60% ~03s             |+++++++++++++++++++++++++++++++                   | 61% ~03s             |++++++++++++++++++++++++++++++++                  | 62% ~03s             |++++++++++++++++++++++++++++++++                  | 63% ~03s             |+++++++++++++++++++++++++++++++++                 | 64% ~03s             |+++++++++++++++++++++++++++++++++                 | 66% ~03s             |++++++++++++++++++++++++++++++++++                | 67% ~03s             |++++++++++++++++++++++++++++++++++                | 68% ~02s             |+++++++++++++++++++++++++++++++++++               | 69% ~02s             |+++++++++++++++++++++++++++++++++++               | 70% ~02s             |++++++++++++++++++++++++++++++++++++              | 71% ~02s             |+++++++++++++++++++++++++++++++++++++             | 72% ~02s             |+++++++++++++++++++++++++++++++++++++             | 73% ~02s             |++++++++++++++++++++++++++++++++++++++            | 74% ~02s             |++++++++++++++++++++++++++++++++++++++            | 76% ~02s             |+++++++++++++++++++++++++++++++++++++++           | 77% ~02s             |+++++++++++++++++++++++++++++++++++++++           | 78% ~02s             |++++++++++++++++++++++++++++++++++++++++          | 79% ~02s             |++++++++++++++++++++++++++++++++++++++++         | 80% ~02s             |+++++++++++++++++++++++++++++++++++++++++         | 81% ~01s             |++++++++++++++++++++++++++++++++++++++++++        | 82% ~01s             |++++++++++++++++++++++++++++++++++++++++++        | 83% ~01s             |+++++++++++++++++++++++++++++++++++++++++++       | 84% ~01s             |+++++++++++++++++++++++++++++++++++++++++++       | 86% ~01s             |++++++++++++++++++++++++++++++++++++++++++++      | 87% ~01s             |++++++++++++++++++++++++++++++++++++++++++++      | 88% ~01s             |+++++++++++++++++++++++++++++++++++++++++++++     | 89% ~01s             |+++++++++++++++++++++++++++++++++++++++++++++    | 90% ~01s             |++++++++++++++++++++++++++++++++++++++++++++++    | 91% ~01s             |+++++++++++++++++++++++++++++++++++++++++++++++   | 92% ~01s             |+++++++++++++++++++++++++++++++++++++++++++++++   | 93% ~01s             |++++++++++++++++++++++++++++++++++++++++++++++++  | 94% ~00s             |++++++++++++++++++++++++++++++++++++++++++++++++  | 96% ~00s             |+++++++++++++++++++++++++++++++++++++++++++++++++ | 97% ~00s             |+++++++++++++++++++++++++++++++++++++++++++++++++ | 98% ~00s             |++++++++++++++++++++++++++++++++++++++++++++++++++| 99% ~00s             |++++++++++++++++++++++++++++++++++++++++++++++++++| 100% elapsed = 08s
#>    |                                                  | 0 % ~calculating     |+                                                 | 1 % ~04s             |++                                                | 3 % ~04s             |++                                                | 4 % ~04s             |+++                                               | 5 % ~04s             |++++                                              | 7 % ~04s             |++++                                              | 8 % ~04s             |+++++                                             | 9 % ~04s             |++++++                                            | 11% ~04s             |++++++                                            | 12% ~04s             |+++++++                                           | 13% ~04s             |++++++++                                          | 14% ~04s             |++++++++                                          | 16% ~04s             |+++++++++                                         | 17% ~04s             |++++++++++                                        | 18% ~04s             |++++++++++                                        | 20% ~04s             |+++++++++++                                       | 21% ~04s             |++++++++++++                                      | 22% ~03s             |++++++++++++                                      | 24% ~03s             |+++++++++++++                                     | 25% ~03s             |++++++++++++++                                    | 26% ~03s             |++++++++++++++                                    | 28% ~03s             |+++++++++++++++                                   | 29% ~03s             |++++++++++++++++                                  | 30% ~03s             |++++++++++++++++                                  | 32% ~03s             |+++++++++++++++++                                 | 33% ~03s             |++++++++++++++++++                                | 34% ~03s             |++++++++++++++++++                                | 36% ~03s             |+++++++++++++++++++                               | 37% ~03s             |++++++++++++++++++++                              | 38% ~03s             |++++++++++++++++++++                              | 39% ~03s             |+++++++++++++++++++++                             | 41% ~03s             |++++++++++++++++++++++                            | 42% ~03s             |++++++++++++++++++++++                            | 43% ~03s             |+++++++++++++++++++++++                           | 45% ~02s             |++++++++++++++++++++++++                          | 46% ~02s             |++++++++++++++++++++++++                          | 47% ~02s             |+++++++++++++++++++++++++                         | 49% ~02s             |+++++++++++++++++++++++++                         | 50% ~02s             |++++++++++++++++++++++++++                        | 51% ~02s             |+++++++++++++++++++++++++++                       | 53% ~02s             |+++++++++++++++++++++++++++                       | 54% ~02s             |++++++++++++++++++++++++++++                      | 55% ~02s             |+++++++++++++++++++++++++++++                     | 57% ~02s             |+++++++++++++++++++++++++++++                     | 58% ~02s             |++++++++++++++++++++++++++++++                    | 59% ~02s             |+++++++++++++++++++++++++++++++                   | 61% ~02s             |+++++++++++++++++++++++++++++++                   | 62% ~02s             |++++++++++++++++++++++++++++++++                  | 63% ~02s             |+++++++++++++++++++++++++++++++++                 | 64% ~02s             |+++++++++++++++++++++++++++++++++                 | 66% ~02s             |++++++++++++++++++++++++++++++++++                | 67% ~01s             |+++++++++++++++++++++++++++++++++++               | 68% ~01s             |+++++++++++++++++++++++++++++++++++               | 70% ~01s             |++++++++++++++++++++++++++++++++++++              | 71% ~01s             |+++++++++++++++++++++++++++++++++++++             | 72% ~01s             |+++++++++++++++++++++++++++++++++++++             | 74% ~01s             |++++++++++++++++++++++++++++++++++++++            | 75% ~01s             |+++++++++++++++++++++++++++++++++++++++           | 76% ~01s             |+++++++++++++++++++++++++++++++++++++++           | 78% ~01s             |++++++++++++++++++++++++++++++++++++++++          | 79% ~01s             |+++++++++++++++++++++++++++++++++++++++++         | 80% ~01s             |+++++++++++++++++++++++++++++++++++++++++         | 82% ~01s             |++++++++++++++++++++++++++++++++++++++++++        | 83% ~01s             |+++++++++++++++++++++++++++++++++++++++++++       | 84% ~01s             |+++++++++++++++++++++++++++++++++++++++++++       | 86% ~01s             |++++++++++++++++++++++++++++++++++++++++++++      | 87% ~01s             |+++++++++++++++++++++++++++++++++++++++++++++     | 88% ~01s             |+++++++++++++++++++++++++++++++++++++++++++++     | 89% ~00s             |++++++++++++++++++++++++++++++++++++++++++++++    | 91% ~00s             |+++++++++++++++++++++++++++++++++++++++++++++++   | 92% ~00s             |+++++++++++++++++++++++++++++++++++++++++++++++   | 93% ~00s             |++++++++++++++++++++++++++++++++++++++++++++++++  | 95% ~00s             |+++++++++++++++++++++++++++++++++++++++++++++++++ | 96% ~00s             |+++++++++++++++++++++++++++++++++++++++++++++++++ | 97% ~00s             |++++++++++++++++++++++++++++++++++++++++++++++++++| 99% ~00s             |++++++++++++++++++++++++++++++++++++++++++++++++++| 100% elapsed = 04s
#>    |                                                  | 0 % ~calculating     |+                                                 | 1 % ~05s             |++                                                | 2 % ~05s             |++                                                | 3 % ~05s             |+++                                               | 4 % ~05s             |+++                                               | 5 % ~05s             |++++                                              | 6 % ~05s             |++++                                              | 7 % ~05s             |+++++                                             | 8 % ~05s             |+++++                                             | 9 % ~05s             |++++++                                            | 11% ~05s             |++++++                                            | 12% ~05s             |+++++++                                           | 13% ~05s             |+++++++                                           | 14% ~05s             |++++++++                                          | 15% ~05s             |++++++++                                          | 16% ~05s             |+++++++++                                         | 17% ~05s             |+++++++++                                         | 18% ~05s             |++++++++++                                        | 19% ~04s             |++++++++++                                        | 20% ~04s             |+++++++++++                                       | 21% ~04s             |++++++++++++                                      | 22% ~04s             |++++++++++++                                      | 23% ~04s             |+++++++++++++                                     | 24% ~04s             |+++++++++++++                                     | 25% ~04s             |++++++++++++++                                    | 26% ~04s             |++++++++++++++                                    | 27% ~04s             |+++++++++++++++                                   | 28% ~04s             |+++++++++++++++                                   | 29% ~04s             |++++++++++++++++                                  | 31% ~04s             |++++++++++++++++                                  | 32% ~04s             |+++++++++++++++++                                 | 33% ~04s             |+++++++++++++++++                                 | 34% ~04s             |++++++++++++++++++                                | 35% ~04s             |++++++++++++++++++                                | 36% ~04s             |+++++++++++++++++++                               | 37% ~03s             |+++++++++++++++++++                               | 38% ~03s             |++++++++++++++++++++                              | 39% ~03s             |++++++++++++++++++++                              | 40% ~03s             |+++++++++++++++++++++                             | 41% ~03s             |++++++++++++++++++++++                            | 42% ~03s             |++++++++++++++++++++++                            | 43% ~03s             |+++++++++++++++++++++++                           | 44% ~03s             |+++++++++++++++++++++++                           | 45% ~03s             |++++++++++++++++++++++++                          | 46% ~03s             |++++++++++++++++++++++++                          | 47% ~03s             |+++++++++++++++++++++++++                         | 48% ~03s             |+++++++++++++++++++++++++                         | 49% ~03s             |++++++++++++++++++++++++++                        | 51% ~03s             |++++++++++++++++++++++++++                        | 52% ~03s             |+++++++++++++++++++++++++++                       | 53% ~03s             |+++++++++++++++++++++++++++                       | 54% ~03s             |++++++++++++++++++++++++++++                      | 55% ~03s             |++++++++++++++++++++++++++++                      | 56% ~02s             |+++++++++++++++++++++++++++++                     | 57% ~02s             |+++++++++++++++++++++++++++++                     | 58% ~02s             |++++++++++++++++++++++++++++++                    | 59% ~02s             |++++++++++++++++++++++++++++++                    | 60% ~02s             |+++++++++++++++++++++++++++++++                   | 61% ~02s             |++++++++++++++++++++++++++++++++                  | 62% ~02s             |++++++++++++++++++++++++++++++++                  | 63% ~02s             |+++++++++++++++++++++++++++++++++                 | 64% ~02s             |+++++++++++++++++++++++++++++++++                 | 65% ~02s             |++++++++++++++++++++++++++++++++++                | 66% ~02s             |++++++++++++++++++++++++++++++++++                | 67% ~02s             |+++++++++++++++++++++++++++++++++++               | 68% ~02s             |+++++++++++++++++++++++++++++++++++               | 69% ~02s             |++++++++++++++++++++++++++++++++++++              | 71% ~02s             |++++++++++++++++++++++++++++++++++++              | 72% ~02s             |+++++++++++++++++++++++++++++++++++++             | 73% ~02s             |+++++++++++++++++++++++++++++++++++++             | 74% ~01s             |++++++++++++++++++++++++++++++++++++++            | 75% ~01s             |++++++++++++++++++++++++++++++++++++++            | 76% ~01s             |+++++++++++++++++++++++++++++++++++++++           | 77% ~01s             |+++++++++++++++++++++++++++++++++++++++           | 78% ~01s             |++++++++++++++++++++++++++++++++++++++++          | 79% ~01s             |++++++++++++++++++++++++++++++++++++++++         | 80% ~01s             |+++++++++++++++++++++++++++++++++++++++++         | 81% ~01s             |++++++++++++++++++++++++++++++++++++++++++        | 82% ~01s             |++++++++++++++++++++++++++++++++++++++++++        | 83% ~01s             |+++++++++++++++++++++++++++++++++++++++++++       | 84% ~01s             |+++++++++++++++++++++++++++++++++++++++++++       | 85% ~01s             |++++++++++++++++++++++++++++++++++++++++++++      | 86% ~01s             |++++++++++++++++++++++++++++++++++++++++++++      | 87% ~01s             |+++++++++++++++++++++++++++++++++++++++++++++     | 88% ~01s             |+++++++++++++++++++++++++++++++++++++++++++++     | 89% ~01s             |++++++++++++++++++++++++++++++++++++++++++++++    | 91% ~01s             |++++++++++++++++++++++++++++++++++++++++++++++    | 92% ~00s             |+++++++++++++++++++++++++++++++++++++++++++++++   | 93% ~00s             |+++++++++++++++++++++++++++++++++++++++++++++++   | 94% ~00s             |++++++++++++++++++++++++++++++++++++++++++++++++  | 95% ~00s             |++++++++++++++++++++++++++++++++++++++++++++++++  | 96% ~00s             |+++++++++++++++++++++++++++++++++++++++++++++++++ | 97% ~00s             |+++++++++++++++++++++++++++++++++++++++++++++++++ | 98% ~00s             |++++++++++++++++++++++++++++++++++++++++++++++++++| 99% ~00s             |++++++++++++++++++++++++++++++++++++++++++++++++++| 100% elapsed = 06s
#>    |                                                  | 0 % ~calculating     |+                                                 | 1 % ~09s             |++                                                | 2 % ~09s             |++                                                | 3 % ~08s             |+++                                               | 5 % ~08s             |+++                                               | 6 % ~08s             |++++                                              | 7 % ~08s             |+++++                                             | 8 % ~08s             |+++++                                             | 9 % ~08s             |++++++                                            | 10% ~08s             |++++++                                            | 12% ~08s             |+++++++                                           | 13% ~08s             |+++++++                                           | 14% ~07s             |++++++++                                          | 15% ~07s             |+++++++++                                         | 16% ~07s             |+++++++++                                         | 17% ~07s             |++++++++++                                        | 19% ~07s             |++++++++++                                        | 20% ~07s             |+++++++++++                                       | 21% ~07s             |++++++++++++                                      | 22% ~07s             |++++++++++++                                      | 23% ~07s             |+++++++++++++                                     | 24% ~07s             |+++++++++++++                                     | 26% ~06s             |++++++++++++++                                    | 27% ~06s             |++++++++++++++                                    | 28% ~06s             |+++++++++++++++                                   | 29% ~06s             |++++++++++++++++                                  | 30% ~06s             |++++++++++++++++                                  | 31% ~06s             |+++++++++++++++++                                 | 33% ~06s             |+++++++++++++++++                                 | 34% ~06s             |++++++++++++++++++                                | 35% ~06s             |+++++++++++++++++++                               | 36% ~05s             |+++++++++++++++++++                               | 37% ~05s             |++++++++++++++++++++                              | 38% ~05s             |++++++++++++++++++++                              | 40% ~05s             |+++++++++++++++++++++                             | 41% ~05s             |+++++++++++++++++++++                             | 42% ~05s             |++++++++++++++++++++++                            | 43% ~05s             |+++++++++++++++++++++++                           | 44% ~05s             |+++++++++++++++++++++++                           | 45% ~05s             |++++++++++++++++++++++++                          | 47% ~05s             |++++++++++++++++++++++++                          | 48% ~05s             |+++++++++++++++++++++++++                         | 49% ~04s             |+++++++++++++++++++++++++                         | 50% ~04s             |++++++++++++++++++++++++++                        | 51% ~04s             |+++++++++++++++++++++++++++                       | 52% ~04s             |+++++++++++++++++++++++++++                       | 53% ~04s             |++++++++++++++++++++++++++++                      | 55% ~04s             |++++++++++++++++++++++++++++                      | 56% ~04s             |+++++++++++++++++++++++++++++                     | 57% ~04s             |++++++++++++++++++++++++++++++                    | 58% ~04s             |++++++++++++++++++++++++++++++                    | 59% ~04s             |+++++++++++++++++++++++++++++++                   | 60% ~03s             |+++++++++++++++++++++++++++++++                   | 62% ~03s             |++++++++++++++++++++++++++++++++                  | 63% ~03s             |++++++++++++++++++++++++++++++++                  | 64% ~03s             |+++++++++++++++++++++++++++++++++                 | 65% ~03s             |++++++++++++++++++++++++++++++++++                | 66% ~03s             |++++++++++++++++++++++++++++++++++                | 67% ~03s             |+++++++++++++++++++++++++++++++++++               | 69% ~03s             |+++++++++++++++++++++++++++++++++++               | 70% ~03s             |++++++++++++++++++++++++++++++++++++              | 71% ~02s             |+++++++++++++++++++++++++++++++++++++             | 72% ~02s             |+++++++++++++++++++++++++++++++++++++             | 73% ~02s             |++++++++++++++++++++++++++++++++++++++            | 74% ~02s             |++++++++++++++++++++++++++++++++++++++            | 76% ~02s             |+++++++++++++++++++++++++++++++++++++++           | 77% ~02s             |+++++++++++++++++++++++++++++++++++++++           | 78% ~02s             |++++++++++++++++++++++++++++++++++++++++          | 79% ~02s             |+++++++++++++++++++++++++++++++++++++++++         | 80% ~02s             |+++++++++++++++++++++++++++++++++++++++++         | 81% ~02s             |++++++++++++++++++++++++++++++++++++++++++        | 83% ~02s             |++++++++++++++++++++++++++++++++++++++++++        | 84% ~01s             |+++++++++++++++++++++++++++++++++++++++++++       | 85% ~01s             |++++++++++++++++++++++++++++++++++++++++++++      | 86% ~01s             |++++++++++++++++++++++++++++++++++++++++++++      | 87% ~01s             |+++++++++++++++++++++++++++++++++++++++++++++     | 88% ~01s             |+++++++++++++++++++++++++++++++++++++++++++++     | 90% ~01s             |++++++++++++++++++++++++++++++++++++++++++++++    | 91% ~01s             |++++++++++++++++++++++++++++++++++++++++++++++    | 92% ~01s             |+++++++++++++++++++++++++++++++++++++++++++++++   | 93% ~01s             |++++++++++++++++++++++++++++++++++++++++++++++++  | 94% ~01s             |++++++++++++++++++++++++++++++++++++++++++++++++  | 95% ~00s             |+++++++++++++++++++++++++++++++++++++++++++++++++ | 97% ~00s             |+++++++++++++++++++++++++++++++++++++++++++++++++ | 98% ~00s             |++++++++++++++++++++++++++++++++++++++++++++++++++| 99% ~00s             |++++++++++++++++++++++++++++++++++++++++++++++++++| 100% elapsed = 09s
#>    |                                                  | 0 % ~calculating     |+                                                 | 1 % ~07s             |++                                                | 2 % ~07s             |++                                                | 3 % ~07s             |+++                                               | 4 % ~07s             |+++                                               | 5 % ~07s             |++++                                              | 7 % ~07s             |++++                                              | 8 % ~06s             |+++++                                             | 9 % ~06s             |+++++                                             | 10% ~06s             |++++++                                            | 11% ~06s             |+++++++                                           | 12% ~06s             |+++++++                                           | 13% ~06s             |++++++++                                          | 14% ~06s             |++++++++                                          | 15% ~06s             |+++++++++                                         | 16% ~06s             |+++++++++                                         | 18% ~06s             |++++++++++                                        | 19% ~06s             |++++++++++                                        | 20% ~06s             |+++++++++++                                       | 21% ~06s             |+++++++++++                                       | 22% ~06s             |++++++++++++                                      | 23% ~05s             |+++++++++++++                                     | 24% ~05s             |+++++++++++++                                     | 25% ~05s             |++++++++++++++                                    | 26% ~05s             |++++++++++++++                                    | 27% ~05s             |+++++++++++++++                                   | 29% ~05s             |+++++++++++++++                                   | 30% ~05s             |++++++++++++++++                                  | 31% ~05s             |++++++++++++++++                                  | 32% ~05s             |+++++++++++++++++                                 | 33% ~05s             |++++++++++++++++++                                | 34% ~05s             |++++++++++++++++++                                | 35% ~05s             |+++++++++++++++++++                               | 36% ~05s             |+++++++++++++++++++                               | 37% ~04s             |++++++++++++++++++++                              | 38% ~04s             |++++++++++++++++++++                              | 40% ~04s             |+++++++++++++++++++++                             | 41% ~04s             |+++++++++++++++++++++                             | 42% ~04s             |++++++++++++++++++++++                            | 43% ~04s             |++++++++++++++++++++++                            | 44% ~04s             |+++++++++++++++++++++++                           | 45% ~04s             |++++++++++++++++++++++++                          | 46% ~04s             |++++++++++++++++++++++++                          | 47% ~04s             |+++++++++++++++++++++++++                         | 48% ~04s             |+++++++++++++++++++++++++                         | 49% ~04s             |++++++++++++++++++++++++++                        | 51% ~04s             |++++++++++++++++++++++++++                        | 52% ~03s             |+++++++++++++++++++++++++++                       | 53% ~03s             |+++++++++++++++++++++++++++                       | 54% ~03s             |++++++++++++++++++++++++++++                      | 55% ~03s             |+++++++++++++++++++++++++++++                     | 56% ~03s             |+++++++++++++++++++++++++++++                     | 57% ~03s             |++++++++++++++++++++++++++++++                    | 58% ~03s             |++++++++++++++++++++++++++++++                    | 59% ~03s             |+++++++++++++++++++++++++++++++                   | 60% ~03s             |+++++++++++++++++++++++++++++++                   | 62% ~03s             |++++++++++++++++++++++++++++++++                  | 63% ~03s             |++++++++++++++++++++++++++++++++                  | 64% ~03s             |+++++++++++++++++++++++++++++++++                 | 65% ~03s             |+++++++++++++++++++++++++++++++++                 | 66% ~02s             |++++++++++++++++++++++++++++++++++                | 67% ~02s             |+++++++++++++++++++++++++++++++++++               | 68% ~02s             |+++++++++++++++++++++++++++++++++++               | 69% ~02s             |++++++++++++++++++++++++++++++++++++              | 70% ~02s             |++++++++++++++++++++++++++++++++++++              | 71% ~02s             |+++++++++++++++++++++++++++++++++++++             | 73% ~02s             |+++++++++++++++++++++++++++++++++++++             | 74% ~02s             |++++++++++++++++++++++++++++++++++++++            | 75% ~02s             |++++++++++++++++++++++++++++++++++++++            | 76% ~02s             |+++++++++++++++++++++++++++++++++++++++           | 77% ~02s             |++++++++++++++++++++++++++++++++++++++++          | 78% ~02s             |++++++++++++++++++++++++++++++++++++++++          | 79% ~01s             |+++++++++++++++++++++++++++++++++++++++++         | 80% ~01s             |+++++++++++++++++++++++++++++++++++++++++         | 81% ~01s             |++++++++++++++++++++++++++++++++++++++++++        | 82% ~01s             |++++++++++++++++++++++++++++++++++++++++++        | 84% ~01s             |+++++++++++++++++++++++++++++++++++++++++++       | 85% ~01s             |+++++++++++++++++++++++++++++++++++++++++++       | 86% ~01s             |++++++++++++++++++++++++++++++++++++++++++++      | 87% ~01s             |++++++++++++++++++++++++++++++++++++++++++++      | 88% ~01s             |+++++++++++++++++++++++++++++++++++++++++++++     | 89% ~01s             |++++++++++++++++++++++++++++++++++++++++++++++    | 90% ~01s             |++++++++++++++++++++++++++++++++++++++++++++++    | 91% ~01s             |+++++++++++++++++++++++++++++++++++++++++++++++   | 92% ~01s             |+++++++++++++++++++++++++++++++++++++++++++++++   | 93% ~00s             |++++++++++++++++++++++++++++++++++++++++++++++++  | 95% ~00s             |++++++++++++++++++++++++++++++++++++++++++++++++  | 96% ~00s             |+++++++++++++++++++++++++++++++++++++++++++++++++ | 97% ~00s             |+++++++++++++++++++++++++++++++++++++++++++++++++ | 98% ~00s             |++++++++++++++++++++++++++++++++++++++++++++++++++| 99% ~00s             |++++++++++++++++++++++++++++++++++++++++++++++++++| 100% elapsed = 07s

all_sanitized <- sanitizeMarkers(
    object = seurat_small,
    markers = all_markers
)

#> Renaming `pVal` column to `pvalue` (matching DESeq2)

#> Renaming `pValAdj` column to `padj` (matching DESeq2)

glimpse(all_sanitized)

#> Observations: 1,380
#> Variables: 18
#> $ cluster        <fct> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0...
#> $ geneID         <chr> "ENSG00000164692", "ENSG00000108821", "ENSG000001000...
#> $ geneName       <chr> "COL1A2", "COL1A1", "LGALS1", "SERPINE2", "COL6A3", ...
#> $ pct1           <dbl> 1.000, 1.000, 0.806, 0.831, 0.927, 0.976, 0.766, 1.0...
#> $ pct2           <dbl> 0.465, 0.596, 0.098, 0.125, 0.229, 0.306, 0.096, 0.5...
#> $ avgLogFC       <dbl> 2.5553118, 2.7949173, 2.2995216, 2.4224445, 2.284212...
#> $ padj           <dbl> 5.313721e-58, 1.979923e-57, 1.215935e-54, 1.019308e-...
#> $ pvalue         <dbl> 1.062744e-60, 3.959847e-60, 2.431870e-57, 2.038616e-...
#> $ rowname        <chr> "COL1A2", "COL1A1", "LGALS1", "SERPINE2", "COL6A3", ...
#> $ seqnames       <fct> 7, 17, 22, 2, 2, 2, 14, 2, 22, 17, 9, 21, 5, 7, 8, 1...
#> $ start          <int> 94394561, 50183289, 37675608, 223975112, 237324003, ...
#> $ end            <int> 94431232, 50201632, 37679806, 224039319, 237414375, ...
#> $ width          <int> 36672, 18344, 4199, 64208, 90373, 75733, 12520, 3842...
#> $ strand         <fct> +, -, +, -, -, -, +, +, +, -, +, +, -, -, -, +, +, +...
#> $ geneBiotype    <fct> protein_coding, protein_coding, protein_coding, prot...
#> $ description    <chr> "collagen type I alpha 2 chain [Source:HGNC Symbol;A...
#> $ seqCoordSystem <fct> chromosome, chromosome, chromosome, chromosome, chro...
#> $ broadClass     <fct> coding, coding, coding, coding, coding, coding, codi...

ident_3_markers <- FindMarkers(seurat_small, ident.1 = "3", ident.2 = NULL)

#>    |                                                  | 0 % ~calculating     |+                                                 | 1 % ~09s             |++                                                | 2 % ~08s             |++                                                | 3 % ~08s             |+++                                               | 5 % ~08s             |+++                                               | 6 % ~08s             |++++                                              | 7 % ~08s             |+++++                                             | 8 % ~08s             |+++++                                             | 9 % ~08s             |++++++                                            | 10% ~08s             |++++++                                            | 12% ~07s             |+++++++                                           | 13% ~07s             |+++++++                                           | 14% ~07s             |++++++++                                          | 15% ~07s             |+++++++++                                         | 16% ~07s             |+++++++++                                         | 17% ~07s             |++++++++++                                        | 19% ~07s             |++++++++++                                        | 20% ~07s             |+++++++++++                                       | 21% ~07s             |++++++++++++                                      | 22% ~07s             |++++++++++++                                      | 23% ~06s             |+++++++++++++                                     | 24% ~06s             |+++++++++++++                                     | 26% ~06s             |++++++++++++++                                    | 27% ~06s             |++++++++++++++                                    | 28% ~06s             |+++++++++++++++                                   | 29% ~06s             |++++++++++++++++                                  | 30% ~06s             |++++++++++++++++                                  | 31% ~06s             |+++++++++++++++++                                 | 33% ~06s             |+++++++++++++++++                                 | 34% ~06s             |++++++++++++++++++                                | 35% ~05s             |+++++++++++++++++++                               | 36% ~06s             |+++++++++++++++++++                               | 37% ~05s             |++++++++++++++++++++                              | 38% ~05s             |++++++++++++++++++++                              | 40% ~05s             |+++++++++++++++++++++                             | 41% ~05s             |+++++++++++++++++++++                             | 42% ~05s             |++++++++++++++++++++++                            | 43% ~05s             |+++++++++++++++++++++++                           | 44% ~05s             |+++++++++++++++++++++++                           | 45% ~05s             |++++++++++++++++++++++++                          | 47% ~05s             |++++++++++++++++++++++++                          | 48% ~05s             |+++++++++++++++++++++++++                         | 49% ~04s             |+++++++++++++++++++++++++                         | 50% ~04s             |++++++++++++++++++++++++++                        | 51% ~04s             |+++++++++++++++++++++++++++                       | 52% ~04s             |+++++++++++++++++++++++++++                       | 53% ~04s             |++++++++++++++++++++++++++++                      | 55% ~04s             |++++++++++++++++++++++++++++                      | 56% ~04s             |+++++++++++++++++++++++++++++                     | 57% ~04s             |++++++++++++++++++++++++++++++                    | 58% ~04s             |++++++++++++++++++++++++++++++                    | 59% ~04s             |+++++++++++++++++++++++++++++++                   | 60% ~03s             |+++++++++++++++++++++++++++++++                   | 62% ~03s             |++++++++++++++++++++++++++++++++                  | 63% ~03s             |++++++++++++++++++++++++++++++++                  | 64% ~03s             |+++++++++++++++++++++++++++++++++                 | 65% ~03s             |++++++++++++++++++++++++++++++++++                | 66% ~03s             |++++++++++++++++++++++++++++++++++                | 67% ~03s             |+++++++++++++++++++++++++++++++++++               | 69% ~03s             |+++++++++++++++++++++++++++++++++++               | 70% ~03s             |++++++++++++++++++++++++++++++++++++              | 71% ~02s             |+++++++++++++++++++++++++++++++++++++             | 72% ~02s             |+++++++++++++++++++++++++++++++++++++             | 73% ~02s             |++++++++++++++++++++++++++++++++++++++            | 74% ~02s             |++++++++++++++++++++++++++++++++++++++            | 76% ~02s             |+++++++++++++++++++++++++++++++++++++++           | 77% ~02s             |+++++++++++++++++++++++++++++++++++++++           | 78% ~02s             |++++++++++++++++++++++++++++++++++++++++          | 79% ~02s             |+++++++++++++++++++++++++++++++++++++++++         | 80% ~02s             |+++++++++++++++++++++++++++++++++++++++++         | 81% ~02s             |++++++++++++++++++++++++++++++++++++++++++        | 83% ~01s             |++++++++++++++++++++++++++++++++++++++++++        | 84% ~01s             |+++++++++++++++++++++++++++++++++++++++++++       | 85% ~01s             |++++++++++++++++++++++++++++++++++++++++++++      | 86% ~01s             |++++++++++++++++++++++++++++++++++++++++++++      | 87% ~01s             |+++++++++++++++++++++++++++++++++++++++++++++     | 88% ~01s             |+++++++++++++++++++++++++++++++++++++++++++++     | 90% ~01s             |++++++++++++++++++++++++++++++++++++++++++++++    | 91% ~01s             |++++++++++++++++++++++++++++++++++++++++++++++    | 92% ~01s             |+++++++++++++++++++++++++++++++++++++++++++++++   | 93% ~01s             |++++++++++++++++++++++++++++++++++++++++++++++++  | 94% ~00s             |++++++++++++++++++++++++++++++++++++++++++++++++  | 95% ~00s             |+++++++++++++++++++++++++++++++++++++++++++++++++ | 97% ~00s             |+++++++++++++++++++++++++++++++++++++++++++++++++ | 98% ~00s             |++++++++++++++++++++++++++++++++++++++++++++++++++| 99% ~00s             |++++++++++++++++++++++++++++++++++++++++++++++++++| 100% elapsed = 09s

ident_3_sanitized <- sanitizeMarkers(
    object = seurat_small,
    markers = ident_3_markers
)

#> Renaming `pVal` column to `pvalue` (matching DESeq2)

#> Renaming `pValAdj` column to `padj` (matching DESeq2)

glimpse(ident_3_sanitized)

#> Observations: 427
#> Variables: 17
#> $ geneID         <chr> "ENSG00000156508", "ENSG00000198899", "ENSG000002485...
#> $ geneName       <chr> "EEF1A1", "MT-ATP6", "MTATP6P1", "MT-ND5", "MTND1P23...
#> $ pct1           <dbl> 0.865, 1.000, 1.000, 0.966, 0.989, 1.000, 0.989, 0.9...
#> $ pct2           <dbl> 0.985, 0.993, 0.988, 0.908, 0.959, 0.976, 0.988, 0.9...
#> $ avgLogFC       <dbl> -1.709205, 1.159734, 1.173533, 1.326033, 1.269361, 1...
#> $ padj           <dbl> 2.710328e-32, 1.796991e-31, 1.869653e-31, 2.357196e-...
#> $ pvalue         <dbl> 5.420656e-35, 3.593981e-34, 3.739306e-34, 4.714391e-...
#> $ rowname        <chr> "EEF1A1", "MT.ATP6", "MTATP6P1", "MT.ND5", "MTND1P23...
#> $ seqnames       <fct> 6, MT, 1, MT, 1, MT, MT, MT, MT, MT, X, MT, 16, MT, ...
#> $ start          <int> 73515750, 8527, 633696, 12337, 629062, 14747, 9207, ...
#> $ end            <int> 73523797, 9207, 634376, 14148, 629433, 15887, 9990, ...
#> $ width          <int> 8048, 681, 681, 1812, 372, 1141, 784, 1042, 1378, 95...
#> $ strand         <fct> -, +, +, +, +, +, +, +, +, +, +, +, -, -, +, +, +, -...
#> $ geneBiotype    <fct> protein_coding, protein_coding, unprocessed_pseudoge...
#> $ description    <chr> "eukaryotic translation elongation factor 1 alpha 1 ...
#> $ seqCoordSystem <fct> chromosome, chromosome, chromosome, chromosome, chro...
#> $ broadClass     <fct> coding, mito, pseudo, mito, pseudo, mito, mito, mito...