Sanitize Seurat Markers

Currently only Seurat markers are supported.

sanitizeSeuratMarkers(data, rowRanges)

Arguments

data
data.frame. Seurat::FindAllMarkers() or Seurat::FindMarkers() return.

rowRanges
GRanges. Gene annotations. Must contain geneName column that corresponds to the gene names in the seurat object used to generate the markers data.frame.

Value

data.frame, arranged by adjusted P value.

Note

Seurat::FindAllMarkers() maps the counts matrix rownames correctly in the gene column, whereas Seurat::FindMarkers() maps them correctly in the rownames of the returned marker data.frame.

See also

Examples

# seurat ====
object <- seurat_small

# `FindAllMarkers()` return
invisible(capture.output(
    all_markers <- Seurat::FindAllMarkers(object)
))
all_sanitized <- sanitizeSeuratMarkers(
    data = all_markers,
    rowRanges = rowRanges(object)
)
glimpse(all_sanitized)

#> Observations: 952
#> Variables: 18
#> $ cluster        <fct> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0...
#> $ avgLogFC       <dbl> 1.9829602, 3.0536375, 2.8619805, 2.2483912, 1.804789...
#> $ pct1           <dbl> 0.921, 1.000, 1.000, 0.872, 1.000, 0.988, 1.000, 1.0...
#> $ pct2           <dbl> 0.137, 0.807, 0.890, 0.173, 0.866, 0.384, 0.446, 0.4...
#> $ rowname        <chr> "VCAN", "S100A8", "S100A9", "S100A12", "LYZ", "AC020...
#> $ pvalue         <dbl> 1.369269e-74, 8.446076e-72, 1.100817e-71, 8.711241e-...
#> $ padj           <dbl> 6.846344e-72, 4.223038e-69, 5.504085e-69, 4.355621e-...
#> $ seqnames       <fct> 5, 1, 1, 1, 12, 12, 3, 9, 22, 3, 1, 19, 22, X, 1, 22...
#> $ start          <int> 83471465, 153390032, 153357854, 153373706, 69348341,...
#> $ end            <int> 83582303, 153391188, 153361027, 153375649, 69354234,...
#> $ width          <int> 110839, 1157, 3174, 1944, 5894, 733, 16729, 12074, 1...
#> $ strand         <fct> +, -, +, -, +, -, +, -, -, -, +, +, +, -, -, +, -, -...
#> $ broadClass     <fct> coding, coding, coding, coding, coding, other, codin...
#> $ description    <fct> versican [Source:HGNC Symbol;Acc:HGNC:2464], S100 ca...
#> $ geneBiotype    <fct> protein_coding, protein_coding, protein_coding, prot...
#> $ geneID         <chr> "ENSG00000038427", "ENSG00000143546", "ENSG000001632...
#> $ geneName       <fct> VCAN, S100A8, S100A9, S100A12, LYZ, AC020656.1, CSTA...
#> $ seqCoordSystem <fct> chromosome, chromosome, chromosome, chromosome, chro...

# `FindMarkers()` return
invisible(capture.output(
    ident_3_markers <- Seurat::FindMarkers(
        object = object,
        ident.1 = "3",
        ident.2 = NULL
    )
))
ident_3_sanitized <- sanitizeSeuratMarkers(
    data = ident_3_markers,
    rowRanges = rowRanges(object)
)
glimpse(ident_3_sanitized)

#> Observations: 181
#> Variables: 16
#> $ avgLogFC       <dbl> 2.0535014, 2.0801853, 3.0662657, -1.3856482, -0.6757...
#> $ pct1           <dbl> 0.900, 0.743, 0.843, 0.657, 1.000, 0.943, 0.357, 0.5...
#> $ pct2           <dbl> 0.153, 0.191, 0.351, 0.965, 1.000, 0.993, 0.881, 0.9...
#> $ pvalue         <dbl> 1.534176e-50, 4.789948e-31, 1.696343e-29, 4.771235e-...
#> $ padj           <dbl> 7.670879e-48, 2.394974e-28, 8.481717e-27, 2.385617e-...
#> $ seqnames       <fct> 14, 19, 2, 1, X, 1, 9, 22, 1, 21, 1, 3, 21, 12, 6, 1...
#> $ start          <int> 105851708, 41877120, 88857161, 151982915, 12975108, ...
#> $ end            <int> 105856218, 41881372, 88857683, 151994390, 12977227, ...
#> $ width          <int> 4511, 4253, 523, 11476, 2120, 6524, 18637, 11729, 15...
#> $ strand         <fct> -, +, -, -, +, -, +, +, -, -, -, +, -, +, +, +, +, -...
#> $ broadClass     <fct> ig, coding, ig, coding, coding, coding, coding, codi...
#> $ description    <fct> immunoglobulin heavy constant mu [Source:HGNC Symbol...
#> $ geneBiotype    <fct> IG_C_gene, protein_coding, IG_C_gene, protein_coding...
#> $ geneID         <chr> "ENSG00000211899", "ENSG00000105369", "ENSG000002115...
#> $ geneName       <fct> IGHM, CD79A, IGKC, S100A10, TMSB4X, S100A4, ANXA1, T...
#> $ seqCoordSystem <fct> chromosome, chromosome, chromosome, chromosome, chro...