Select Samples

Select Samples

# S4 method for bcbioSingleCell
selectSamples(object, ...)

Arguments

object
Object.

...
Columns to use for grep pattern matching. Supply a named character vector containing the column name and the grep pattern.

Value

bcbioSingleCell.

Details

Internally, pattern matching against sample and file names is applied using logical grep matching.

Note

Bracket based subsetting with [ also works on [bcbioSingleCell] objects. In this case, provide cellular barcode identifiers for columns and Ensembl gene identifiers for rows.

[` also works on [bcbioSingleCell]: R:

Examples

load(system.file(
    file.path("extdata", "filtered.rda"),
    package = "bcbioSingleCell"))

# bcbioSingleCell
# Quality control filtering must be applied first!
selectSamples(filtered, sampleName = "M1")

#> 1 sample(s) matched: M1

#> 243 cells matched

#> class: bcbioSingleCell 
#> dim: 1000 243 
#> metadata(38): version pipeline ... filterParams selectSamples
#> assays(1): assay
#> rownames(1000): ENSG00000000457 ENSG00000002586 ... ENSG00000283957
#>   ENSG00000284526
#> rowData names(11): ensgene symbol ... seqCoordSystem entrez
#> colnames(243): M1_AAACACTA_CTTAGGTA M1_AAACACTA_CTTCGATT ...
#>   M1_TGTTATCA_ACGCAGAG M1_TGTTATCA_CAAGTCAT
#> colData names(7): nUMI nGene ... mitoRatio nCount