bcbioSingleCell Object and Constructor

bcbioSingleCell is an S4 class that extends SingleCellExperiment, and is designed to store a bcbio single-cell RNA-seq analysis. This class contains read counts saved as a sparse matrix (dgCMatrix), sample metadata, and cell quality control metrics.

bcbioSingleCell(uploadDir, organism = NULL, sampleMetadataFile = NULL,
  interestingGroups = "sampleName", ensemblRelease = NULL,
  genomeBuild = NULL, transgeneNames = NULL, spikeNames = NULL,
  gffFile = NULL, ...)

Arguments

uploadDir
string. Path to final upload directory. This path is set when running "bcbio_nextgen -w template".

organism
string or NULL. Organism name. Use the full Latin name (e.g. "Homo sapiens"), since this will be input downstream to AnnotationHub and ensembldb, unless gffFile is set. If left NULL (not recommended), the function call will skip loading gene-level annotations into rowRanges(). This can be useful for poorly annotation genomes or experiments involving multiple genomes.

sampleMetadataFile
string or NULL. Sample barcode metadata file. Optional for runs with demultiplixed index barcodes (e.g. SureCell) and can be left NULL, but otherwise required for runs with multipliexed FASTQs containing multiple index barcodes (e.g. inDrop).

interestingGroups
character or NULL. Character vector of interesting groups. Must be formatted in camel case and intersect with sampleData() colnames.

ensemblRelease
scalar integer or NULL. Ensembl release version. If NULL, defaults to current release, and does not typically need to be user-defined. Passed to AnnotationHub for EnsDb annotation matching, unless gffFile is set.

genomeBuild
string or NULL. Ensembl genome build name (e.g. "GRCh38"). This will be passed to AnnotationHub for EnsDb annotation matching, unless gffFile is set.

transgeneNames
character or NULL. Vector indicating which assay() rows denote transgenes (e.g. EGFP, TDTOMATO).

spikeNames
character or NULL. Vector indicating which assay() rows denote spike-in sequences (e.g. ERCCs).

gffFile
string or NULL. Advanced use only; not recommended. By default, we recommend leaving this NULL for genomes that are supported on Ensembl. In this case, the row annotations (rowRanges()) will be obtained automatically from Ensembl by passing the organism, genomeBuild, and ensemblRelease arguments to AnnotationHub and ensembldb. For a genome that is not supported on Ensembl and/or AnnotationHub, a GFF/GTF (General Feature Format) file is required. Generally, we recommend using a GTF (GFFv2) file here over a GFF3 file if possible, although all GFF formats are supported. The function will internally generate transcript-to-gene mappings and construct a GRanges object containing the genomic ranges (rowRanges()).

...
Additional arguments, to be stashed in the metadata() slot.

Value

bcbioSingleCell.

Note

bcbioSingleCell extended SummarizedExperiment prior to v0.1.0, where we migrated to SingleCellExperiment.

Remote Data

When working in RStudio, we recommend connecting to the bcbio-nextgen run directory as a remote connection over sshfs.

See also

Examples

uploadDir <- system.file("extdata/indrops", package = "bcbioSingleCell")
x <- bcbioSingleCell(
    uploadDir = uploadDir,
    organism = "Homo sapiens",
    sampleMetadataFile = file.path(uploadDir, "metadata.csv"),
    ensemblRelease = 87L
)

#> 2018-01-01_bcbio

#> 1 samples detected:
#> multiplexed_AAAAAAAA

#> Reading project-summary.yaml

#> Data versions are missing

#> 1000 reads per cellular barcode cutoff detected

#> UMI type: surecell

#> Reading metadata.csv

#> Multiplexed samples detected

#> Reading counts as genes

#> Reading counts in mtx format

#> Reading cellular barcode distributions

#> Making GRanges from Ensembl with AnnotationHub 2.12.0 (2018-04-30)

#> Ensembl 87 EnsDb for Homo Sapiens

#> id: "AH53211"
#> organism: "Homo sapiens"
#> build: "GRCh38"
#> release: "87"
#> format: "genes"

#> Standardizing the metadata columns

#> Defining broad class using: geneName, geneBiotype, seqnames

#> Sorting ranges by geneID

#> Calculating cellular barcode metrics

#> 500 cells detected

#> 449 coding genes

#> 13 mitochondrial genes

#> 500 / 500 cellular barcodes passed pre-filtering (100%)

show(x)

#> bcbioSingleCell 0.2.0
#> class: SingleCellExperiment 
#> dim: 500 500 
#> metadata(31): version pipeline ... utilsSessionInfo devtoolsSessionInfo
#> assays(1): counts
#> rownames(500): ENSG00000000457 ENSG00000002586 ... ENSG00000283060
#>   ENSG00000283391
#> rowData names(7): broadClass description ... geneName seqCoordSystem
#> colnames(500): multiplexed_AAAAAAAA_AAACACTA_ATCCCACG
#>   multiplexed_AAAAAAAA_AAACACTA_CTTAGGTA ...
#>   multiplexed_AAAAAAAA_TGTTATCA_ACGCAGAG
#>   multiplexed_AAAAAAAA_TGTTATCA_CAAGTCAT
#> colData names(15): nCount sampleID ... aggregate revcomp
#> reducedDimNames(0):
#> spikeNames(0):
#> ====================================================================================================
#> Upload Dir: "/Users/mike/GitHub/hbc/bcbioSingleCell/inst/extdata/indrops"
#> Upload Date: 2018-01-01
#> R Load Date: 2018-08-02
#> Level: "genes"
#> Organism: "Homo sapiens"
#> Interesting Groups: "sampleName"
#> Metadata File: "/Users/mike/GitHub/hbc/bcbioSingleCell/inst/extdata/indrops/metadata.csv"
#> AnnotationHub: "AH53211"
#> Ensembl Release: "87"
#> Genome Build: "GRCh38"
#> Filtered: FALSE
#> zinbwave: FALSE

validObject(x)

#> [1] TRUE