Coerce Object

Coerce Object

Arguments

from
Class for which the coerce method will perform coercion.

Upgrade bcbioSingleCell to current version

This method adds support for upgrading bcbioSCDataSet objects to the latest bcbioSingleCell class version. This should be backwards compatible to bcbioSingleCell version 0.0.17. Previous objects saved using bcbioSinglecell (note case) will likely fail to load with newer versions of the package.

bcbioSingleCell to seurat

Interally Seurat::CreateSeuratObject() is called without applying any additional filtering cutoffs, since we have already defined them during our quality control analysis. Here we are passing the raw gene-level counts of the filtered cells into a new seurat class object, using as() object coercion.

bcbioSingleCell to SummarizedExperiment

Since bcbioSingleCell is an extension of SummarizedExperiment, this coercion method is very simple. Here we're simply dropping our @bcbio slot, which contains raw cellular barcodes and other bcbio-specific metadata.

See also

help(topic = "coerce", package = "methods").

Examples

bcb <- examples[["bcb"]]
pooled <- aggregateReplicates(bcb)

#> New sample IDs: M1

#> Remapping cellular barcodes to aggregate sample IDs

#> Aggregating counts

#> Aggregating raw cellular barcode counts

#> Calculating barcode metrics

#> 10874 cellular barcodes detected

#> 13 mitochondrial genes detected

#> 10874 cellular barcodes passed pre-filtering (100%)

#> Updating metadata

#> Preparing aggregated bcbioSingleCell object

filtered <- filterCells(pooled)

#> 10874 cells before filtering
#> 39297 genes before filtering

#>     1465 cells | minUMIs >= 1000

#>      983 cells | minGenes >= 500

#>      983 cells | maxGenes <= Inf

#>      591 cells | maxMitoRatio <= 0.1

#>      591 cells | minNovelty <= 0.8

#>    18893 genes | minCellsPerGene <= 3

#> 591 / 10874 cells passed filtering (5.43%)
#> 18893 / 39297 genes passed filtering (48.1%)

#> Filtering parameters:
#>   - >= 1000 UMI counts per cell
#>   - >= 500 genes per cell
#>   - <= Inf genes per cell
#>   - <= 0.1 mitochondrial abundance
#>   - >= 0.8 novelty score
#>   - >= 3 cells per gene

seurat <- as(filtered, "seurat")
seurat

#> An object of class seurat in project bcbioSingleCell 
#>  18893 genes across 591 samples.