Coerce Object

Coerce Object

Arguments

from
Class for which the coerce method will perform coercion.

bcbioSingleCell to seurat

Interally Seurat::CreateSeuratObject() is called without applying any additional filtering cutoffs, since we have already defined them during our quality control analysis. Here we are passing the raw gene-level counts of the filtered cells into a new seurat class object, using as() object coercion.

See also

help(topic = "coerce", package = "methods").

Examples

load(system.file("extdata/bcb_small.rda", package = "bcbioSingleCell"))

# bcbioSingleCell to seurat
as(bcb_small, "seurat")

#> `sampleNameAggregate` metadata column detected.
#> Use `aggregateReplicates()` to combine technical replicates.

#> An object of class seurat in project bcbioSingleCell 
#>  500 genes across 500 samples.