Methods for Coercing an Object to a Class

Methods for Coercing an Object to a Class

Value

Object of new class.

SingleCellExperiment to seurat

Interally Seurat::CreateSeuratObject() is called without applying any additional filtering cutoffs, since we have already defined them during our quality control analysis. Here we are passing the raw gene-level counts of the filtered cells into a new seurat class object.

seurat to SingleCellExperiment

Super basic S4 coercion support for taking the raw counts matrix from a seurat class object and coercing to a SingleCellExperiment.

See also

Examples

# SingleCellExperiment to seurat ====
x <- as(indrops_small, "seurat")
class(x)

#> [1] "seurat"
#> attr(,"package")
#> [1] "Seurat"

print(x)

#> An object of class seurat in project bcbioSingleCell 
#>  500 genes across 500 samples.

# seurat to SingleCellExperiment ====
x <- as(Seurat::pbmc_small, "SingleCellExperiment")
class(x)

#> [1] "SingleCellExperiment"
#> attr(,"package")
#> [1] "SingleCellExperiment"

print(x)

#> class: SingleCellExperiment 
#> dim: 230 80 
#> metadata(0):
#> assays(2): counts logcounts
#> rownames(230): MS4A1 CD79B ... SPON2 S100B
#> rowData names(4): gene gene.mean gene.dispersion gene.dispersion.scaled
#> colnames(80): ATGCCAGAACGACT CATGGCCTGTGCAT ... GGAACACTTCAGAC
#>   CTTGATTGATCTTC
#> colData names(6): nGene nUMI ... res.1 ident
#> reducedDimNames(2): PCA TSNE
#> spikeNames(0):