Coerce Object

Coerce Object

Value

New class, specified by the "to" argument.

bcbioSingleCell to seurat

Interally Seurat::CreateSeuratObject() is called without applying any additional filtering cutoffs, since we have already defined them during our quality control analysis. Here we are passing the raw gene-level counts of the filtered cells into a new seurat class object, using as() object coercion.

See also

Examples

# SingleCellExperiment to list ====
x <- as(bcb_small, "list")
class(x)

#> [1] "list"

names(x)

#>  [1] "int_elementMetadata" "int_colData"         "int_metadata"       
#>  [4] "reducedDims"         "rowRanges"           "colData"            
#>  [7] "assays"              "NAMES"               "elementMetadata"    
#> [10] "metadata"           

# SingleCellExperiment to seurat ====
x <- as(bcb_small, "seurat")

#> `aggregate` metadata column detected.
#> Use `aggregateReplicates()` to combine technical replicates.

class(x)

#> [1] "seurat"
#> attr(,"package")
#> [1] "Seurat"

print(x)

#> An object of class seurat in project bcbioSingleCell 
#>  500 genes across 500 samples.