Extract or replace parts of an object

Extract genes by row and cells by column.

# S4 method for bcbioSingleCell,ANY,ANY,ANY
[(x, i, j, ..., drop = FALSE)

Arguments

x
object from which to extract element(s) or in which to replace element(s).

i
indices specifying elements to extract or replace. Indices are numeric or character vectors or empty (missing) or NULL. Numeric values are coerced to integer as by as.integer (and hence truncated towards zero). Character vectors will be matched to the names of the object (or for matrices/arrays, the dimnames): see ‘Character indices’ below for further details.

For [-indexing only: i, j, … can be logical vectors, indicating elements/slices to select. Such vectors are recycled if necessary to match the corresponding extent. i, j, … can also be negative integers, indicating elements/slices to leave out of the selection.

When indexing arrays by [ a single argument i can be a matrix with as many columns as there are dimensions of x; the result is then a vector with elements corresponding to the sets of indices in each row of i.

An index value of NULL is treated as if it were integer(0).

j
indices specifying elements to extract or replace. Indices are numeric or character vectors or empty (missing) or NULL. Numeric values are coerced to integer as by as.integer (and hence truncated towards zero). Character vectors will be matched to the names of the object (or for matrices/arrays, the dimnames): see ‘Character indices’ below for further details.

For [-indexing only: i, j, … can be logical vectors, indicating elements/slices to select. Such vectors are recycled if necessary to match the corresponding extent. i, j, … can also be negative integers, indicating elements/slices to leave out of the selection.

When indexing arrays by [ a single argument i can be a matrix with as many columns as there are dimensions of x; the result is then a vector with elements corresponding to the sets of indices in each row of i.

An index value of NULL is treated as if it were integer(0).

...
Additional arguments.

drop
For matrices and arrays. If TRUE the result is coerced to the lowest possible dimension (see the examples). This only works for extracting elements, not for the replacement. See drop for further details.

Value

bcbioSingleCell.

Details

Refer to cell2sample() and selectSamples() if sample-level extraction is desired. Note that sampleID is slotted into colData and defines the cell-to-sample mappings.

Unfiltered cellular barcode distributions for the entire dataset, including cells not kept in the matrix will be dropped in favor of the nCount column of colData().

References

Becker, R. A., Chambers, J. M. and Wilks, A. R. (1988) The New S Language. Wadsworth & Brooks/Cole.

Examples

data(indrops)

cells <- head(colnames(indrops), 100L)
head(cells)

#> [1] "AAACACTA_CTTCGATT" "AAACTACA_CCACATTA" "AACTGCCT_GCAAGGAC"
#> [4] "AAGAAGGT_TCTGTGGT" "AAGCCTTC_TAAATAGG" "AATAAGGA_CCACATTA"

genes <- head(rownames(indrops), 100L)
head(genes)

#> [1] "ENSG00000071082" "ENSG00000100316" "ENSG00000106631" "ENSG00000108821"
#> [5] "ENSG00000112306" "ENSG00000115414"

## Subset by cell identifiers.
indrops[, cells]

#> bcbioSingleCell 0.3.9
#> uploadDir: /data00/packages/bcbioSingleCell/inst/extdata/indrops
#> dates(2): [bcbio] 2018-01-01; [R] 2019-04-02
#> level: genes
#> sampleMetadataFile: inst/extdata/indrops/metadata.csv
#> organism: Homo sapiens
#> interestingGroups: sampleName
#> filtered: FALSE
#> class: SingleCellExperiment 
#> dim: 50 100 
#> metadata(27): version pipeline ... wd sessionInfo
#> assays(1): counts
#> rownames(50): ENSG00000071082 ENSG00000100316 ... ENSG00000269028
#>   ENSG00000282105
#> rowData names(4): broadClass geneBiotype geneID geneName
#> colnames(100): AAACACTA_CTTCGATT AAACTACA_CCACATTA ...
#>   TGGGAATT_ATATAGGA TGTTATCA_ACGCAGAG
#> colData names(15): nUMI nGene ... aggregate revcomp
#> reducedDimNames(0):
#> spikeNames(0):

## Subset by genes.
indrops[genes, ]

#> bcbioSingleCell 0.3.9
#> uploadDir: /data00/packages/bcbioSingleCell/inst/extdata/indrops
#> dates(2): [bcbio] 2018-01-01; [R] 2019-04-02
#> level: genes
#> sampleMetadataFile: inst/extdata/indrops/metadata.csv
#> organism: Homo sapiens
#> interestingGroups: sampleName
#> filtered: FALSE
#> class: SingleCellExperiment 
#> dim: 50 100 
#> metadata(27): version pipeline ... wd sessionInfo
#> assays(1): counts
#> rownames(50): ENSG00000071082 ENSG00000100316 ... ENSG00000269028
#>   ENSG00000282105
#> rowData names(4): broadClass geneBiotype geneID geneName
#> colnames(100): AAACACTA_CTTCGATT AAACTACA_CCACATTA ...
#>   TGGGAATT_ATATAGGA TGTTATCA_ACGCAGAG
#> colData names(15): nUMI nGene ... aggregate revcomp
#> reducedDimNames(0):
#> spikeNames(0):

## Subset by both genes and cells.
indrops[genes, cells]

#> bcbioSingleCell 0.3.9
#> uploadDir: /data00/packages/bcbioSingleCell/inst/extdata/indrops
#> dates(2): [bcbio] 2018-01-01; [R] 2019-04-02
#> level: genes
#> sampleMetadataFile: inst/extdata/indrops/metadata.csv
#> organism: Homo sapiens
#> interestingGroups: sampleName
#> filtered: FALSE
#> class: SingleCellExperiment 
#> dim: 50 100 
#> metadata(27): version pipeline ... wd sessionInfo
#> assays(1): counts
#> rownames(50): ENSG00000071082 ENSG00000100316 ... ENSG00000269028
#>   ENSG00000282105
#> rowData names(4): broadClass geneBiotype geneID geneName
#> colnames(100): AAACACTA_CTTCGATT AAACTACA_CCACATTA ...
#>   TGGGAATT_ATATAGGA TGTTATCA_ACGCAGAG
#> colData names(15): nUMI nGene ... aggregate revcomp
#> reducedDimNames(0):
#> spikeNames(0):