Extract or Replace Parts of an Object

Extract genes by row and cells by column.

# S4 method for bcbioSingleCell,ANY,ANY,ANY
[(x, i, j, ..., drop = FALSE)

# S4 method for CellRanger,ANY,ANY,ANY
[(x, i, j, ..., drop = FALSE)

Arguments

x
object from which to extract element(s) or in which to replace element(s).

i
indices specifying elements to extract or replace. Indices are numeric or character vectors or empty (missing) or NULL. Numeric values are coerced to integer as by as.integer (and hence truncated towards zero). Character vectors will be matched to the names of the object (or for matrices/arrays, the dimnames): see ‘Character indices’ below for further details.

For [-indexing only: i, j, … can be logical vectors, indicating elements/slices to select. Such vectors are recycled if necessary to match the corresponding extent. i, j, … can also be negative integers, indicating elements/slices to leave out of the selection.

When indexing arrays by [ a single argument i can be a matrix with as many columns as there are dimensions of x; the result is then a vector with elements corresponding to the sets of indices in each row of i.

An index value of NULL is treated as if it were integer(0).

j
indices specifying elements to extract or replace. Indices are numeric or character vectors or empty (missing) or NULL. Numeric values are coerced to integer as by as.integer (and hence truncated towards zero). Character vectors will be matched to the names of the object (or for matrices/arrays, the dimnames): see ‘Character indices’ below for further details.

For [-indexing only: i, j, … can be logical vectors, indicating elements/slices to select. Such vectors are recycled if necessary to match the corresponding extent. i, j, … can also be negative integers, indicating elements/slices to leave out of the selection.

When indexing arrays by [ a single argument i can be a matrix with as many columns as there are dimensions of x; the result is then a vector with elements corresponding to the sets of indices in each row of i.

An index value of NULL is treated as if it were integer(0).

...
Additional arguments.

drop
For matrices and arrays. If TRUE the result is coerced to the lowest possible dimension (see the examples). This only works for extracting elements, not for the replacement. See drop for further details.

Value

SingleCellExperiment.

Details

Refer to cell2sample() and selectSamples() if sample-level extraction is desired. Note that sampleID is slotted into colData() and defines the cell-to-sample mappings.

Unfiltered cellular barcode distributions for the entire dataset, including cells not kept in the matrix will be dropped in favor of the nCount column of colData().

References

Becker, R. A., Chambers, J. M. and Wilks, A. R. (1988) The New S Language. Wadsworth & Brooks/Cole.

Examples

cells <- head(colnames(indrops_small), 100L)
head(cells)

#> [1] "AAACACTA_ATCCCACG" "AAACACTA_CTTAGGTA" "AAACACTA_CTTCGATT"
#> [4] "AAACCGCC_GAGTTTCG" "AAACCGCC_TAGCCTCG" "AAACGATC_TTTACCCT"

genes <- head(rownames(indrops_small), 100L)
head(genes)

#> [1] "ENSG00000000457" "ENSG00000002586" "ENSG00000004939" "ENSG00000005022"
#> [5] "ENSG00000005100" "ENSG00000008988"

## Subset by cell identifiers.
indrops_small[, cells]

#> bcbioSingleCell 0.99.0
#> http://bioinformatics.sph.harvard.edu/bcbioSingleCell
#> citation("bcbioSingleCell")
#> ──────────────────────────────────────────────────────────
#> Upload Dir: "/home/bioinfo/git/bcbioSingleCell-private/inst/extdata/indrops"
#> Upload Date: 2018-01-01
#> R Load Date: 2018-10-14
#> Level: "genes"
#> Organism: "Homo sapiens"
#> Interesting Groups: "sampleName"
#> Metadata File: "inst/extdata/indrops/metadata.csv"
#> Filtered: FALSE
#> ──────────────────────────────────────────────────────────
#> class: SingleCellExperiment 
#> dim: 500 100 
#> metadata(29): version pipeline ... sessionInfo subset
#> assays(1): counts
#> rownames(500): ENSG00000000457 ENSG00000002586 ... ENSG00000283060
#>   ENSG00000283391
#> rowData names(4): broadClass geneBiotype geneID geneName
#> colnames(100): AAACACTA_ATCCCACG AAACACTA_CTTAGGTA ...
#>   AGTTGAAC_CTTCGCAC ATAACAGG_GGATAAAG
#> colData names(15): nUMI nGene ... aggregate revcomp
#> reducedDimNames(0):
#> spikeNames(0):

## Subset by genes.
indrops_small[genes, ]

#> bcbioSingleCell 0.99.0
#> http://bioinformatics.sph.harvard.edu/bcbioSingleCell
#> citation("bcbioSingleCell")
#> ──────────────────────────────────────────────────────────
#> Upload Dir: "/home/bioinfo/git/bcbioSingleCell-private/inst/extdata/indrops"
#> Upload Date: 2018-01-01
#> R Load Date: 2018-10-14
#> Level: "genes"
#> Organism: "Homo sapiens"
#> Interesting Groups: "sampleName"
#> Metadata File: "inst/extdata/indrops/metadata.csv"
#> Filtered: FALSE
#> ──────────────────────────────────────────────────────────
#> class: SingleCellExperiment 
#> dim: 100 500 
#> metadata(29): version pipeline ... sessionInfo subset
#> assays(1): counts
#> rownames(100): ENSG00000000457 ENSG00000002586 ... ENSG00000111716
#>   ENSG00000111775
#> rowData names(4): broadClass geneBiotype geneID geneName
#> colnames(500): AAACACTA_ATCCCACG AAACACTA_CTTAGGTA ...
#>   TGTTATCA_ACGCAGAG TGTTATCA_CAAGTCAT
#> colData names(15): nUMI nGene ... aggregate revcomp
#> reducedDimNames(0):
#> spikeNames(0):

## Subset by both genes and cells.
indrops_small[genes, cells]

#> bcbioSingleCell 0.99.0
#> http://bioinformatics.sph.harvard.edu/bcbioSingleCell
#> citation("bcbioSingleCell")
#> ──────────────────────────────────────────────────────────
#> Upload Dir: "/home/bioinfo/git/bcbioSingleCell-private/inst/extdata/indrops"
#> Upload Date: 2018-01-01
#> R Load Date: 2018-10-14
#> Level: "genes"
#> Organism: "Homo sapiens"
#> Interesting Groups: "sampleName"
#> Metadata File: "inst/extdata/indrops/metadata.csv"
#> Filtered: FALSE
#> ──────────────────────────────────────────────────────────
#> class: SingleCellExperiment 
#> dim: 100 100 
#> metadata(29): version pipeline ... sessionInfo subset
#> assays(1): counts
#> rownames(100): ENSG00000000457 ENSG00000002586 ... ENSG00000111716
#>   ENSG00000111775
#> rowData names(4): broadClass geneBiotype geneID geneName
#> colnames(100): AAACACTA_ATCCCACG AAACACTA_CTTAGGTA ...
#>   AGTTGAAC_CTTCGCAC ATAACAGG_GGATAAAG
#> colData names(15): nUMI nGene ... aggregate revcomp
#> reducedDimNames(0):
#> spikeNames(0):