Fetch Data Functions

Fetch Data Functions

fetchGeneData(object, ...)

fetchPCAData(object, ...)

fetchTSNEData(object, ...)

fetchTSNEExpressionData(object, ...)

fetchUMAPData(object, ...)

fetchUMAPExpressionData(object, ...)

# S4 method for SingleCellExperiment
fetchGeneData(object, genes)

# S4 method for seurat
fetchGeneData(object, genes)

# S4 method for SingleCellExperiment
fetchPCAData(object, minimal = FALSE)

# S4 method for seurat
fetchPCAData(object, minimal = FALSE)

# S4 method for SingleCellExperiment
fetchTSNEData(object, minimal = FALSE)

# S4 method for seurat
fetchTSNEData(object, minimal = FALSE)

# S4 method for SingleCellExperiment
fetchTSNEExpressionData(object, genes)

# S4 method for seurat
fetchTSNEExpressionData(object, genes)

# S4 method for SingleCellExperiment
fetchUMAPData(object, minimal = FALSE)

# S4 method for seurat
fetchUMAPData(object, minimal = FALSE)

# S4 method for SingleCellExperiment
fetchUMAPExpressionData(object, genes)

# S4 method for seurat
fetchUMAPExpressionData(object, genes)

Arguments

object
Object.

...
Additional arguments.

genes
character. Gene identifiers. Must match the rownames of the object.

minimal
boolean. Return minimal data without metrics.

Value

fetchGeneData(): matrix.

fetchTSNEExpressionData(): data.frame containing t-SNE coordinates, sample metadata, and aggregate marker expression values (mean, median, and sum).

Other functions: data.frame, containing metrics.

See also

Examples

# SingleCellExperiment ====
object <- indrops_small
genes <- head(rownames(object))

# fetchGeneData
x <- fetchGeneData(object, genes = genes)
glimpse(x)

#>  num [1:500, 1:6] 1 1 4 3 3 2 5 1 3 3 ...
#>  - attr(*, "dimnames")=List of 2
#>   ..$ : chr [1:500] "multiplexed_AAAAAAAA_AAACACTA_ATCCCACG" "multiplexed_AAAAAAAA_AAACACTA_CTTAGGTA" "multiplexed_AAAAAAAA_AAACACTA_CTTCGATT" "multiplexed_AAAAAAAA_AAACCGCC_GAGTTTCG" ...
#>   ..$ : chr [1:6] "ENSG00000000457" "ENSG00000002586" "ENSG00000004939" "ENSG00000005022" ...

# seurat ====
object <- seurat_small
genes <- head(rownames(object))

# t-SNE
x <- fetchTSNEData(object)
glimpse(x)

#> Observations: 500
#> Variables: 20
#> $ sampleID          <fct> pbmc_1, pbmc_1, pbmc_1, pbmc_1, pbmc_1, pbmc_1, p...
#> $ nGene             <int> 2785, 2278, 1639, 2645, 2194, 2724, 1708, 2583, 1...
#> $ nUMI              <int> 12147, 8588, 7157, 10995, 7025, 11698, 7014, 1006...
#> $ nCoding           <int> 11413, 7791, 6659, 10388, 6504, 10945, 6610, 9084...
#> $ nMito             <int> 394, 385, 173, 346, 199, 416, 148, 331, 440, 187,...
#> $ log10GenesPerUMI  <dbl> 0.8433961, 0.8534941, 0.8339308, 0.8468845, 0.868...
#> $ mitoRatio         <dbl> 0.03243599, 0.04483000, 0.02417214, 0.03146885, 0...
#> $ orig.ident        <fct> pbmc, pbmc, pbmc, pbmc, pbmc, pbmc, pbmc, pbmc, p...
#> $ res.0.4           <chr> "4", "0", "1", "0", "3", "0", "1", "2", "0", "3",...
#> $ res.0.8           <chr> "4", "0", "1", "0", "3", "0", "1", "2", "0", "3",...
#> $ res.1.2           <chr> "5", "2", "0", "2", "4", "2", "0", "3", "1", "4",...
#> $ ident             <fct> 4, 0, 1, 0, 3, 0, 1, 2, 0, 3, 0, 3, 1, 3, 1, 0, 0...
#> $ sampleName        <fct> pbmc, pbmc, pbmc, pbmc, pbmc, pbmc, pbmc, pbmc, p...
#> $ description       <fct> pbmc, pbmc, pbmc, pbmc, pbmc, pbmc, pbmc, pbmc, p...
#> $ index             <fct> 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1...
#> $ interestingGroups <fct> pbmc, pbmc, pbmc, pbmc, pbmc, pbmc, pbmc, pbmc, p...
#> $ tSNE_1            <dbl> -1.173195, 11.621926, -23.954647, 5.792514, -4.33...
#> $ tSNE_2            <dbl> 5.83516248, -11.74702169, 0.09422639, -0.30668500...
#> $ centerX           <dbl> -0.2199901, 12.0832021, -21.1648072, 12.0832021, ...
#> $ centerY           <dbl> 5.805863, -5.746378, 1.267467, -5.746378, -9.6642...

# PCA
x <- fetchPCAData(object)
glimpse(x)

#> Observations: 500
#> Variables: 20
#> $ sampleID          <fct> pbmc_1, pbmc_1, pbmc_1, pbmc_1, pbmc_1, pbmc_1, p...
#> $ nGene             <int> 2785, 2278, 1639, 2645, 2194, 2724, 1708, 2583, 1...
#> $ nUMI              <int> 12147, 8588, 7157, 10995, 7025, 11698, 7014, 1006...
#> $ nCoding           <int> 11413, 7791, 6659, 10388, 6504, 10945, 6610, 9084...
#> $ nMito             <int> 394, 385, 173, 346, 199, 416, 148, 331, 440, 187,...
#> $ log10GenesPerUMI  <dbl> 0.8433961, 0.8534941, 0.8339308, 0.8468845, 0.868...
#> $ mitoRatio         <dbl> 0.03243599, 0.04483000, 0.02417214, 0.03146885, 0...
#> $ orig.ident        <fct> pbmc, pbmc, pbmc, pbmc, pbmc, pbmc, pbmc, pbmc, p...
#> $ res.0.4           <chr> "4", "0", "1", "0", "3", "0", "1", "2", "0", "3",...
#> $ res.0.8           <chr> "4", "0", "1", "0", "3", "0", "1", "2", "0", "3",...
#> $ res.1.2           <chr> "5", "2", "0", "2", "4", "2", "0", "3", "1", "4",...
#> $ ident             <fct> 4, 0, 1, 0, 3, 0, 1, 2, 0, 3, 0, 3, 1, 3, 1, 0, 0...
#> $ sampleName        <fct> pbmc, pbmc, pbmc, pbmc, pbmc, pbmc, pbmc, pbmc, p...
#> $ description       <fct> pbmc, pbmc, pbmc, pbmc, pbmc, pbmc, pbmc, pbmc, p...
#> $ index             <fct> 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1...
#> $ interestingGroups <fct> pbmc, pbmc, pbmc, pbmc, pbmc, pbmc, pbmc, pbmc, p...
#> $ PC1               <dbl> 2.4403582, 4.5195187, -7.3410868, 4.0058345, 0.46...
#> $ PC2               <dbl> -2.47091325, -0.18189599, 2.60355188, -0.27214185...
#> $ centerX           <dbl> 3.026150, 4.382754, -7.130071, 4.382754, -3.54459...
#> $ centerY           <dbl> -2.520556, 1.126445, 1.792681, 1.126445, -4.49748...

# UMAP
x <- fetchUMAPData(object)
glimpse(x)

#> Observations: 500
#> Variables: 20
#> $ sampleID          <fct> pbmc_1, pbmc_1, pbmc_1, pbmc_1, pbmc_1, pbmc_1, p...
#> $ nGene             <int> 2785, 2278, 1639, 2645, 2194, 2724, 1708, 2583, 1...
#> $ nUMI              <int> 12147, 8588, 7157, 10995, 7025, 11698, 7014, 1006...
#> $ nCoding           <int> 11413, 7791, 6659, 10388, 6504, 10945, 6610, 9084...
#> $ nMito             <int> 394, 385, 173, 346, 199, 416, 148, 331, 440, 187,...
#> $ log10GenesPerUMI  <dbl> 0.8433961, 0.8534941, 0.8339308, 0.8468845, 0.868...
#> $ mitoRatio         <dbl> 0.03243599, 0.04483000, 0.02417214, 0.03146885, 0...
#> $ orig.ident        <fct> pbmc, pbmc, pbmc, pbmc, pbmc, pbmc, pbmc, pbmc, p...
#> $ res.0.4           <chr> "4", "0", "1", "0", "3", "0", "1", "2", "0", "3",...
#> $ res.0.8           <chr> "4", "0", "1", "0", "3", "0", "1", "2", "0", "3",...
#> $ res.1.2           <chr> "5", "2", "0", "2", "4", "2", "0", "3", "1", "4",...
#> $ ident             <fct> 4, 0, 1, 0, 3, 0, 1, 2, 0, 3, 0, 3, 1, 3, 1, 0, 0...
#> $ sampleName        <fct> pbmc, pbmc, pbmc, pbmc, pbmc, pbmc, pbmc, pbmc, p...
#> $ description       <fct> pbmc, pbmc, pbmc, pbmc, pbmc, pbmc, pbmc, pbmc, p...
#> $ index             <fct> 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1...
#> $ interestingGroups <fct> pbmc, pbmc, pbmc, pbmc, pbmc, pbmc, pbmc, pbmc, p...
#> $ UMAP1             <dbl> 4.6446381, 5.3386016, -7.5365976, 2.7442835, -4.2...
#> $ UMAP2             <dbl> -0.08474138, 0.08443071, -3.77034678, -1.71767301...
#> $ centerX           <dbl> 4.382514, 1.194854, -6.451458, 1.194854, -3.14338...
#> $ centerY           <dbl> -0.5180848, -1.5345286, -4.6021462, -1.5345286, -...

# t-SNE gene expression
x <- fetchTSNEExpressionData(seurat_small, genes = genes)
glimpse(x)

#> Observations: 500
#> Variables: 23
#> $ sampleID          <fct> pbmc_1, pbmc_1, pbmc_1, pbmc_1, pbmc_1, pbmc_1, p...
#> $ nGene             <int> 2785, 2278, 1639, 2645, 2194, 2724, 1708, 2583, 1...
#> $ nUMI              <int> 12147, 8588, 7157, 10995, 7025, 11698, 7014, 1006...
#> $ nCoding           <int> 11413, 7791, 6659, 10388, 6504, 10945, 6610, 9084...
#> $ nMito             <int> 394, 385, 173, 346, 199, 416, 148, 331, 440, 187,...
#> $ log10GenesPerUMI  <dbl> 0.8433961, 0.8534941, 0.8339308, 0.8468845, 0.868...
#> $ mitoRatio         <dbl> 0.03243599, 0.04483000, 0.02417214, 0.03146885, 0...
#> $ orig.ident        <fct> pbmc, pbmc, pbmc, pbmc, pbmc, pbmc, pbmc, pbmc, p...
#> $ res.0.4           <chr> "4", "0", "1", "0", "3", "0", "1", "2", "0", "3",...
#> $ res.0.8           <chr> "4", "0", "1", "0", "3", "0", "1", "2", "0", "3",...
#> $ res.1.2           <chr> "5", "2", "0", "2", "4", "2", "0", "3", "1", "4",...
#> $ ident             <fct> 4, 0, 1, 0, 3, 0, 1, 2, 0, 3, 0, 3, 1, 3, 1, 0, 0...
#> $ sampleName        <fct> pbmc, pbmc, pbmc, pbmc, pbmc, pbmc, pbmc, pbmc, p...
#> $ description       <fct> pbmc, pbmc, pbmc, pbmc, pbmc, pbmc, pbmc, pbmc, p...
#> $ index             <fct> 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1...
#> $ interestingGroups <fct> pbmc, pbmc, pbmc, pbmc, pbmc, pbmc, pbmc, pbmc, p...
#> $ tSNE_1            <dbl> -1.173195, 11.621926, -23.954647, 5.792514, -4.33...
#> $ tSNE_2            <dbl> 5.83516248, -11.74702169, 0.09422639, -0.30668500...
#> $ centerX           <dbl> -0.2199901, 12.0832021, -21.1648072, 12.0832021, ...
#> $ centerY           <dbl> 5.805863, -5.746378, 1.267467, -5.746378, -9.6642...
#> $ mean              <dbl> 7.833333, 4.333333, 6.833333, 7.166667, 3.166667,...
#> $ median            <dbl> 2.0, 2.0, 2.0, 1.5, 2.0, 2.5, 0.5, 4.0, 0.5, 3.0,...
#> $ sum               <dbl> 47, 26, 41, 43, 19, 41, 42, 46, 13, 24, 36, 30, 6...

# UMAP gene expession
genes <- head(rownames(seurat_small))
x <- fetchUMAPExpressionData(seurat_small, genes = genes)
glimpse(x)

#> Observations: 500
#> Variables: 23
#> $ sampleID          <fct> pbmc_1, pbmc_1, pbmc_1, pbmc_1, pbmc_1, pbmc_1, p...
#> $ nGene             <int> 2785, 2278, 1639, 2645, 2194, 2724, 1708, 2583, 1...
#> $ nUMI              <int> 12147, 8588, 7157, 10995, 7025, 11698, 7014, 1006...
#> $ nCoding           <int> 11413, 7791, 6659, 10388, 6504, 10945, 6610, 9084...
#> $ nMito             <int> 394, 385, 173, 346, 199, 416, 148, 331, 440, 187,...
#> $ log10GenesPerUMI  <dbl> 0.8433961, 0.8534941, 0.8339308, 0.8468845, 0.868...
#> $ mitoRatio         <dbl> 0.03243599, 0.04483000, 0.02417214, 0.03146885, 0...
#> $ orig.ident        <fct> pbmc, pbmc, pbmc, pbmc, pbmc, pbmc, pbmc, pbmc, p...
#> $ res.0.4           <chr> "4", "0", "1", "0", "3", "0", "1", "2", "0", "3",...
#> $ res.0.8           <chr> "4", "0", "1", "0", "3", "0", "1", "2", "0", "3",...
#> $ res.1.2           <chr> "5", "2", "0", "2", "4", "2", "0", "3", "1", "4",...
#> $ ident             <fct> 4, 0, 1, 0, 3, 0, 1, 2, 0, 3, 0, 3, 1, 3, 1, 0, 0...
#> $ sampleName        <fct> pbmc, pbmc, pbmc, pbmc, pbmc, pbmc, pbmc, pbmc, p...
#> $ description       <fct> pbmc, pbmc, pbmc, pbmc, pbmc, pbmc, pbmc, pbmc, p...
#> $ index             <fct> 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1...
#> $ interestingGroups <fct> pbmc, pbmc, pbmc, pbmc, pbmc, pbmc, pbmc, pbmc, p...
#> $ UMAP1             <dbl> 4.6446381, 5.3386016, -7.5365976, 2.7442835, -4.2...
#> $ UMAP2             <dbl> -0.08474138, 0.08443071, -3.77034678, -1.71767301...
#> $ centerX           <dbl> 4.382514, 1.194854, -6.451458, 1.194854, -3.14338...
#> $ centerY           <dbl> -0.5180848, -1.5345286, -4.6021462, -1.5345286, -...
#> $ mean              <dbl> 7.833333, 4.333333, 6.833333, 7.166667, 3.166667,...
#> $ median            <dbl> 2.0, 2.0, 2.0, 1.5, 2.0, 2.5, 0.5, 4.0, 0.5, 3.0,...
#> $ sum               <dbl> 47, 26, 41, 43, 19, 41, 42, 46, 13, 24, 36, 30, 6...