Fetch t-SNE Locations, Cellular Metadata and Expression of Genes

This gets t-SNE locations, cellular metadata, expression of genes and the geometric mean of the gene expression from a seurat object.

fetchTSNEExpressionData(object, ...)

# S4 method for seurat
fetchTSNEExpressionData(object, genes)

Arguments

object
Object.

...
Additional arguments (for the S4 generic definition).

genes
Genes (by symbol name) of which to get expression data.

Value

tibble grouped by gene, containing t-SNE points, cellular metadata, and gene expression.

See also

Other t-SNE Utilities: fetchTSNEData, plotMarkerTSNE, plotTSNE

Examples

load(system.file(
    file.path("extdata", "seurat.rda"),
    package = "bcbioSingleCell"))

genes <- counts(seurat) %>% rownames() %>% head()
print(genes)

#> ENSG00000000003 ENSG00000000005 ENSG00000000419 ENSG00000000457 ENSG00000000460 
#>        "TSPAN6"          "TNMD"          "DPM1"         "SCYL3"      "C1orf112" 
#> ENSG00000000971 
#>           "CFH" 

fetchTSNEExpressionData(seurat, genes = genes) %>% glimpse()

#> Observations: 3,546
#> Variables: 22
#> $ gene              <chr> "C1orf112", "C1orf112", "C1orf112", "C1orf112", "...
#> $ cellID            <chr> "M1_AAACACGG_GTGGAGCT", "M1_AAACACTA_ATCCCACG", "...
#> $ expression        <dbl> 0.0000000, 0.0000000, 0.6179993, 0.0000000, 0.884...
#> $ geomean           <dbl> 0.3014372, 0.6351767, 0.2059998, 0.3426279, 0.747...
#> $ tSNE1             <dbl> -2.8040828, 7.5588808, 2.1844768, -17.2397443, -1...
#> $ tSNE2             <dbl> -9.7700315, 6.5948968, 15.9177634, -5.7968148, -0...
#> $ ident             <fctr> 5, 4, 4, 3, 3, 4, 2, 0, 1, 3, 0, 1, 2, 2, 4, 1, ...
#> $ nGene             <int> 946, 1530, 2859, 3296, 2075, 1984, 846, 2311, 848...
#> $ nUMI              <dbl> 1963, 4994, 11694, 16157, 7054, 7082, 2182, 8560,...
#> $ sampleID          <fctr> M1, M1, M1, M1, M1, M1, M1, M1, M1, M1, M1, M1, ...
#> $ nCoding           <dbl> 1749, 4653, 10820, 14808, 6547, 6506, 1862, 7295,...
#> $ nMito             <dbl> 87, 46, 442, 438, 184, 170, 142, 830, 68, 423, 94...
#> $ log10GenesPerUMI  <dbl> 0.9037240, 0.8610885, 0.8496179, 0.8359519, 0.861...
#> $ mitoRatio         <dbl> 0.044319918, 0.009211053, 0.037797161, 0.02710899...
#> $ nCount            <int> 28588, 43996, 95155, 124786, 60593, 59845, 18388,...
#> $ sampleName        <fctr> M1, M1, M1, M1, M1, M1, M1, M1, M1, M1, M1, M1, ...
#> $ description       <fctr> M1, M1, M1, M1, M1, M1, M1, M1, M1, M1, M1, M1, ...
#> $ interestingGroups <fctr> M1, M1, M1, M1, M1, M1, M1, M1, M1, M1, M1, M1, ...
#> $ orig.ident        <fctr> M1, M1, M1, M1, M1, M1, M1, M1, M1, M1, M1, M1, ...
#> $ res.0.8           <chr> "5", "4", "4", "3", "3", "4", "2", "0", "1", "3",...
#> $ centerX           <dbl> -3.804437, 8.679370, 8.679370, -14.453529, -14.45...
#> $ centerY           <dbl> -10.510737, 6.908902, 6.908902, -3.813257, -3.813...