Fetch t-SNE Locations, Cellular Metadata and Expression of Genes

This gets t-SNE locations, cellular metadata, expression of genes and the geometric mean of the gene expression from a seurat object.

fetchTSNEExpressionData(object, ...)

# S4 method for seurat
fetchTSNEExpressionData(object, genes)

Arguments

object
Object.

...
Additional arguments.

genes
Gene identifiers. Must match the rownames of the object.

Value

data.frame containing tSNE coordinates, sample metadata, and aggregate marker expression values (mean, median, and sum).

See also

Examples

# seurat ====
genes <- head(rownames(Seurat::pbmc_small))
fetchTSNEExpressionData(Seurat::pbmc_small, genes = genes) %>% glimpse()

#> Observations: 80
#> Variables: 16
#> $ nGene             <int> 47, 52, 50, 56, 53, 48, 36, 45, 36, 41, 31, 61, 4...
#> $ nUMI              <int> 70, 85, 87, 127, 173, 70, 64, 72, 52, 100, 62, 19...
#> $ origIdent         <fct> SeuratProject, SeuratProject, SeuratProject, Seur...
#> $ res0x8            <fct> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0...
#> $ res1              <fct> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0...
#> $ sampleID          <fct> SeuratProject, SeuratProject, SeuratProject, Seur...
#> $ sampleName        <fct> SeuratProject, SeuratProject, SeuratProject, Seur...
#> $ interestingGroups <fct> SeuratProject, SeuratProject, SeuratProject, Seur...
#> $ ident             <fct> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 1, 0...
#> $ tSNE1             <dbl> 14.4219083, 16.7237708, 0.9080303, -12.2713037, 1...
#> $ tSNE2             <dbl> 8.336022, 11.254907, 21.274887, 8.603055, 11.9711...
#> $ centerX           <dbl> -0.3324792, -0.3324792, -0.3324792, -0.3324792, -...
#> $ centerY           <dbl> 18.75982, 18.75982, 18.75982, 18.75982, 18.75982,...
#> $ mean              <dbl> 0.3333333, 0.1666667, 0.0000000, 0.0000000, 0.166...
#> $ median            <dbl> 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0,...
#> $ sum               <dbl> 2, 1, 0, 0, 1, 1, 0, 1, 0, 1, 22, 45, 31, 22, 26,...