Filter Cells

Apply gene detection, mitochondrial abundance, and novelty score cutoffs to cellular barcodes.

filterCells(object, ...)

# S4 method for bcbioSingleCell
filterCells(object, minUMIs = 1000L,
  maxUMIs = Inf, minGenes = 500L, maxGenes = Inf, maxMitoRatio = 0.1,
  minNovelty = 0.75, minCellsPerGene = 3L, quiet = FALSE)

# S4 method for bcbioSCDataSet
filterCells(object)

Arguments

object
Object.

...
Additional arguments (for the S4 generic definition).

minUMIs
Minimum number of UMI disambiguated counts per cell.

maxUMIs
Maximum number of UMI disambiguated counts per cell.

minGenes
Minimum number of genes detected.

maxGenes
Maximum number of genes detected.

maxMitoRatio
Maximum relative mitochondrial abundance (0-1 scale).

minNovelty
Minimum novelty score.

minCellsPerGene
Include genes with non-zero expression in at least this many cells.

quiet
If TRUE, don't show the filtering parameter summary.

Value

bcbioSingleCell object, with filtering information slotted into metadata() as filterCells and filterParams.

Details

The filtering cutoff values now support a named numeric vector. By default we recommend applying the same filtering cutoff to all samples. When matching the samples, be sure to use sampleID column (i.e. the rownames of sampleMetadata()).

See also

Seurat::CreateSeuratObject().

Other Data Management Utilities: aggregateReplicates, readCellTypeMarkersFile, subsetPerSample

Examples

load(system.file(
    file.path("extdata", "bcb.rda"),
    package = "bcbioSingleCell"))

# bcbioSingleCell
print(bcb)

#> class: bcbioSingleCell 
#> dim: 1000 500 
#> metadata(33): version pipeline ... devtoolsSessionInfo subset
#> assays(1): assay
#> rownames(1000): ENSG00000000457 ENSG00000002586 ... ENSG00000283957
#>   ENSG00000284526
#> rowData names(11): ensgene symbol ... seqCoordSystem entrez
#> colnames(500): run1_AGAGGATA_AAACACTA_CTTAGGTA
#>   run1_AGAGGATA_AAACACTA_CTTCGATT ... run2_AGAGGATA_TGTTATCA_ACGCAGAG
#>   run2_AGAGGATA_TGTTATCA_CAAGTCAT
#> colData names(7): nUMI nGene ... mitoRatio nCount

filterCells(bcb)

#> Filtering parameters:
#>   - >= 1000 UMI counts per cell
#>   - <= Inf UMI counts per cell
#>   - >= 500 genes per cell
#>   - <= Inf genes per cell
#>   - <= 0.1 mitochondrial abundance
#>   - >= 0.75 novelty score
#>   - >= 3 cells per gene
#> ============================================================
#> 1000 genes / 500 cells
#>      500 cells | minUMIs >= 1000
#>      500 cells | maxUMIs <= Inf
#>      499 cells | minGenes >= 500
#>      499 cells | maxGenes <= Inf
#>      421 cells | maxMitoRatio <= 0.1
#>      421 cells | minNovelty <= 0.75
#>     1000 genes | minCellsPerGene <= 3
#> ============================================================
#> 421 / 500 cells passed filtering (84.2%)
#> 1000 / 1000 genes passed filtering (100%)

#> class: bcbioSingleCell 
#> dim: 1000 421 
#> metadata(38): version pipeline ... filterSummary originalVersion
#> assays(1): assay
#> rownames(1000): ENSG00000000457 ENSG00000002586 ... ENSG00000283957
#>   ENSG00000284526
#> rowData names(11): ensgene symbol ... seqCoordSystem entrez
#> colnames(421): run1_AGAGGATA_AAACACTA_CTTAGGTA
#>   run1_AGAGGATA_AAACACTA_CTTCGATT ... run2_AGAGGATA_TGTTATCA_ACGCAGAG
#>   run2_AGAGGATA_TGTTATCA_CAAGTCAT
#> colData names(7): nUMI nGene ... mitoRatio nCount