Filter Cells

Apply gene detection, novelty score, and mitochondrial abundance cutoffs to cellular barcodes. By default we recommend applying the same filtering cutoff to all samples. The filtering parameters now support per-sample cutoffs, defined using a named numeric vector. When matching per sample, be sure to use the sampleNames() return values (i.e. the sampleName column in sampleData()).

filterCells(object, ...)

# S4 method for SingleCellExperiment
filterCells(object, nCells = Inf,
  minUMIs = 1L, maxUMIs = Inf, minGenes = 1L, maxGenes = Inf,
  minNovelty = 0L, maxMitoRatio = 1L, minCellsPerGene = 1L)

Arguments

object
Object.

...
Additional arguments.

nCells
scalar integer. Expected number of cells per sample.

minUMIs
scalar integer. Minimum number of UMI disambiguated counts per cell.

maxUMIs
scalar integer. Maximum number of UMI disambiguated counts per cell.

minGenes
scalar integer. Minimum number of genes detected.

maxGenes
scalar integer. Maximum number of genes detected.

minNovelty
scalar integer (0-1). Minimum novelty score (log10 genes per UMI).

maxMitoRatio
scalar integer (0-1). Maximum relative mitochondrial abundance.

minCellsPerGene
scalar integer. Include genes with non-zero expression in at least this many cells.

Value

bcbioSingleCell, with filtering information slotted into metadata() as filterCells and filterParams.

See also

Examples

object <- indrops_small
show(object)

#> bcbioSingleCell 0.99.0
#> http://bioinformatics.sph.harvard.edu/bcbioSingleCell
#> citation("bcbioSingleCell")
#> ────────────────────────────────────────────────────────────────────────
#> Upload Dir: "/home/bioinfo/git/bcbioSingleCell-private/inst/extdata/indrops"
#> Upload Date: 2018-01-01
#> R Load Date: 2018-09-27
#> Level: "genes"
#> Organism: "Homo sapiens"
#> Interesting Groups: "sampleName"
#> Metadata File: "inst/extdata/indrops/metadata.csv"
#> Filtered: FALSE
#> ────────────────────────────────────────────────────────────────────────
#> class: SingleCellExperiment 
#> dim: 500 500 
#> metadata(29): version pipeline ... wd sessionInfo
#> assays(1): counts
#> rownames(500): ENSG00000000457 ENSG00000002586 ... ENSG00000283060
#>   ENSG00000283391
#> rowData names(4): broadClass geneBiotype geneID geneName
#> colnames(500): AAACACTA_ATCCCACG AAACACTA_CTTAGGTA ...
#>   TGTTATCA_ACGCAGAG TGTTATCA_CAAGTCAT
#> colData names(15): nUMI nGene ... aggregate revcomp
#> reducedDimNames(0):
#> spikeNames(0):

x <- filterCells(object)

#> Returning object unmodified.

#> Returning object unmodified.

#> Parameters:
#>   - >= 1 UMIs per cell
#>   - <= Inf UMIs per cell
#>   - >= 1 genes per cell
#>   - <= Inf genes per cell
#>   - >= 0 novelty score
#>   - <= 1 mitochondrial abundance
#>   - == Inf cells per sample
#>   - >= 1 cells per gene
#> ────────────────────────────────────────────────────────────────────────
#> Cells:
#>      500 cells | prefilter
#>      500 cells | minUMIs >= 1
#>      500 cells | maxUMIs <= Inf
#>      500 cells | minGenes >= 1
#>      500 cells | maxGenes <= Inf
#>      500 cells | minNovelty <= 0
#>      500 cells | maxMitoRatio <= 1
#>      500 cells | nCells == Inf
#>      500 of 500 cells passed filtering (100%)
#> rep_1 
#>   500 
#> ────────────────────────────────────────────────────────────────────────
#> Genes:
#>      500 genes | prefilter
#>      500 genes | minCellsPerGene >= 1
#>      500 of 500 genes passed filtering (100%)

show(x)

#> bcbioSingleCell 0.99.0
#> http://bioinformatics.sph.harvard.edu/bcbioSingleCell
#> citation("bcbioSingleCell")
#> ────────────────────────────────────────────────────────────────────────
#> Upload Dir: "/home/bioinfo/git/bcbioSingleCell-private/inst/extdata/indrops"
#> Upload Date: 2018-01-01
#> R Load Date: 2018-09-27
#> Level: "genes"
#> Organism: "Homo sapiens"
#> Interesting Groups: "sampleName"
#> Metadata File: "inst/extdata/indrops/metadata.csv"
#> Filtered: TRUE
#> ────────────────────────────────────────────────────────────────────────
#> class: SingleCellExperiment 
#> dim: 500 500 
#> metadata(32): version pipeline ... filterParams filterSummary
#> assays(1): counts
#> rownames(500): ENSG00000000457 ENSG00000002586 ... ENSG00000283060
#>   ENSG00000283391
#> rowData names(4): broadClass geneBiotype geneID geneName
#> colnames(500): AAACACTA_ATCCCACG AAACACTA_CTTAGGTA ...
#>   TGTTATCA_ACGCAGAG TGTTATCA_CAAGTCAT
#> colData names(15): nUMI nGene ... aggregate revcomp
#> reducedDimNames(0):
#> spikeNames(0):

metadata(x)$filterParams

#> $nCells
#> [1] Inf
#> 
#> $minUMIs
#> [1] 1
#> 
#> $maxUMIs
#> [1] Inf
#> 
#> $minGenes
#> [1] 1
#> 
#> $maxGenes
#> [1] Inf
#> 
#> $minNovelty
#> [1] 0
#> 
#> $maxMitoRatio
#> [1] 1
#> 
#> $minCellsPerGene
#> [1] 1
#> 

# Per sample cutoffs
sampleNames(object)

#> AAACACTA_ATCCCACG AAACACTA_CTTAGGTA AAACACTA_CTTCGATT AAACCGCC_GAGTTTCG 
#>           "rep_1"           "rep_1"           "rep_1"           "rep_1" 
#> AAACCGCC_TAGCCTCG AAACGATC_TTTACCCT AAACTACA_CCACATTA AAAGAGGC_ACCTTGCC 
#>           "rep_1"           "rep_1"           "rep_1"           "rep_1" 
#> AAAGTCAT_TTGAGGGT AAATTCCG_TTTATCAC AACAGAAC_TTTACCCT AACGATTT_ATTCAGTA 
#>           "rep_1"           "rep_1"           "rep_1"           "rep_1" 
#> AACGATTT_TCCAGGGA AACGCCAA_TAAGACGG AACGGTAG_CACGGACT AACGTTAC_ACCCACGA 
#>           "rep_1"           "rep_1"           "rep_1"           "rep_1" 
#> AACGTTAC_GCAGTCGA AACGTTAC_TCGGTTTA AACTGCCT_GAGAGTAT AACTGCCT_GCAAGGAC 
#>           "rep_1"           "rep_1"           "rep_1"           "rep_1" 
#> AACTGCCT_GGGTGTAT AACTGCCT_TTAGGACC AAGAAGGT_ATTCAGTA AAGAAGGT_TCTGTGGT 
#>           "rep_1"           "rep_1"           "rep_1"           "rep_1" 
#> AAGAGTAT_CCCTAACC AAGCCTTC_TAAATAGG AAGCCTTC_TCACGTTT AAGCTCCT_AGACACCA 
#>           "rep_1"           "rep_1"           "rep_1"           "rep_1" 
#> AAGGATGA_CTAGTAAC AAGGGACC_CAGGTGTC AAGTCCAA_AACCTGAC AAGTCGGG_ACGCTCAG 
#>           "rep_1"           "rep_1"           "rep_1"           "rep_1" 
#> AAGTCGGG_GGAAACAG AAGTGAGA_GAGGCTGA AATAAGGA_AACCTGAC AATAAGGA_CCACATTA 
#>           "rep_1"           "rep_1"           "rep_1"           "rep_1" 
#> AATATGAC_GTAATCCC AATCGAAG_CCCAAGCA AATCGTTC_CCCTAACC AATGTATG_ATACTCTT 
#>           "rep_1"           "rep_1"           "rep_1"           "rep_1" 
#> AATGTATG_CACAAGGC ACAAGTAG_GACAACTT ACAAGTAG_TAACCCGT ACAAGTAG_TCTGTGGT 
#>           "rep_1"           "rep_1"           "rep_1"           "rep_1" 
#> ACAGATAA_TGGTTTCT ACCAACCC_CACAACAG ACCAACCC_CCATTACT ACCAAGGG_GCACCTCT 
#>           "rep_1"           "rep_1"           "rep_1"           "rep_1" 
#> ACCACAGA_ATACCCAG ACCACAGA_CGATTGAT ACCAGTTT_GCGCATTC ACCCATGC_TGCCATCG 
#>           "rep_1"           "rep_1"           "rep_1"           "rep_1" 
#> ACCCGATT_GGGAACGA ACCCTCAA_CTGCGTTG ACCTGAAG_CCTACTAG ACCTGAAG_GAGCGGTA 
#>           "rep_1"           "rep_1"           "rep_1"           "rep_1" 
#> ACCTTCCC_ATACTCTT ACGGGTTA_CATCTCCC ACGTAAAC_AGAACGGG ACTAATTG_CTTTAATC 
#>           "rep_1"           "rep_1"           "rep_1"           "rep_1" 
#> ACTAATTG_TGACCAGT ACTACCCG_CGTACCTA ACTAGAGC_TCGACACC ACTGGTCA_AGTTTAGA 
#>           "rep_1"           "rep_1"           "rep_1"           "rep_1" 
#> ACTGGTCA_TGATAACA AGAAACCA_ATACTCTT AGAAACCA_GTGTCGGA AGAAGCTT_CGGAATTT 
#>           "rep_1"           "rep_1"           "rep_1"           "rep_1" 
#> AGAAGCTT_GATGTGAG AGAGAGAC_ATTGGGCC AGAGAGAC_GTATGAGT AGCACGTA_TTCGCTGG 
#>           "rep_1"           "rep_1"           "rep_1"           "rep_1" 
#> AGCATGCC_GGGAAATC AGCCATCT_GTGTCGGA AGCCATCT_TTTAACAG AGCGTGGT_CTTCGATT 
#>           "rep_1"           "rep_1"           "rep_1"           "rep_1" 
#> AGCTCCAC_CCTGACAC AGCTCCAC_GTCCATGT AGCTTCGA_AGATGGCT AGGCAATA_CTATAGAG 
#>           "rep_1"           "rep_1"           "rep_1"           "rep_1" 
#> AGGCAATA_GCAAAGCC AGGCGTTT_AGCACCAC AGGGACTG_CTAGACCC AGGGACTG_TCCGTCCA 
#>           "rep_1"           "rep_1"           "rep_1"           "rep_1" 
#> AGGGTAAA_GGGAAGGT AGGGTAAA_TGGCCTGT AGGTAAGC_TCATGAGG AGGTAAGC_TCCCAATC 
#>           "rep_1"           "rep_1"           "rep_1"           "rep_1" 
#> AGGTATAT_CTTACGGG AGGTATAT_TACTTGTG AGGTTCCC_GCCTCTTT AGGTTCCC_GTGTCGGA 
#>           "rep_1"           "rep_1"           "rep_1"           "rep_1" 
#> AGTAATGG_ATACCCAG AGTAATGG_TGTGCTCA AGTAGTTA_AGAACGGG AGTCACAA_AACAATCC 
#>           "rep_1"           "rep_1"           "rep_1"           "rep_1" 
#> AGTCCGTG_TATTGCCT AGTGCTTC_GGAATAGA AGTTGAAC_CTTCGCAC ATAACAGG_GGATAAAG 
#>           "rep_1"           "rep_1"           "rep_1"           "rep_1" 
#> ATAACAGG_TAACCATC ATACACCC_GCATGGGT ATAGATGT_ATCCGCTA ATAGATGT_CTGTCTGG 
#>           "rep_1"           "rep_1"           "rep_1"           "rep_1" 
#> ATATGCAA_GGCCCTTA ATATGCAA_GGCGGTTT ATCAATCG_ATTCGACG ATCAATCG_GTTGTCAT 
#>           "rep_1"           "rep_1"           "rep_1"           "rep_1" 
#> ATCCCACC_CCCATAGC ATCCTAGT_TCGAAGCT ATCGCGCT_ACAGCGGA ATCGCGCT_AGAGGTGG 
#>           "rep_1"           "rep_1"           "rep_1"           "rep_1" 
#> ATCGCGCT_CACAGTTT ATCGCGCT_TCCCAATC ATCTTGGC_TCCTTATT ATGACAAC_AAAGTCGG 
#>           "rep_1"           "rep_1"           "rep_1"           "rep_1" 
#> ATGACTTG_TGCAAGGG ATGCGGAG_AACCCTTG ATGCGGAG_GATTTACC ATGGTCTG_TTTACCCT 
#>           "rep_1"           "rep_1"           "rep_1"           "rep_1" 
#> ATGTGCCG_CCGCAACT ATGTGCCG_GGTCACAG ATTACCTT_CGTAGAAG ATTACCTT_TTGCATAT 
#>           "rep_1"           "rep_1"           "rep_1"           "rep_1" 
#> ATTCTGAG_GAGCGGTA ATTCTGAG_TAACCCGT ATTGAAGT_ACGTATAC ATTGGCCC_AACAATCC 
#>           "rep_1"           "rep_1"           "rep_1"           "rep_1" 
#> ATTTGTTG_CTTTAATC CAAACATT_CTCAGAAT CAAACATT_GGGAGGTA CAACGCAG_CTCGCGTA 
#>           "rep_1"           "rep_1"           "rep_1"           "rep_1" 
#> CAACGCAG_TGATGCCC CAAGGAAT_TTATAGCC CAAGGGTT_ACCCACGA CAAGGGTT_CAGACCAT 
#>           "rep_1"           "rep_1"           "rep_1"           "rep_1" 
#> CAAGGTAC_CAGTTTGC CAATTCTC_ACGCCATT CAATTCTC_GTTCTGTT CAATTCTC_TGATAACA 
#>           "rep_1"           "rep_1"           "rep_1"           "rep_1" 
#> CACAACCT_GGAGAAGC CACTAACC_AGTGAAAG CAGATGGG_TGTCTTTC CAGGTTGC_ATATACCT 
#>           "rep_1"           "rep_1"           "rep_1"           "rep_1" 
#> CAGTTTAA_TAGTGGAC CATGCTGC_TTTATCAC CCACGTTG_TGGAGGAG CCATATGA_TGGCTACC 
#>           "rep_1"           "rep_1"           "rep_1"           "rep_1" 
#> CCATGCAT_CAGTTTGC CCATGCAT_GCGCATTC CCATGCAT_TTCCGCTC CCCGTTCT_AAAGCCTA 
#>           "rep_1"           "rep_1"           "rep_1"           "rep_1" 
#> CCCTCTTG_CTTACGGG CCCTGTTT_TGGTATGA CCCTTGCA_AGCAGAAC CCGACTTT_GGCCTAAG 
#>           "rep_1"           "rep_1"           "rep_1"           "rep_1" 
#> CCGAGATC_ATGGGCAC CCGAGATC_GGAAACAG CCGATACG_CAAGAGGG CCGATACG_CTCGATGC 
#>           "rep_1"           "rep_1"           "rep_1"           "rep_1" 
#> CCGGAAAT_AACCTGAC CCGGAAAT_ATCAGCGC CCGGAAAT_GTTGTCAT CCGGAAAT_TACCGGCA 
#>           "rep_1"           "rep_1"           "rep_1"           "rep_1" 
#> CCGGAAAT_TCGACGGT CCGTAGCT_ACGCCATT CCGTAGCT_CGTATTTC CCGTAGCT_TTATCTGT 
#>           "rep_1"           "rep_1"           "rep_1"           "rep_1" 
#> CCGTCTTA_GTTGTCAT CCTACGCT_ACTCACCG CCTACGCT_AGCAGAAC CCTACGCT_CCTCGTCT 
#>           "rep_1"           "rep_1"           "rep_1"           "rep_1" 
#> CCTACGCT_TTCCGCTC CCTATTTA_ACGCAGAG CCTATTTA_GCAGTCGA CCTATTTA_GCGTTGCT 
#>           "rep_1"           "rep_1"           "rep_1"           "rep_1" 
#> CCTATTTA_GGGAAATC CCTATTTA_TGCAAGGG CCTCATGA_CTTCGCAC CGACATTT_CGTGGATA 
#>           "rep_1"           "rep_1"           "rep_1"           "rep_1" 
#> CGAGGCTA_CACAAGGC CGATGGCA_TCCCGTAG CGGACTAA_GTCCGTAC CGGCTGTA_GCCGGATT 
#>           "rep_1"           "rep_1"           "rep_1"           "rep_1" 
#> CGTCGAAT_CGGACAAC CGTCGAAT_GAGTTTCG CGTGTACA_CCGTGTTT CGTGTACA_CGTACCTA 
#>           "rep_1"           "rep_1"           "rep_1"           "rep_1" 
#> CGTGTACA_GCAACCTG CGTGTGTT_AGACACCA CGTGTGTT_AGGAAGAC CGTGTGTT_CTCTTGAC 
#>           "rep_1"           "rep_1"           "rep_1"           "rep_1" 
#> CGTTGCCT_CGTGGATA CGTTGCCT_CTTCTTCG CGTTTCGT_GACAGATA CGTTTCGT_GAGAGTAT 
#>           "rep_1"           "rep_1"           "rep_1"           "rep_1" 
#> CTAACGCC_TAGCCTCG CTAGCACG_AATCGGGT CTAGCACG_GGCCCTTA CTAGCACG_TCTGTGGT 
#>           "rep_1"           "rep_1"           "rep_1"           "rep_1" 
#> CTAGCACG_TTGGCGTT CTAGTAGG_TTGAGGGT CTAGTAGG_TTTGTGTC CTCAAACA_GTGGATGG 
#>           "rep_1"           "rep_1"           "rep_1"           "rep_1" 
#> CTCACATC_ACCCACGA CTCACATC_CCTCGTCT CTCCCAAA_ATGGGCAC CTCCTCCA_CGGGTAGT 
#>           "rep_1"           "rep_1"           "rep_1"           "rep_1" 
#> CTCCTCCA_CGTATTTC CTCCTCCA_CTTCGCAC CTCCTCCA_TGTACACG CTCTATAG_CTTCGCAC 
#>           "rep_1"           "rep_1"           "rep_1"           "rep_1" 
#> CTCTATAG_GCAAAGCC CTCTATAG_TGACGGAC CTGAAGGG_CCTACTAG CTGAAGGG_CGGCACAT 
#>           "rep_1"           "rep_1"           "rep_1"           "rep_1" 
#> CTGAGCGT_CGACGTCA CTGAGCGT_GGTCCCTT CTGGAACA_TGGTTTCT CTGTCGCA_AGTGATGC 
#>           "rep_1"           "rep_1"           "rep_1"           "rep_1" 
#> CTGTCGCA_CCTATTCA CTGTCGCA_TGATAACA CTGTCGCA_TTTAACAG CTGTTAAA_AATGAATG 
#>           "rep_1"           "rep_1"           "rep_1"           "rep_1" 
#> CTGTTAAA_GACAGATA CTTAGGCC_AGAAGTCC CTTAGGCC_CCTATTCA CTTAGTGT_GAGAGTAT 
#>           "rep_1"           "rep_1"           "rep_1"           "rep_1" 
#> CTTAGTGT_GTCGTCGT CTTAGTGT_TCCAGGGA CTTAGTGT_TTTAACAG CTTCTACG_CTTCTTCG 
#>           "rep_1"           "rep_1"           "rep_1"           "rep_1" 
#> CTTCTACG_GCAGGGTA CTTTATCC_AGGATTCA CTTTATCC_CGAACGTA CTTTATCC_CTACCGTT 
#>           "rep_1"           "rep_1"           "rep_1"           "rep_1" 
#> CTTTATCC_TCGGTTTA CTTTCACT_TGTAAAGG CTTTGGAC_AAACTGGT GAAAGACA_TTCCGCTC 
#>           "rep_1"           "rep_1"           "rep_1"           "rep_1" 
#> GAAATACG_AGCTATGA GAAATACG_CAGTCCCT GAATCCCA_TCCCAATC GAATGCGC_TGCGACAG 
#>           "rep_1"           "rep_1"           "rep_1"           "rep_1" 
#> GACACAAA_ACGCCATT GACACAAA_GCAGGGTA GACACCTG_CTCAGAAT GACTAGCG_GAAGTGCC 
#>           "rep_1"           "rep_1"           "rep_1"           "rep_1" 
#> GAGAAACC_ACAGCGGA GAGAAACC_GCAAGGAC GAGAAACC_TTGATTCT GAGCGGAA_GTAACGTT 
#>           "rep_1"           "rep_1"           "rep_1"           "rep_1" 
#> GAGGGTCA_ACGAAACG GAGGGTCA_GCAGTCGA GAGTGTAC_ACGCAGAG GATACGCA_GATCGTTT 
#>           "rep_1"           "rep_1"           "rep_1"           "rep_1" 
#> GATGCAGA_AACCTGAC GATGCAGA_CAGATTAC GATTAAAG_ACTAGCCA GATTAAAG_GTTACGAT 
#>           "rep_1"           "rep_1"           "rep_1"           "rep_1" 
#> GATTACTT_ATGAAGCA GATTACTT_CTAGATTG GATTACTT_GAGAATTG GATTACTT_GATTGCGA 
#>           "rep_1"           "rep_1"           "rep_1"           "rep_1" 
#> GATTACTT_TTGGACTT GATTACTT_TTTACCCT GATTGGGA_AACACACG GATTGGGA_AATGTTTG 
#>           "rep_1"           "rep_1"           "rep_1"           "rep_1" 
#> GATTGGGA_AGTGATGC GATTTCCC_ATGTTGGC GATTTCCC_TGGAGGAG GCAAACTG_CTTCAGGT 
#>           "rep_1"           "rep_1"           "rep_1"           "rep_1" 
#> GCAAACTG_GGGAACCT GCAAACTG_TTTACCCT GCACTCAG_AAGTAATC GCACTCAG_CCCATAGC 
#>           "rep_1"           "rep_1"           "rep_1"           "rep_1" 
#> GCACTCAG_CGTACCTA GCATCGAG_CATCTCCC GCATCGAG_GCTTACCT GCCAAAGC_GCCTCTTT 
#>           "rep_1"           "rep_1"           "rep_1"           "rep_1" 
#> GCCAACAT_CGTGGATA GCCAACAT_GATCGTTT GCCAACAT_TGAAATGA GCCTGGTA_CGCTCTCA 
#>           "rep_1"           "rep_1"           "rep_1"           "rep_1" 
#> GCCTGGTA_CTATAGAG GCCTGGTA_GGCTACTA GCCTTGTG_AGAACGGG GCCTTGTG_AGCGCGAT 
#>           "rep_1"           "rep_1"           "rep_1"           "rep_1" 
#> GCGCTGAT_CTCGATGC GCGCTGAT_GACAAAGG GCGCTGAT_GTCAATAC GCGTGCAA_CGCTCTCA 
#>           "rep_1"           "rep_1"           "rep_1"           "rep_1" 
#> GCGTGCAA_TCTGTGGT GCGTGCAA_TTGTTTAC GCTAAGTT_ATTCCTTG GCTAAGTT_GCGCATTC 
#>           "rep_1"           "rep_1"           "rep_1"           "rep_1" 
#> GCTAAGTT_TCTCAACC GCTCGTAG_ACCTTGCC GGAACGAA_CGTAGAAG GGAACGAA_CTAGTAAC 
#>           "rep_1"           "rep_1"           "rep_1"           "rep_1" 
#> GGAACGAA_TCCGTCCA GGAACGAA_TTGCACGC GGACTTCT_CCCGACTT GGAGTAAG_CTTCTGGA 
#>           "rep_1"           "rep_1"           "rep_1"           "rep_1" 
#> GGATTGTT_CAGTTTGC GGATTGTT_GCGTTGCT GGCACTTC_GAATTCGT GGCCCAAT_GTTTGTTT 
#>           "rep_1"           "rep_1"           "rep_1"           "rep_1" 
#> GGCGACAA_AAGTAATC GGCGACAA_CGTCAGCA GGCGACAA_GACAGATA GGCGACAA_TTCCGAGT 
#>           "rep_1"           "rep_1"           "rep_1"           "rep_1" 
#> GGCGACAA_TTGATTCT GGCTTTGC_AACCCTTG GGCTTTGC_AGCGAAGT GGCTTTGC_TGGAGGAG 
#>           "rep_1"           "rep_1"           "rep_1"           "rep_1" 
#> GGCTTTGC_TTAGGACC GGGATTAC_AAATGTCG GGGATTAC_ATTCGACG GGGCATCA_GCTTACCT 
#>           "rep_1"           "rep_1"           "rep_1"           "rep_1" 
#> GGGTCATT_TACGTGCT GGGTCTAG_AGCGAAGT GGGTCTAG_CTTCTGGA GGTAAATC_AAACAGGG 
#>           "rep_1"           "rep_1"           "rep_1"           "rep_1" 
#> GGTAGCCA_CTGTCTGG GGTAGCCA_TAACTACT GGTAGCCA_TTTGTGTC GGTCCTAA_AGCAGAAC 
#>           "rep_1"           "rep_1"           "rep_1"           "rep_1" 
#> GGTCCTAA_GACGATGG GGTCCTAA_GGTTTCTC GGTCCTAA_TGGCCTGT GGTCTTTC_AGCGCGAT 
#>           "rep_1"           "rep_1"           "rep_1"           "rep_1" 
#> GGTGTCGA_ATATGCAT GGTGTCGA_CCTATTCA GGTGTCGA_CTGTTCTT GGTTAGGG_AAACTCGA 
#>           "rep_1"           "rep_1"           "rep_1"           "rep_1" 
#> GGTTAGGG_AAATCAGA GGTTAGGG_TTCGCTGG GGTTGAGA_AAACCTCC GGTTGAGA_AGTGATGC 
#>           "rep_1"           "rep_1"           "rep_1"           "rep_1" 
#> GGTTGAGA_CTGTCTGG GGTTGAGA_GAGAGTAT GTAAACAA_TACAGCCG GTAAGCCG_AAAGCCTA 
#>           "rep_1"           "rep_1"           "rep_1"           "rep_1" 
#> GTAAGCCG_AATGACCC GTAAGCCG_CGATTGAT GTAAGCCG_GGCGTTAG GTAATCTG_CGCTAATA 
#>           "rep_1"           "rep_1"           "rep_1"           "rep_1" 
#> GTACGCTT_CCCAAGCA GTACGCTT_GATTTACC GTACGCTT_GCTTTGGC GTACGGAC_ATGACCGA 
#>           "rep_1"           "rep_1"           "rep_1"           "rep_1" 
#> GTACGGAC_CAACAAAT GTACGGAC_CGCTAGTC GTACGGAC_TACCGGCA GTATACGT_CGCTCTCA 
#>           "rep_1"           "rep_1"           "rep_1"           "rep_1" 
#> GTATTGAC_AACACACG GTCAGACC_TAGCCTCG GTCAGACC_TTGGATCG GTCAGGTT_GGATAAAG 
#>           "rep_1"           "rep_1"           "rep_1"           "rep_1" 
#> GTCCACTA_CGCTCTCA GTCCACTA_CTTCTGGA GTCCACTA_GACAAAGG GTCCGTCA_AGCACCAC 
#>           "rep_1"           "rep_1"           "rep_1"           "rep_1" 
#> GTCCGTCA_ATACTCTT GTCCGTCA_CGTGGATA GTCCGTCA_GGGCCAAT GTCCTTGC_TAACCATC 
#>           "rep_1"           "rep_1"           "rep_1"           "rep_1" 
#> GTCCTTGC_TACCAGGC GTCTAATC_CCATTACT GTCTAATC_GATGTGAG GTCTAATC_GATTAGAC 
#>           "rep_1"           "rep_1"           "rep_1"           "rep_1" 
#> GTCTAATC_GGCCCTTA GTGAACTC_CAAGAGGG GTGAACTC_GCCGGATT GTGAACTC_GTCAAAGA 
#>           "rep_1"           "rep_1"           "rep_1"           "rep_1" 
#> GTGAACTC_TCGACGGT GTGAGGCA_ACCCACGA GTGAGGCA_CAGTTTGC GTGATAAA_ATCAGCGC 
#>           "rep_1"           "rep_1"           "rep_1"           "rep_1" 
#> GTGATAAA_CCCAAGCA GTGATAAA_CGCTCTCA GTGATAAA_CGTGCTAG GTGCCCAT_GCTCTAGT 
#>           "rep_1"           "rep_1"           "rep_1"           "rep_1" 
#> GTGCCCAT_GTGTCGGA GTGCGAAG_TGGCCTGT GTGGTGCT_AAGTAATC GTGGTGCT_GGAGAAGC 
#>           "rep_1"           "rep_1"           "rep_1"           "rep_1" 
#> GTGTCACC_TGTGGCTA GTGTCAGG_TCCAGGGA GTTACTAG_AGAAGTTA GTTACTAG_CCCTTGGT 
#>           "rep_1"           "rep_1"           "rep_1"           "rep_1" 
#> GTTACTAG_TGACCCTC GTTCTGCT_CGAACGTA GTTCTGCT_TGGCTACC GTTGTCCG_ATGGGCAC 
#>           "rep_1"           "rep_1"           "rep_1"           "rep_1" 
#> GTTGTCCG_CGATTGAT TAAACCGA_GATCGTTT TAAGGGCC_ACCACGCT TAAGGGCC_GGTCACAG 
#>           "rep_1"           "rep_1"           "rep_1"           "rep_1" 
#> TAAGGGCC_GGTCTGAC TAATCCAT_ATATGCAT TAATCCAT_CGGAATTT TAATGTGG_GGGAACGA 
#>           "rep_1"           "rep_1"           "rep_1"           "rep_1" 
#> TAATGTGG_GTGTAACC TACCCTGC_AATCGGGT TACCCTGC_CTAGACCC TACCCTGC_GAAAGACC 
#>           "rep_1"           "rep_1"           "rep_1"           "rep_1" 
#> TACCGCTC_CCCATAGC TACCGCTC_CTGAGTGC TACCGCTC_CTTCGCAC TACCGCTC_TCCCAATC 
#>           "rep_1"           "rep_1"           "rep_1"           "rep_1" 
#> TACCTAAG_AAATCAGA TACCTAAG_AGGATTCA TACCTAAG_CTGGTGCA TACCTAAG_TGACGGAC 
#>           "rep_1"           "rep_1"           "rep_1"           "rep_1" 
#> TACCTAAG_TGGCTACC TACCTCCC_AACCTGAC TACCTCCC_TGGGATTC TACGCGAG_ACGCCATT 
#>           "rep_1"           "rep_1"           "rep_1"           "rep_1" 
#> TACGCGAG_CCTATTCA TACGCGAG_TGTAGTTT TACGTTCG_TGCCTCAC TACGTTCG_TTGATCTA 
#>           "rep_1"           "rep_1"           "rep_1"           "rep_1" 
#> TAGATCAA_CGCTCTCA TAGATCAA_CTTCAGGT TAGATCAA_GAAAGACC TAGCCACA_GCAGTCGA 
#>           "rep_1"           "rep_1"           "rep_1"           "rep_1" 
#> TAGGCTTT_CGGACAAC TAGGCTTT_CTTCAGGT TAGGCTTT_TACCGGCA TAGGTACG_ATGAAGCA 
#>           "rep_1"           "rep_1"           "rep_1"           "rep_1" 
#> TAGGTACG_TGAAATGA TAGTAGCC_TAGTGTTT TATTAGCG_ACTCACCG TATTAGCG_CCCTAACC 
#>           "rep_1"           "rep_1"           "rep_1"           "rep_1" 
#> TATTAGCG_GACAACTT TATTAGCG_GGTTAGTG TATTAGCG_TGTGCTCA TCAGCCTC_TGCAAGGG 
#>           "rep_1"           "rep_1"           "rep_1"           "rep_1" 
#> TCATACCA_ACACTAAG TCATAGCT_ACAGCGGA TCATTTCA_GGCCCTTA TCATTTCA_TTGGCGTT 
#>           "rep_1"           "rep_1"           "rep_1"           "rep_1" 
#> TCCAGAAG_GATGTATT TCCGACAC_GGGAGGTA TCCGCTGT_GTTACCGC TCGACTGC_AAACAGGG 
#>           "rep_1"           "rep_1"           "rep_1"           "rep_1" 
#> TCGACTGC_ACTAGGAT TCGAGTTT_CCGCAACT TCGCAATC_AGTGATGC TCGCAATC_CATCTCCC 
#>           "rep_1"           "rep_1"           "rep_1"           "rep_1" 
#> TCGCAATC_CGAACGTA TCGCAATC_GTACACTC TCGGTCAT_AGCACCAC TCGGTCAT_GACAGATA 
#>           "rep_1"           "rep_1"           "rep_1"           "rep_1" 
#> TCGTTCCC_GTGTAACC TCGTTCCC_TGAGGTCT TCTAAACT_CTCTTGAC TCTAAACT_TCGGTTTA 
#>           "rep_1"           "rep_1"           "rep_1"           "rep_1" 
#> TCTATTCC_AGACACCA TCTGATTT_AGGAGCTT TCTGATTT_CCTGACAC TCTGATTT_GGGCCAAT 
#>           "rep_1"           "rep_1"           "rep_1"           "rep_1" 
#> TCTGATTT_TGTAGTTT TCTTTGAC_AAACAGGG TCTTTGAC_AGTGATGC TCTTTGAC_CGCTCTCA 
#>           "rep_1"           "rep_1"           "rep_1"           "rep_1" 
#> TCTTTGAC_CTACGAGC TCTTTGAC_GGCTACTA TGAATAGG_AGAGACTA TGAATAGG_TTATCTGT 
#>           "rep_1"           "rep_1"           "rep_1"           "rep_1" 
#> TGACGTCG_ACCTTCTT TGACGTCG_GATCTCGG TGAGAGCG_CCTATTCA TGAGAGCG_GAAGTGCC 
#>           "rep_1"           "rep_1"           "rep_1"           "rep_1" 
#> TGAGCACA_TGCTATTT TGCACCAG_AGTTTAGA TGCACCAG_CCTCGTCT TGCACCAG_GCAGTCGA 
#>           "rep_1"           "rep_1"           "rep_1"           "rep_1" 
#> TGCCGGTA_GTTAACCA TGCCGGTA_GTTTGTTT TGCCGGTA_TGGTTTCT TGCGACTA_CCCTTGGT 
#>           "rep_1"           "rep_1"           "rep_1"           "rep_1" 
#> TGCGACTA_CCGTGTTT TGCGACTA_TGTGCTCA TGCGACTA_TTCACATA TGCTGACG_CGTCAGCA 
#>           "rep_1"           "rep_1"           "rep_1"           "rep_1" 
#> TGCTTCAT_GCAGGGTA TGCTTGGG_CAACAAAT TGGACGGA_TTGTTTAC TGGCTAGT_AATGAATG 
#>           "rep_1"           "rep_1"           "rep_1"           "rep_1" 
#> TGGGAATT_ATATAGGA TGGGAATT_GGTCCCTT TGGGAATT_TTTAACAG TGGTGTCT_AAACTGGT 
#>           "rep_1"           "rep_1"           "rep_1"           "rep_1" 
#> TGGTGTCT_ATTCGACG TGGTTAAC_TGTACACG TGTTATCA_ACGCAGAG TGTTATCA_CAAGTCAT 
#>           "rep_1"           "rep_1"           "rep_1"           "rep_1" 

x <- filterCells(
    object = object,
    minUMIs = c(rep_1 = 100)
)

#> minUMIs: per sample mode
#> rep_1 
#>   100 

#> Returning object unmodified.

#> Returning object unmodified.

#> Parameters:
#>   - >= 100 UMIs per cell
#>   - <= Inf UMIs per cell
#>   - >= 1 genes per cell
#>   - <= Inf genes per cell
#>   - >= 0 novelty score
#>   - <= 1 mitochondrial abundance
#>   - == Inf cells per sample
#>   - >= 1 cells per gene
#> ────────────────────────────────────────────────────────────────────────
#> Cells:
#>      500 cells | prefilter
#>      500 cells | minUMIs >= 100
#>      500 cells | maxUMIs <= Inf
#>      500 cells | minGenes >= 1
#>      500 cells | maxGenes <= Inf
#>      500 cells | minNovelty <= 0
#>      500 cells | maxMitoRatio <= 1
#>      500 cells | nCells == Inf
#>      500 of 500 cells passed filtering (100%)
#> rep_1 
#>   500 
#> ────────────────────────────────────────────────────────────────────────
#> Genes:
#>      500 genes | prefilter
#>      500 genes | minCellsPerGene >= 1
#>      500 of 500 genes passed filtering (100%)

show(x)

#> bcbioSingleCell 0.99.0
#> http://bioinformatics.sph.harvard.edu/bcbioSingleCell
#> citation("bcbioSingleCell")
#> ────────────────────────────────────────────────────────────────────────
#> Upload Dir: "/home/bioinfo/git/bcbioSingleCell-private/inst/extdata/indrops"
#> Upload Date: 2018-01-01
#> R Load Date: 2018-09-27
#> Level: "genes"
#> Organism: "Homo sapiens"
#> Interesting Groups: "sampleName"
#> Metadata File: "inst/extdata/indrops/metadata.csv"
#> Filtered: TRUE
#> ────────────────────────────────────────────────────────────────────────
#> class: SingleCellExperiment 
#> dim: 500 500 
#> metadata(32): version pipeline ... filterParams filterSummary
#> assays(1): counts
#> rownames(500): ENSG00000000457 ENSG00000002586 ... ENSG00000283060
#>   ENSG00000283391
#> rowData names(4): broadClass geneBiotype geneID geneName
#> colnames(500): AAACACTA_ATCCCACG AAACACTA_CTTAGGTA ...
#>   TGTTATCA_ACGCAGAG TGTTATCA_CAAGTCAT
#> colData names(15): nUMI nGene ... aggregate revcomp
#> reducedDimNames(0):
#> spikeNames(0):

metadata(x)$filterParams

#> $nCells
#> [1] Inf
#> 
#> $minUMIs
#> rep_1 
#>   100 
#> 
#> $maxUMIs
#> [1] Inf
#> 
#> $minGenes
#> [1] 1
#> 
#> $maxGenes
#> [1] Inf
#> 
#> $minNovelty
#> [1] 0
#> 
#> $maxMitoRatio
#> [1] 1
#> 
#> $minCellsPerGene
#> [1] 1
#>