Filter Cells

Apply gene detection, novelty score, and mitochondrial abundance cutoffs to cellular barcodes. By default we recommend applying the same filtering cutoff to all samples. The filtering parameters now support per-sample cutoffs, defined using a named numeric vector. When matching per sample, be sure to use the sampleNames() return values (i.e. the sampleName column in sampleData()).

filterCells(object, ...)

# S4 method for SingleCellExperiment
filterCells(object, nCells = Inf,
  minUMIs = 0L, maxUMIs = Inf, minGenes = 0L, maxGenes = Inf,
  minNovelty = 0L, maxMitoRatio = 1L, minCellsPerGene = 1L,
  zinbwave = FALSE)

Arguments

object
Object.

...
Additional arguments.

nCells
scalar integer. Expected number of cells per sample.

minUMIs
scalar integer. Minimum number of UMI disambiguated counts per cell.

maxUMIs
scalar integer. Maximum number of UMI disambiguated counts per cell.

minGenes
scalar integer. Minimum number of genes detected.

maxGenes
scalar integer. Maximum number of genes detected.

minNovelty
scalar integer (0-1). Minimum novelty score (log10 genes per UMI).

maxMitoRatio
scalar integer (0-1). Maximum relative mitochondrial abundance.

minCellsPerGene
scalar integer. Include genes with non-zero expression in at least this many cells.

zinbwave
boolean. Run zinbwave::zinbwave() to automatically apply a ZINB regression model to calculate normalizedValues and weights matrices to by used for differential expression with DESeq2 or edgeR. Note that this calculation should only be performed on filtered data. For large datasets this can take a long time and use a lot of memory, so this calculation is disabled by default.

Value

bcbioSingleCell, with filtering information slotted into metadata() as filterCells and filterParams.

See also

Examples

# SingleCellExperiment ====
object <- cellranger_small
show(object)

#> class: SingleCellExperiment 
#> dim: 500 500 
#> metadata(20): version pipeline ... utilsSessionInfo devtoolsSessionInfo
#> assays(1): counts
#> rownames(500): ENSG00000002586 ENSG00000005022 ... ENSG00000277734
#>   ENSG00000279483
#> rowData names(7): broadClass description ... geneName seqCoordSystem
#> colnames(500): pbmc_1_AAACCTGCAGGCGATA pbmc_1_AAACCTGCATGAAGTA ...
#>   pbmc_1_TTTGGTTTCGCTAGCG pbmc_1_TTTGTCACAGGTCCAC
#> colData names(15): nGene nUMI ... res.1.2 ident
#> reducedDimNames(3): PCA TSNE UMAP
#> spikeNames(0):

x <- filterCells(object)

#> Parameters:
#>   - >= 0 UMIs per cell
#>   - <= Inf UMIs per cell
#>   - >= 0 genes per cell
#>   - <= Inf genes per cell
#>   - >= 0 novelty score
#>   - <= 1 mitochondrial abundance
#>   - == Inf cells per sample
#>   - >= 1 cells per gene
#> ================================================================================
#> Cells:
#>      500 cells | prefilter
#>      500 cells | minUMIs >= 0
#>      500 cells | maxUMIs <= Inf
#>      500 cells | minGenes >= 0
#>      500 cells | maxGenes <= Inf
#>      500 cells | minNovelty <= 0
#>      500 cells | maxMitoRatio <= 1
#>      500 cells | nCells == Inf
#> 
#> pbmc 
#>  500 
#> ================================================================================
#> Genes:
#>      500 genes | prefilter
#>      500 genes | minCellsPerGene <= 1
#> ================================================================================
#> Summary:
#>   - 500 of 500 cells passed filtering (100%)
#>   - 500 of 500 genes passed filtering (100%)

show(x)

#> class: SingleCellExperiment 
#> dim: 500 500 
#> metadata(24): version pipeline ... filterParams filterSummary
#> assays(1): counts
#> rownames(500): ENSG00000002586 ENSG00000005022 ... ENSG00000277734
#>   ENSG00000279483
#> rowData names(7): broadClass description ... geneName seqCoordSystem
#> colnames(500): pbmc_1_AAACCTGCAGGCGATA pbmc_1_AAACCTGCATGAAGTA ...
#>   pbmc_1_TTTGGTTTCGCTAGCG pbmc_1_TTTGTCACAGGTCCAC
#> colData names(15): nGene nUMI ... res.1.2 ident
#> reducedDimNames(3): PCA TSNE UMAP
#> spikeNames(0):

metadata(x)$filterParams

#> $nCells
#> [1] Inf
#> 
#> $minUMIs
#> [1] 0
#> 
#> $maxUMIs
#> [1] Inf
#> 
#> $minGenes
#> [1] 0
#> 
#> $maxGenes
#> [1] Inf
#> 
#> $minNovelty
#> [1] 0
#> 
#> $maxMitoRatio
#> [1] 1
#> 
#> $minCellsPerGene
#> [1] 1
#> 

# Per sample cutoffs
sampleNames(object)

#> pbmc_1 
#> "pbmc" 

x <- filterCells(
    object = object,
    minUMIs = c(pbmc = 100)
)

#> minUMIs: per sample mode
#> pbmc 
#>  100 

#> Parameters:
#>   - >= 100 UMIs per cell
#>   - <= Inf UMIs per cell
#>   - >= 0 genes per cell
#>   - <= Inf genes per cell
#>   - >= 0 novelty score
#>   - <= 1 mitochondrial abundance
#>   - == Inf cells per sample
#>   - >= 1 cells per gene
#> ================================================================================
#> Cells:
#>      500 cells | prefilter
#>      500 cells | minUMIs >= 100
#>      500 cells | maxUMIs <= Inf
#>      500 cells | minGenes >= 0
#>      500 cells | maxGenes <= Inf
#>      500 cells | minNovelty <= 0
#>      500 cells | maxMitoRatio <= 1
#>      500 cells | nCells == Inf
#> 
#> pbmc 
#>  500 
#> ================================================================================
#> Genes:
#>      500 genes | prefilter
#>      500 genes | minCellsPerGene <= 1
#> ================================================================================
#> Summary:
#>   - 500 of 500 cells passed filtering (100%)
#>   - 500 of 500 genes passed filtering (100%)

show(x)

#> class: SingleCellExperiment 
#> dim: 500 500 
#> metadata(24): version pipeline ... filterParams filterSummary
#> assays(1): counts
#> rownames(500): ENSG00000002586 ENSG00000005022 ... ENSG00000277734
#>   ENSG00000279483
#> rowData names(7): broadClass description ... geneName seqCoordSystem
#> colnames(500): pbmc_1_AAACCTGCAGGCGATA pbmc_1_AAACCTGCATGAAGTA ...
#>   pbmc_1_TTTGGTTTCGCTAGCG pbmc_1_TTTGTCACAGGTCCAC
#> colData names(15): nGene nUMI ... res.1.2 ident
#> reducedDimNames(3): PCA TSNE UMAP
#> spikeNames(0):

metadata(x)$filterParams

#> $nCells
#> [1] Inf
#> 
#> $minUMIs
#> pbmc 
#>  100 
#> 
#> $maxUMIs
#> [1] Inf
#> 
#> $minGenes
#> [1] 0
#> 
#> $maxGenes
#> [1] Inf
#> 
#> $minNovelty
#> [1] 0
#> 
#> $maxMitoRatio
#> [1] 1
#> 
#> $minCellsPerGene
#> [1] 1
#>