Known Markers Detected

Known Markers Detected

knownMarkersDetected(all, known, ...)

# S4 method for grouped_df,tbl_df
knownMarkersDetected(all, known, alpha = 0.05,
  filterPromiscuous = FALSE, promiscuousCutoff = 5)

Arguments

all
All markers, generated by Seurat::FindAllMarkers() return.

known
Known markers data.frame imported by readMarkersFile() or pulled from internal cellCycleMarkers data.

...
Additional arguments (for the S4 generic definition).

alpha
Alpha cutoff (adjusted P value; false discovery rate).

filterPromiscuous
Remove markers with poor specificity, that are present in as many clusters as defined by promiscuousCutoff.

promiscuousCutoff
Minimum number of clusters required to consider a gene marker promiscuous.

Value

Grouped tibble.

Note

Both the all and known objects must contain Ensembl gene identifiers in the ensgene column. We must avoid any matching operations based on the gene symbols, since these change often and can mismatch easily.

See also

Other Clustering Utilities: plotKnownMarkersDetected, plotMarkers, plotPCElbow, plotTopMarkers, topBarcodes, topMarkers

Examples

load(system.file(
    file.path("extdata", "seuratAllMarkers.rda"),
    package = "bcbioSingleCell"))

# grouped_df
knownMarkersDetected(
    all = seuratAllMarkers,
    known = cellTypeMarkers[["hsapiens"]]) %>%
    glimpse()

#> Observations: 2
#> Variables: 18
#> $ cell           <chr> "Natural Killer Cell", "Natural Killer Cell"
#> $ cluster        <fctr> 1, 2
#> $ ensgene        <chr> "ENSG00000149294", "ENSG00000149294"
#> $ symbol         <chr> "NCAM1", "NCAM1"
#> $ pct1           <dbl> 0.613, 0.848
#> $ pct2           <dbl> 0.862, 0.787
#> $ avgLogFC       <dbl> -1.0176323, 0.5827054
#> $ pvalue         <dbl> 3.611922e-11, 7.177981e-06
#> $ padj           <dbl> 3.611922e-08, 7.177981e-03
#> $ description    <chr> "neural cell adhesion molecule 1 [Source:HGNC Symbol...
#> $ biotype        <chr> "protein_coding", "protein_coding"
#> $ broadClass     <chr> "coding", "coding"
#> $ geneSeqStart   <int> 112961247, 112961247
#> $ geneSeqEnd     <int> 113278436, 113278436
#> $ seqName        <chr> "11", "11"
#> $ seqStrand      <int> 1, 1
#> $ seqCoordSystem <chr> "chromosome", "chromosome"
#> $ entrez         <list> [4684, 4684]