Load 10X Genomics CellRanger Data

Read 10x Genomics Chromium cell counts from barcodes.tsv, genes.tsv, and matrix.mtx files.

loadCellRanger(uploadDir, refDataDir, interestingGroups = "sampleName",
  prefilter = TRUE, sampleMetadataFile, annotable, ...)

Arguments

uploadDir
Path to CellRanger output directory. This directory path must contain filtered_gene_bc_matrices/ as a child.

refDataDir
Directory path to cellranger reference annotation data.

interestingGroups
Character vector of interesting groups. First entry is used for plot colors during quality control (QC) analysis. Entire vector is used for PCA and heatmap QC functions.

prefilter
Prefilter counts prior to quality control analysis.

sampleMetadataFile
Sample barcode metadata file. Optional for runs with demultiplixed index barcodes (e.g. SureCell), but otherwise required for runs with multipliexed FASTQs containing multiple index barcodes (e.g. inDrop). Consult the GitHub repo for examples and additional information.

annotable
Optional. User-defined gene annotations (a.k.a. "annotable"), which will be slotted into rowData(). Typically this should be left undefined. By default, the function will automatically generate an annotable from the annotations available on Ensembl. If set NULL, then rowData() inside the resulting bcbioSingleCell object will be left empty. This is recommended for projects dealing with genes or transcripts that are poorly annotated.

...
Additional arguments, to be stashed in the metadata() slot.

Value

bcbioSingleCell.

Details

This function is a simplified version of loadSingleCell() optimized for handling CellRanger output.

Examples

# NOT RUN {
# Mus musculus
# Run with Ensembl 88 transcriptome FASTA and GTF files
bcb <- loadSingleCell(
    uploadDir = file.path("indrop_rnaseq", "final"),
    interestingGroups = c("genotype", "treatment"),
    sampleMetadataFile = file.path("meta", "sample_metadata.xlsx"),
    gtfFile = file.path(
        "annotations",
        "Mus_musculus.GRCm38.88.chr_patch_hapl_scaff.gtf.gz"),
    ensemblVersion = 88)
# }