Load bcbio Single-Cell RNA-Seq Data

Load bcbio Single-Cell RNA-Seq Data

loadSingleCell(uploadDir, sampleMetadataFile,
  interestingGroups = "sampleName", ensemblVersion = "current",
  gffFile = NULL, wellMetadataFile = NULL, prefilter = TRUE, ...)

Arguments

uploadDir
Path to final upload directory. This path is set when running bcbio_nextgen -w template.

sampleMetadataFile
Sample barcode metadata file.

interestingGroups
Character vector of interesting groups. First entry is used for plot colors during quality control (QC) analysis. Entire vector is used for PCA and heatmap QC functions.

ensemblVersion
Ensembl release version. Defaults to current, and does not typically need to be user-defined. This parameter can be useful for matching Ensembl annotations against an outdated bcbio annotation build.

gffFile
Optional. GFF (General Feature Format) or GTF (Gene Transfer Format) file, which will be used for transcript-to-gene (tx2gene) and gene-to-symbol (gene2symbol) annotation mappings. A GTF file is GFF v2.

wellMetadataFile
Optional. Well identifier metadata file.

prefilter
Prefilter counts prior to quality control analysis.

...
Additional arguments, to be stashed in the metadata() slot.

Value

bcbioSingleCell.

Note

When working in RStudio, we recommend connecting to the bcbio-nextgen run directory as a remote connection over sshfs.

Examples

# NOT RUN {
# Mus musculus
# Run with Ensembl 88 transcriptome FASTA and GTF files
bcb <- loadSingleCell(
    uploadDir = file.path("indrop_rnaseq", "final"),
    sampleMetadataFile = file.path("meta", "sample_metadata.xlsx"),
    interestingGroups = c("genotype", "treatment"),
    ensemblVersion = 88,
    gffFile = file.path(
        "annotations",
        "Mus_musculus.GRCm38.88.chr_patch_hapl_scaff.gtf.gz"))
# }