Load bcbio Single-Cell RNA-Seq Data

Load bcbio Single-Cell RNA-Seq Data

loadSingleCell(uploadDir, interestingGroups = "sampleName",
  prefilter = TRUE, sampleMetadataFile, gtfFile, ensemblVersion, annotable,
  ...)

Arguments

uploadDir
Path to final upload directory. This path is set when running bcbio_nextgen -w template.

interestingGroups
Character vector of interesting groups. First entry is used for plot colors during quality control (QC) analysis. Entire vector is used for PCA and heatmap QC functions.

prefilter
Prefilter counts prior to quality control analysis.

sampleMetadataFile
Sample barcode metadata file. Optional for runs with demultiplixed index barcodes (e.g. SureCell), but otherwise required for runs with multipliexed FASTQs containing multiple index barcodes (e.g. inDrop). Consult the GitHub repo for examples and additional information.

gtfFile
Optional but recommended. GTF (Gene Transfer Format) file, which will be used for gene-to-symbol (gene2symbol) and transcript-to-gene (tx2gene) annotation mappings.

ensemblVersion
Optional. Ensembl release version. If NULL, defaults to current release, and does not typically need to be user-defined. This parameter can be useful for matching Ensembl annotations against an outdated bcbio annotation build.

annotable
Optional. User-defined gene annotations (a.k.a. "annotable"), which will be slotted into rowData(). Typically this should be left undefined. By default, the function will automatically generate an annotable from the annotations available on Ensembl. If set NULL, then rowData() inside the resulting bcbioSingleCell object will be left empty. This is recommended for projects dealing with genes or transcripts that are poorly annotated.

...
Additional arguments, to be stashed in the metadata() slot.

Value

bcbioSingleCell.

Note

When working in RStudio, we recommend connecting to the bcbio-nextgen run directory as a remote connection over sshfs.

Examples

# Homo sapiens
# Minimal working example dataset
extdataDir <- system.file("extdata", package = "bcbioSingleCell")
uploadDir <- file.path(extdataDir, "harvard_indrop_v3")
sampleMetadataFile <- file.path(extdataDir, "harvard_indrop_v3.xlsx")
annotable <- annotable("Homo sapiens", release = 90)

#> Loading Ensembl annotations from AnnotationHub
#> 2017-10-27

#> EnsDB AH57757: Homo sapiens Ensembl 90

bcb <- loadSingleCell(
    uploadDir = uploadDir,
    sampleMetadataFile = sampleMetadataFile,
    annotable = annotable)

#> 2017-11-10_bcbio

#> 2 samples detected

#> Reading project-summary.yaml

#> Reading log files

#> 1000 reads per cellular barcode cutoff detected

#> Reading data and program versions

#> Parsed with column specification:
#> cols(
#>   genome = col_character(),
#>   resource = col_character(),
#>   version = col_character()
#> )

#> Genome: Homo sapiens (hg19)

#> UMI type: harvard-indrop-v3

#> Reading harvard_indrop_v3.xlsx

#> Reading cellular barcode distributions

#>    |                                                  | 0 % ~calculating  

#> Reading run1-AGAGGATA-barcodes.tsv

#>    |+++++++++++++++++++++++++                         | 50% ~01s          

#> Reading run2-AGAGGATA-barcodes.tsv

#>    |++++++++++++++++++++++++++++++++++++++++++++++++++| 100% elapsed = 02s

#> Reading counts at gene level

#>    |                                                  | 0 % ~calculating  

#> run1_AGAGGATA

#> Reading run1-AGAGGATA.mtx

#> Reading run1-AGAGGATA.mtx.colnames

#> Reading run1-AGAGGATA.mtx.rownames

#>    |+++++++++++++++++++++++++                         | 50% ~04s          

#> run2_AGAGGATA

#> Reading run2-AGAGGATA.mtx

#> Reading run2-AGAGGATA.mtx.colnames

#> Reading run2-AGAGGATA.mtx.rownames

#>    |++++++++++++++++++++++++++++++++++++++++++++++++++| 100% elapsed = 09s

#> Calculating barcode metrics

#> 18244 cellular barcodes detected

#> 13 mitochondrial genes detected

#> 18240 cellular barcodes passed pre-filtering (100%)

# Mus musculus
# Run with Ensembl 88 transcriptome FASTA and GTF files

# NOT RUN {
bcb <- loadSingleCell(
    uploadDir = file.path("indrop_rnaseq", "final"),
    interestingGroups = c("genotype", "treatment"),
    sampleMetadataFile = file.path("meta", "sample_metadata.xlsx"),
    gtfFile = file.path(
        "annotations",
        "Mus_musculus.GRCm38.88.chr_patch_hapl_scaff.gtf.gz"),
    ensemblVersion = 88)
# }