Load bcbio Single-Cell RNA-Seq Data
Source: R/loadSingleCell.R

Load bcbio Single-Cell RNA-Seq Data

loadSingleCell(uploadDir, sampleMetadataFile,
  interestingGroups = "sampleName", organism, genomeBuild = NULL,
  ensemblRelease = NULL, isSpike = NULL, gffFile = NULL,
  prefilter = TRUE, ...)

Arguments

uploadDir
Path to final upload directory. This path is set when running bcbio_nextgen -w template.

sampleMetadataFile
Sample barcode metadata file. Optional for runs with demultiplixed index barcodes (e.g. SureCell), but otherwise required for runs with multipliexed FASTQs containing multiple index barcodes (e.g. inDrop).

interestingGroups
Character vector of interesting groups. Must be formatted in camel case and intersect with sampleData() colnames.

organism
Organism name. Use the full latin name (e.g. "Homo sapiens"), since this will be input downstream to AnnotationHub and ensembldb, unless gffFile is set.

genomeBuild
Optional. Ensembl genome build name (e.g. "GRCh38"). This will be passed to AnnotationHub for EnsDb annotation matching, unless gffFile is set.

ensemblRelease
Optional. Ensembl release version. If NULL, defaults to current release, and does not typically need to be user-defined. Passed to AnnotationHub for EnsDb annotation matching, unless gffFile is set.

isSpike
Optional. Gene names corresponding to FASTA spike-in sequences (e.g. ERCCs, EGFP, TDTOMATO).

gffFile
Optional, not recommended. By default, we recommend leaving this NULL for genomes that are supported on Ensembl. In this case, the row annotations (rowRanges()) will be obtained automatically from Ensembl by passing the organism, genomeBuild, and ensemblRelease arguments to AnnotationHub and ensembldb. For a genome that is not supported on Ensembl and/or AnnotationHub, a GFF/GTF (General Feature Format) file is required. Generally, we recommend using a GTF (GFFv2) file here over a GFF3 file if possible, although all GFF formats are supported. The function will internally generate a TxDb containing transcript-to-gene mappings and construct a GRanges object containing the genomic ranges (rowRanges()).

prefilter
Prefilter counts prior to quality control analysis.

...
Additional arguments, to be stashed in the metadata() slot.

Value

bcbioSingleCell.

Note

When working in RStudio, we recommend connecting to the bcbio-nextgen run directory as a remote connection over sshfs.

Examples

uploadDir <- system.file("extdata/indrop", package = "bcbioSingleCell")
loadSingleCell(
    uploadDir = uploadDir,
    sampleMetadataFile = file.path(uploadDir, "metadata.csv"),
    organism = "Homo sapiens"
)

#> 2018-01-01_bcbio

#> 1 samples detected:
#> multiplexed_AAAAAAAA

#> Reading project-summary.yaml

#> Data versions are missing

#> 1000 reads per cellular barcode cutoff detected

#> UMI type: harvard-indrop-v3

#> Reading metadata.csv

#> Parsed with column specification:
#> cols(
#>   fileName = col_character(),
#>   description = col_character(),
#>   index = col_integer(),
#>   sequence = col_character(),
#>   sampleName = col_character(),
#>   sampleNameAggregate = col_character()
#> )

#> Multiplexed samples detected

#> Reading counts as genes

#> Reading cellular barcode distributions

#> Making GRanges from Ensembl with AnnotationHub 2.10.1 (2017-10-27)

#> id: "AH57757"
#> organism: "Homo sapiens"
#> build: "GRCh38"
#> release: "90"
#> format: "genes"

#> Using geneName instead of symbol

#> Setting geneBiotype as factor

#> Setting seqCoordSystem as factor

#> Defining broadClass using: geneName, geneBiotype, seqnames

#> Calculating cellular barcode metrics

#> 500 cells detected

#> 449 coding genes detected

#> 13 mitochondrial genes detected

#> 500 cellular barcodes passed pre-filtering (100%)

#> bcbioSingleCell 0.1.0
#> samples: 1
#> cells: 500
#> genes: 500
#> organism: Homo sapiens
#> bcbio dir: /Users/mike/GitHub/hbc/bcbioSingleCell/inst/extdata/indrop
#> bcbio date: 2018-01-01
#> load date: 2018-04-02