Sample Barcode Metrics

Sample Barcode Metrics

# S4 method for bcbioSingleCell
metrics(object, interestingGroups)

# S4 method for seurat
metrics(object, interestingGroups)

Arguments

object
Object.

interestingGroups
Interesting group, to use for colors.

Value

data.frame with cellular barcodes as rows.

See also

sampleMetadata().

Other Quality Control Metrics: plotCellCounts, plotGenesPerCell, plotMitoRatio, plotMitoVsCoding, plotNovelty, plotReadsPerCell, plotUMIsPerCell, plotUMIsVsGenes, plotZerosVsDepth

Examples

load(system.file(
    file.path("extdata", "bcb.rda"),
    package = "bcbioSingleCell"))
load(system.file(
    file.path("extdata", "seurat.rda"),
    package = "bcbioSingleCell"))

# bcbioSingleCell
metrics(bcb) %>% glimpse()

#> Loading required package: bcbioSingleCell

#> Error in (function (classes, fdef, mtable) {    methods <- .findInheritedMethods(classes, fdef, mtable)    if (length(methods) == 1L)         return(methods[[1L]])    else if (length(methods) == 0L) {        cnames <- paste0("\"", vapply(classes, as.character,             ""), "\"", collapse = ", ")        stop(gettextf("unable to find an inherited method for function %s for signature %s",             sQuote(fdef@generic), sQuote(cnames)), domain = NA)    }    else stop("Internal error in finding inherited methods; didn't return a unique method",         domain = NA)})(list(structure("bcbioSingleCell", package = "bcbioSingleCell")),     structure(function (object, ...)     {        standardGeneric("metrics")    }, generic = structure("metrics", package = "bcbioBase"), package = "bcbioBase", group = list(), valueClass = character(0), signature = "object", default = `\001NULL\001`, skeleton = (function (object,         ...)     stop("invalid call in method dispatch to 'metrics' (no default method)",         domain = NA))(object, ...), class = structure("nonstandardGenericFunction", package = "methods")),     <environment>): unable to find an inherited method for function 'metrics' for signature '"bcbioSingleCell"'

# seurat
metrics(seurat) %>% glimpse()

#> Observations: 243
#> Variables: 13
#> $ nGene             <int> 734, 774, 541, 545, 653, 816, 525, 725, 614, 566,...
#> $ nUMI              <int> 6220, 9622, 1896, 2137, 4918, 9247, 1891, 6031, 4...
#> $ sampleID          <fct> M1, M1, M1, M1, M1, M1, M1, M1, M1, M1, M1, M1, M...
#> $ nCoding           <int> 5495, 8536, 1682, 1908, 3770, 8336, 1484, 5280, 3...
#> $ nMito             <int> 442, 438, 103, 83, 830, 423, 246, 436, 543, 208, ...
#> $ log10GenesPerUMI  <dbl> 0.7553649, 0.7252193, 0.8338414, 0.8217889, 0.762...
#> $ mitoRatio         <dbl> 0.07106109, 0.04552068, 0.05432489, 0.03883949, 0...
#> $ nCount            <int> 95155, 124786, 34607, 34005, 77663, 132228, 41090...
#> $ sampleName        <fct> M1, M1, M1, M1, M1, M1, M1, M1, M1, M1, M1, M1, M...
#> $ description       <fct> M1, M1, M1, M1, M1, M1, M1, M1, M1, M1, M1, M1, M...
#> $ origIdent         <fct> M1, M1, M1, M1, M1, M1, M1, M1, M1, M1, M1, M1, M...
#> $ res0x8            <fct> 0, 1, 1, 3, 2, 1, 3, 0, 1, 2, 3, 0, 1, 2, 1, 0, 3...
#> $ interestingGroups <fct> M1, M1, M1, M1, M1, M1, M1, M1, M1, M1, M1, M1, M...