Read 10X Genomics Cell Ranger Data

Read 10x Genomics Chromium cell counts from barcodes.tsv, genes.tsv, and matrix.mtx files.

readCellRanger(uploadDir, format = c("mtx", "hdf5"), filtered = TRUE,
  organism = NULL, sampleMetadataFile = NULL, refdataDir = NULL,
  interestingGroups = "sampleName", transgeneNames = NULL,
  spikeNames = NULL, ...)

Arguments

uploadDir
string. Path to Cell Ranger output directory. This directory path must contain filtered_gene_bc_matrices* as a child directory.

format
string. Output format, either MatrixMarket ("mtx") or HDF5 ("hdf5").

filtered
boolean. Use filtered (recommended) or raw counts. Note that raw counts still contain only whitelisted cellular barcodes.

organism
string or NULL. Organism name. Use the full Latin name (e.g. "Homo sapiens"), since this will be input downstream to AnnotationHub and ensembldb, unless gffFile is set. If left NULL (not recommended), the function call will skip loading gene-level annotations into rowRanges(). This can be useful for poorly annotation genomes or experiments involving multiple genomes.

sampleMetadataFile
string or NULL. Sample barcode metadata file. Optional for runs with demultiplixed index barcodes (e.g. SureCell) and can be left NULL, but otherwise required for runs with multipliexed FASTQs containing multiple index barcodes (e.g. inDrop).

refdataDir
string or NULL. Directory path to Cell Ranger reference annotation data.

interestingGroups
character or NULL. Character vector of interesting groups. Must be formatted in camel case and intersect with sampleData() colnames.

transgeneNames
character or NULL. Vector indicating which assay() rows denote transgenes (e.g. EGFP, TDTOMATO).

spikeNames
character or NULL. Vector indicating which assay() rows denote spike-in sequences (e.g. ERCCs).

...
Additional arguments, to be stashed in the metadata() slot.

Value

SingleCellExperiment.

Directory structure

Cell Ranger can vary in its output directory structure, but we're requiring a single, consistent directory structure for all datasets, even those that only contain a single sample:

file.path(
    "<uploadDir>",
    "<sampleName>",
    "outs",
    "filtered_gene_bc_matrices*",
    "outs",
    "<genomeBuild>",
    "matrix.mtx"
)

Sample metadata

A user-supplied sample metadata file defined by sampleMetadataFile is required for multiplexed datasets. Otherwise this can be left NULL, and minimal sample data will be used, based on the directory names.

Reference data

We strongly recommend supplying the corresponding reference data required for Cell Ranger with the refdataDir argument. It will convert the gene annotations defined in the GTF file into a GRanges object, which get slotted in rowRanges(). Otherwise, the function will attempt to use the most current annotations available from Ensembl, and some gene IDs may not match, due to deprecation in the current Ensembl release.

Examples

uploadDir <- system.file("extdata/cellranger", package = "bcbioSingleCell")
x <- readCellRanger(
    uploadDir = uploadDir,
    organism = "Homo sapiens"
)

#> 1 sample(s) detected

#> Reading counts at gene level

#> Reading counts in mtx format

#> /Users/mike/GitHub/hbc/bcbioSingleCell/inst/extdata/cellranger/pbmc/outs/filtered_gene_bc_matrices

#> Multiplexed (aggregated) samples detected

#> Making GRanges from Ensembl with AnnotationHub 2.12.0 (2018-04-30)

#> Ensembl 92 EnsDb for Homo Sapiens

#> id: "AH60977"
#> organism: "Homo sapiens"
#> build: "GRCh38"
#> release: "92"
#> format: "genes"

#> Standardizing the metadata columns

#> Defining broad class using: geneName, geneBiotype, seqnames

#> Sorting ranges by geneID

#> Calculating cellular barcode metrics

#> 500 cells detected

#> 312 coding genes

#> 0 mitochondrial genes

#> 500 / 500 cellular barcodes passed pre-filtering (100%)

#> Warning: Unannotated rows detected (9):
#> ENSG00000236743, ENSG00000278566, ENSG00000273547, ENSG00000279244, ENSG00000...

show(x)

#> class: SingleCellExperiment 
#> dim: 500 500 
#> metadata(20): version pipeline ... utilsSessionInfo devtoolsSessionInfo
#> assays(1): counts
#> rownames(500): ENSG00000243485 ENSG00000237613 ... ENSG00000230703
#>   ENSG00000142677
#> rowData names(7): broadClass description ... geneName seqCoordSystem
#> colnames(500): pbmc_1_AAACCTGAGAAGGCCT pbmc_1_AAACCTGAGACAGACC ...
#>   pbmc_1_ACTTACTCACTGCCAG pbmc_1_ACTTACTCAGTACACT
#> colData names(10): sampleID nUMI ... description index
#> reducedDimNames(0):
#> spikeNames(0):