Sanitize Markers Output

Sanitize Markers Output

sanitizeMarkers(object, ...)

# S4 method for seurat
sanitizeMarkers(object, markers)

Arguments

object
Object.

...
Additional arguments.

markers
Original Seurat::FindAllMarkers() data.frame.

Value

data.frame, arranged by adjusted P value.

Note

Seurat::FindAllMarkers() maps the counts matrix rownames correctly in the gene column, whereas Seurat::FindMarkers() maps them correctly in the rownames of the returned marker data.frame.

See also

Examples

# seurat ====
object <- seurat_small

# `FindAllMarkers()` return
invisible(capture.output(
    all_markers <- Seurat::FindAllMarkers(object)
))
all_sanitized <- sanitizeMarkers(
    object = object,
    markers = all_markers
)
glimpse(all_sanitized)

#> Observations: 1,215
#> Variables: 13
#> $ cluster        <fct> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0...
#> $ avgLogFC       <dbl> 2.3278016, 2.8655976, 2.6441318, 1.8779831, 1.622767...
#> $ pct1           <dbl> 0.917, 1.000, 1.000, 0.910, 1.000, 1.000, 0.985, 1.0...
#> $ pct2           <dbl> 0.215, 0.823, 0.899, 0.207, 0.507, 0.877, 0.436, 0.4...
#> $ rowname        <chr> "S100A12", "S100A8", "S100A9", "VCAN", "FCN1", "LYZ"...
#> $ pvalue         <dbl> 1.028977e-63, 3.497403e-62, 6.912426e-62, 4.509736e-...
#> $ padj           <dbl> 5.144887e-61, 1.748702e-59, 3.456213e-59, 2.254868e-...
#> $ geneID         <chr> "ENSG00000163221", "ENSG00000143546", "ENSG000001632...
#> $ geneName       <chr> "S100A12", "S100A8", "S100A9", "VCAN", "FCN1", "LYZ"...
#> $ geneBiotype    <fct> protein_coding, protein_coding, protein_coding, prot...
#> $ description    <chr> "S100 calcium binding protein A12 [Source:HGNC Symbo...
#> $ seqCoordSystem <fct> chromosome, chromosome, chromosome, chromosome, chro...
#> $ broadClass     <fct> coding, coding, coding, coding, coding, coding, othe...

# `FindMarkers()` return
invisible(capture.output(
    ident_3_markers <- Seurat::FindMarkers(
        object = object,
        ident.1 = "3",
        ident.2 = NULL
    )
))
ident_3_sanitized <- sanitizeMarkers(
    object = object,
    markers = ident_3_markers
)
glimpse(ident_3_sanitized)

#> Observations: 273
#> Variables: 11
#> $ avgLogFC       <dbl> 1.6645500, 1.5023792, 1.3624811, 1.0638911, 1.488073...
#> $ pct1           <dbl> 1.000, 1.000, 1.000, 1.000, 1.000, 1.000, 1.000, 1.0...
#> $ pct2           <dbl> 0.728, 0.723, 0.919, 1.000, 0.773, 0.890, 0.933, 0.9...
#> $ pvalue         <dbl> 1.047295e-43, 1.189784e-41, 2.095776e-39, 3.430040e-...
#> $ padj           <dbl> 5.236475e-41, 5.948919e-39, 1.047888e-36, 1.715020e-...
#> $ geneID         <chr> "ENSG00000204482", "ENSG00000204472", "ENSG000001300...
#> $ geneName       <chr> "LST1", "AIF1", "SAT1", "FTH1", "IFITM2", "PSAP", "C...
#> $ geneBiotype    <fct> protein_coding, protein_coding, protein_coding, prot...
#> $ description    <chr> "leukocyte specific transcript 1 [Source:HGNC Symbol...
#> $ seqCoordSystem <fct> chromosome, chromosome, chromosome, chromosome, chro...
#> $ broadClass     <fct> coding, coding, coding, coding, coding, coding, codi...