Sanitize Markers Output
Source: R/AllGenerics.R, R/methods-sanitizeMarkers.R

Sanitize Markers Output

sanitizeMarkers(object, ...)

# S4 method for seurat
sanitizeMarkers(object, markers)

Arguments

object
Object.

...
Additional arguments.

markers
Original Seurat::FindAllMarkers() data.frame.

Value

data.frame, arranged by adjusted P value.

Note

Seurat::FindAllMarkers() maps the counts matrix rownames correctly in the gene column, whereas Seurat::FindMarkers() maps them correctly in the rownames of the returned marker data.frame.

See also

Examples

# seurat ====
invisible(capture.output(
    all_markers <- FindAllMarkers(seurat_small)
))
all_sanitized <- sanitizeMarkers(
    object = seurat_small,
    markers = all_markers
)

#> Renaming `pVal` column to `pvalue` (matching DESeq2)

#> Renaming `pValAdj` column to `padj` (matching DESeq2)

glimpse(all_sanitized)

#> Observations: 1,380
#> Variables: 18
#> $ cluster        <fct> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0...
#> $ geneID         <chr> "ENSG00000164692", "ENSG00000108821", "ENSG000001000...
#> $ geneName       <chr> "COL1A2", "COL1A1", "LGALS1", "SERPINE2", "COL6A3", ...
#> $ pct1           <dbl> 1.000, 1.000, 0.806, 0.831, 0.927, 0.976, 0.766, 1.0...
#> $ pct2           <dbl> 0.465, 0.596, 0.098, 0.125, 0.229, 0.306, 0.096, 0.5...
#> $ avgLogFC       <dbl> 2.5553118, 2.7949173, 2.2995216, 2.4224445, 2.284212...
#> $ padj           <dbl> 5.313721e-58, 1.979923e-57, 1.215935e-54, 1.019308e-...
#> $ pvalue         <dbl> 1.062744e-60, 3.959847e-60, 2.431870e-57, 2.038616e-...
#> $ rowname        <chr> "COL1A2", "COL1A1", "LGALS1", "SERPINE2", "COL6A3", ...
#> $ seqnames       <fct> 7, 17, 22, 2, 2, 2, 14, 2, 22, 17, 9, 21, 5, 7, 8, 1...
#> $ start          <int> 94394561, 50183289, 37675608, 223975112, 237324003, ...
#> $ end            <int> 94431232, 50201632, 37679806, 224039319, 237414375, ...
#> $ width          <int> 36672, 18344, 4199, 64208, 90373, 75733, 12520, 3842...
#> $ strand         <fct> +, -, +, -, -, -, +, +, +, -, +, +, -, -, -, +, +, +...
#> $ geneBiotype    <fct> protein_coding, protein_coding, protein_coding, prot...
#> $ description    <chr> "collagen type I alpha 2 chain [Source:HGNC Symbol;A...
#> $ seqCoordSystem <fct> chromosome, chromosome, chromosome, chromosome, chro...
#> $ broadClass     <fct> coding, coding, coding, coding, coding, coding, codi...

invisible(capture.output(
    ident_3_markers <- FindMarkers(
        object = seurat_small,
        ident.1 = "3",
        ident.2 = NULL
    )
))
ident_3_sanitized <- sanitizeMarkers(
    object = seurat_small,
    markers = ident_3_markers
)

#> Renaming `pVal` column to `pvalue` (matching DESeq2)

#> Renaming `pValAdj` column to `padj` (matching DESeq2)

glimpse(ident_3_sanitized)

#> Observations: 427
#> Variables: 17
#> $ geneID         <chr> "ENSG00000156508", "ENSG00000198899", "ENSG000002485...
#> $ geneName       <chr> "EEF1A1", "MT-ATP6", "MTATP6P1", "MT-ND5", "MTND1P23...
#> $ pct1           <dbl> 0.865, 1.000, 1.000, 0.966, 0.989, 1.000, 0.989, 0.9...
#> $ pct2           <dbl> 0.985, 0.993, 0.988, 0.908, 0.959, 0.976, 0.988, 0.9...
#> $ avgLogFC       <dbl> -1.709205, 1.159734, 1.173533, 1.326033, 1.269361, 1...
#> $ padj           <dbl> 2.710328e-32, 1.796991e-31, 1.869653e-31, 2.357196e-...
#> $ pvalue         <dbl> 5.420656e-35, 3.593981e-34, 3.739306e-34, 4.714391e-...
#> $ rowname        <chr> "EEF1A1", "MT.ATP6", "MTATP6P1", "MT.ND5", "MTND1P23...
#> $ seqnames       <fct> 6, MT, 1, MT, 1, MT, MT, MT, MT, MT, X, MT, 16, MT, ...
#> $ start          <int> 73515750, 8527, 633696, 12337, 629062, 14747, 9207, ...
#> $ end            <int> 73523797, 9207, 634376, 14148, 629433, 15887, 9990, ...
#> $ width          <int> 8048, 681, 681, 1812, 372, 1141, 784, 1042, 1378, 95...
#> $ strand         <fct> -, +, +, +, +, +, +, +, +, +, +, +, -, -, +, +, +, -...
#> $ geneBiotype    <fct> protein_coding, protein_coding, unprocessed_pseudoge...
#> $ description    <chr> "eukaryotic translation elongation factor 1 alpha 1 ...
#> $ seqCoordSystem <fct> chromosome, chromosome, chromosome, chromosome, chro...
#> $ broadClass     <fct> coding, mito, pseudo, mito, pseudo, mito, mito, mito...