Update an Object to Its Current Class Definition

updateObject is a generic function that returns an instance of object updated to its current class definition.

# S4 method for bcbioSingleCell
updateObject(object)

Arguments

object
Object to be updated for updateObject and updateObjectFromSlots.

Object for slot information to be extracted from for getObjectSlots.

Value

Valid object.

Details

Updating objects is primarily useful when an object has been serialized (e.g., stored to disk) for some time (e.g., months), and the class definition has in the mean time changed. Because of the changed class definition, the serialized instance is no longer valid.

updateObject requires that the class of the returned object be the same as the class of the argument object, and that the object is valid (see validObject). By default, updateObject has the following behaviors:

updateObject(ANY, …, verbose=FALSE)

By default, updateObject uses heuristic methods to determine whether the object should be the `new' S4 type (introduced in R 2.4.0), but is not. If the heuristics indicate an update is required, the updateObjectFromSlots function tries to update the object. The default method returns the original S4 object or the successfully updated object, or issues an error if an update is required but not possible. The optional named argument verbose causes a message to be printed describing the action. Arguments … are passed to updateObjectFromSlots.

updateObject(list, …, verbose=FALSE)

Visit each element in list, applying updateObject(list[[elt]], …, verbose=verbose).

updateObject(environment, …, verbose=FALSE)

Visit each element in environment, applying updateObject(environment[[elt]], …, verbose=verbose)

updateObject(formula, …, verbose=FALSE)

Do nothing; the environment of the formula may be too general (e.g., R_GlobalEnv) to attempt an update.

updateObject(envRefClass, …, verbose=FALSE)

Attempt to update objects from fields using a strategy like updateObjectFromSlots Method 1.

updateObjectFromSlots(object, objclass=class(object), …, verbose=FALSE) is a utility function that identifies the intersection of slots defined in the object instance and objclass definition. Under Method 1, the corresponding elements in object are then updated (with updateObject(elt, …, verbose=verbose)) and used as arguments to a call to new(class, …), with … replaced by slots from the original object. If this fails, then Method 2 tries new(class) and assigns slots of object to the newly created instance.

getObjectSlots(object) extracts the slot names and contents from object. This is useful when object was created by a class definition that is no longer current, and hence the contents of object cannot be determined by accessing known slots.

See also

Examples

# bcbioSingleCell ====
updateObject(indrops_small)

#> Upgrading from 0.1.18 to 0.1.18

#> bcbioSingleCell 0.1.18
#> class: SingleCellExperiment 
#> dim: 500 500 
#> metadata(35): version pipeline ... filterSummary previousVersion
#> assays(4): counts normalizedValues weights logcounts
#> rownames(500): ENSG00000000457 ENSG00000002586 ... ENSG00000283060
#>   ENSG00000283391
#> rowData names(7): broadClass description ... geneName seqCoordSystem
#> colnames(500): multiplexed_AAAAAAAA_AAACACTA_ATCCCACG
#>   multiplexed_AAAAAAAA_AAACACTA_CTTAGGTA ...
#>   multiplexed_AAAAAAAA_TGTTATCA_ACGCAGAG
#>   multiplexed_AAAAAAAA_TGTTATCA_CAAGTCAT
#> colData names(15): sampleID nCount ... aggregate revcomp
#> reducedDimNames(0):
#> spikeNames(0):
#> ============================================================
#> Upload Dir: "/Users/mike/GitHub/hbc/bcbioSingleCell/inst/extdata/indrops"
#> Upload Date: 2018-01-01
#> R Load Date: 2018-07-31
#> Level: "genes"
#> Organism: "Homo sapiens"
#> Interesting Groups: "sampleName"
#> Metadata File: "inst/extdata/indrops/metadata.csv"
#> AnnotationHub: "AH57757"
#> Ensembl Release: "90"
#> Genome Build: "GRCh38"
#> Filtered: TRUE
#> zinbwave: TRUE